### 2020-01-07 #2001 @whedon
[PRE REVIEW]: EvoMaster: A Search-Based System Test Generation Tool
**Submitting author:** @arcuri82 (<a href="http://orcid.org/0000-0003-0799-2930">Andrea Arcuri</a>)
**Repository:** <a href="https://github.com/EMResearch/EvoMaster" target ="_blank">https://github.com/EMResearch/EvoMaster</a>
**Version:** v1.0.0
