### 2020-01-08 fbdfb79
new spring-rest-mongodb module
