+++ e2e-tests/spring-rest-mongodb/src/main/java/com/foo/customer/CustomerMongoRepository.java
@@ -1,0 +1,13 @@
+package com.foo.customer;
+
+import org.springframework.data.mongodb.repository.MongoRepository;
+
+import java.util.List;
+
+public interface CustomerMongoRepository extends MongoRepository<CustomerEntity, String> {
+
+    CustomerEntity findByFirstName(String firstName);
+
+    List<CustomerEntity> findByLastName(String lastName);
+
+}
