+++ e2e-tests/spring-rest-mongodb/src/test/java/com/foo/customer/CustomerEmbeddedController.java
@@ -1,0 +1,15 @@
+package com.foo.customer;
+
+import com.foo.spring.rest.mongodb.SpringRestMongodbController;
+
+public class CustomerEmbeddedController extends SpringRestMongodbController {
+
+    public CustomerEmbeddedController() {
+        super(CustomerApplication.class);
+    }
+
+    @Override
+    public String getPackagePrefixesToCover() {
+        return "com.foo.customer";
+    }
+}
