+++ e2e-tests/spring-rest-mongodb/src/test/java/org/evomaster/e2etests/spring/rest/mongodb/JDBCMongoDriverTest.java
@@ -1,0 +1,11 @@
+package org.evomaster.e2etests.spring.rest.mongodb;
+
+import org.junit.jupiter.api.Test;
+
+public class JDBCMongoDriverTest {
+
+    @Test
+    public void testLoad() throws ClassNotFoundException {
+        Class.forName("mongodb.jdbc.MongoDriver");
+    }
+}
