+++ e2e-tests/spring-rest-mongodb/src/main/java/com/foo/customer/CustomerApplication.java
@@ -1,0 +1,16 @@
+package com.foo.customer;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
+import springfox.documentation.swagger2.annotations.EnableSwagger2;
+
+@EnableSwagger2
+@SpringBootApplication(exclude = SecurityAutoConfiguration.class)
+public class CustomerApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(CustomerApplication.class, args);
+    }
+
+}

[…]
