### 2020-01-08 a7f8bd6
reseting state of mongodb after test execution
