### 2020-01-08 45fa733
refactor: mongodb renamed simply as mongo in package and class names.
