+++ e2e-tests/spring-rest-mongodb/src/test/java/org/evomaster/e2etests/spring/rest/mongo/MongoContainerTest.java
@@ -1,4 +1,4 @@
-package org.evomaster.e2etests.spring.rest.mongodb;
+package org.evomaster.e2etests.spring.rest.mongo;
 
 import com.mongodb.MongoClient;
 import com.mongodb.client.MongoCollection;
