### 2020-01-08 1cf8dc8
refactor: removing unnecessary code
