### 2020-01-10 29d7d87
Updating the options.md
