### 2020-01-11 #2001 @mado89
I could be a reviewer. However:
* I'd need some time. I don't think I can give this software a thorough review in just two weeks. By looking at it I think I can manage to do it until the end of February
* I'd possibly need more instructions on how to use the software. By looking at it I was not able to get a grasp how to use it (but I did not spend too much time on it)

So its up to you @gkthiruvathukal as editor and and @arcuri82
