+++ e2e-tests/spring-examples/src/main/java/com/foo/rest/examples/spring/strings/StringsRest.java
@@ -72,5 +72,33 @@ public class StringsRest {
         return dto;
     }
 
+    @RequestMapping(
+            value = "/nope/{s}",
+            method = RequestMethod.GET,
+            produces = MediaType.APPLICATION_JSON
+    )
+    public StringsResponseDto nope(
+            @PathVariable("s") String s
+    ) throws NopeException {
+        StringsResponseDto dto = new StringsResponseDto();
+        if(s == null || s.isEmpty()){
+            dto.valid = false;
+        }
+        else if(s.length() == 3 && "123456789".contains(s)){
+            dto.valid = true;
+        } else {
+            dto.valid = false;
+        }
+
+        throw new NopeException("Nope");
+
+        //return dto;
+    }
+
+    class NopeException extends Exception{
+        NopeException(String text){
+            super(text);
+        }
+    }
 
 }
