### 2020-01-13 104be8b
Minor test update (mostly to rerun github tests).
