+++ e2e-tests/spring-examples/src/test/java/org/evomaster/e2etests/spring/examples/strings/StringsEMTest.java
@@ -26,6 +26,7 @@ public class StringsEMTest extends StringsTestBase {
                     assertHasAtLeastOne(solution, HttpVerb.GET, 200, "/api/strings/equalsFoo/{s}", "false");
                     assertHasAtLeastOne(solution, HttpVerb.GET, 200, "/api/strings/contains/{s}", "false");
                     assertHasAtLeastOne(solution, HttpVerb.GET, 200, "/api/strings/startEnds/{s}", "false");
+                    assertHasAtLeastOne(solution, HttpVerb.GET, 500);
 
                     assertHasAtLeastOne(solution, HttpVerb.GET, 200, "/api/strings/equalsFoo/{s}", "true");
                     assertHasAtLeastOne(solution, HttpVerb.GET, 200, "/api/strings/contains/{s}", "true");
