+++ client-java/instrumentation/src/main/java/org/evomaster/client/java/instrumentation/coverage/ThirdPartyClassVisitor.java
@@ -1,9 +1,8 @@
 package org.evomaster.client.java.instrumentation.coverage;
 
 import org.evomaster.client.java.instrumentation.Constants;
+import org.evomaster.client.java.instrumentation.mongo.CallbackMethodVisitor;
 import org.evomaster.client.java.instrumentation.shared.ClassName;
-import org.evomaster.client.java.instrumentation.shared.ObjectiveNaming;
-import org.evomaster.client.java.instrumentation.staticstate.ObjectiveRecorder;
 import org.objectweb.asm.ClassVisitor;
 import org.objectweb.asm.MethodVisitor;
 
@@ -11,7 +10,7 @@ import org.objectweb.asm.MethodVisitor;
  * Visitor for classes that are not SUT, eg third-party libraries.
  * We do not want to create coverage targets for them, but might still
  * do some instrumentations
- *
+ * <p>
  * Created by arcuri82 on 06-Sep-19.
  */
 public class ThirdPartyClassVisitor extends ClassVisitor {
@@ -42,6 +41,8 @@ public class ThirdPartyClassVisitor extends ClassVisitor {
         }
 
         mv = new MethodReplacementMethodVisitor(false, mv, bytecodeClassName, name, descriptor);
+
+        mv = new CallbackMethodVisitor(mv, bytecodeClassName, name, descriptor);
 
         return mv;
     }
