+++ client-java/controller/src/main/java/org/evomaster/client/java/controller/internal/db/MongoHandler.java
@@ -1,0 +1,12 @@
+package org.evomaster.client.java.controller.internal.db;
+
+public class MongoHandler {
+
+    public void handleFind() {
+
+    }
+
+    public void reset() {
+
+    }
+}
