### 2020-01-13 666eaf7
mongo method replacements are stored in a mapping owner->name->desc->method
