### 2020-01-14 2d219ca
Tweaks to clustering. Clustering should only be done on actions that have a 500 code (i.e. we are only clustering those individuals that have a least an action with a 500 code and only based on those actions that do have a 500 code)
This means that the clusters should be based exclusively on the message
in failed actions.
