### 2020-01-15 23810ec
Refactoring away usedObj.
