### 2020-01-15 bbff5a8
MongoHandler processes the MongoOperation and compute corresponding distances
