+++ client-java/instrumentation/src/main/java/org/evomaster/client/java/instrumentation/mongo/MongoCollectionClassReplacement.java
@@ -3,7 +3,6 @@ package org.evomaster.client.java.instrumentation.mongo;
 import com.mongodb.client.ClientSession;
 import com.mongodb.client.FindIterable;
 import com.mongodb.client.MongoCollection;
-import org.bson.BsonDocument;
 import org.bson.conversions.Bson;
 import org.evomaster.client.java.instrumentation.coverage.methodreplacement.MethodReplacementClass;
 
@@ -15,21 +14,15 @@ public class MongoCollectionClassReplacement implements MethodReplacementClass {
                 MongoCollection.class;
     }
 
-    public static FindIterable<?> find(MongoCollection mongoCollection, Bson bson) {
-        FindIterable<?> rv = mongoCollection.find(bson);
-
-        BsonDocument query = bson.toBsonDocument(BsonDocument.class, mongoCollection.getCodecRegistry());
-        boolean queryHasDocuments = mongoCollection.find(bson).iterator().hasNext();
-
-
+    public static FindIterable<?> find(MongoCollection mongoCollection, Bson bson, Class<?> documentClass) {
+        FindIterable<?> rv = mongoCollection.find(bson, documentClass);
+        MongoLogger.getInstance().logFind(mongoCollection, bson);
         return rv;
     }
 
-    public static FindIterable<?> find(MongoCollection mongoCollection, Bson bson, Class<?> documentClass) {
-        return mongoCollection.find(bson, documentClass);
-    }
-
     public static FindIterable<?> find(MongoCollection mongoCollection, ClientSession clientSession, Bson bson, Class<?> documentClass) {
-        return mongoCollection.find(clientSession, bson, documentClass);
+        FindIterable<?> rv = mongoCollection.find(clientSession, bson, documentClass);
+        MongoLogger.getInstance().logFind(mongoCollection, bson);
+        return rv;
     }
 }
