+++ client-java/instrumentation/src/main/java/org/evomaster/client/java/instrumentation/coverage/ThirdPartyClassVisitor.java
@@ -1,7 +1,7 @@
 package org.evomaster.client.java.instrumentation.coverage;
 
 import org.evomaster.client.java.instrumentation.Constants;
-import org.evomaster.client.java.instrumentation.mongo.MongoMethodReplacementMethodVisitor;
+import org.evomaster.client.java.instrumentation.mongo.MongoReplacementMethodVisitor;
 import org.evomaster.client.java.instrumentation.shared.ClassName;
 import org.objectweb.asm.ClassVisitor;
 import org.objectweb.asm.MethodVisitor;
@@ -21,6 +21,7 @@ public class ThirdPartyClassVisitor extends ClassVisitor {
         super(Constants.ASM, cv);
         bytecodeClassName = className.getBytecodeName();
     }
+
 
     @Override
     public MethodVisitor visitMethod(int methodAccess,
@@ -42,7 +43,7 @@ public class ThirdPartyClassVisitor extends ClassVisitor {
 
         mv = new MethodReplacementMethodVisitor(false, mv, bytecodeClassName, name, descriptor);
 
-        mv = new MongoMethodReplacementMethodVisitor(mv, bytecodeClassName, name, descriptor);
+        mv = new MongoReplacementMethodVisitor(mv, bytecodeClassName, name, descriptor);
 
         return mv;
     }
