+++ client-java/controller-api/src/main/java/org/evomaster/client/java/controller/api/dto/HeuristicEntryDto.java
@@ -9,7 +9,7 @@ public class HeuristicEntryDto {
      * The type of extra heuristic.
      * Note: for the moment, we only have heuristics on SQL commands
      */
-    public enum Type {SQL}
+    public enum Type {SQL, MONGO_DB}
 
     /**
      * Should we try to minimize or maximize the heuristic?

[…]
