+++ client-js/evomaster-client-js/src/instrumentation/staticstate/ExecutionTracer.ts
@@ -6,6 +6,7 @@ import TargetInfo from "../TargetInfo";
 import Action from "../Action";
 import ObjectiveNaming from "../ObjectiveNaming";
 import AdditionalInfoDto from "../../controller/api/dto/AdditionalInfoDto";
+import AdditionalInfo from "../AdditionalInfo";
 
 export default class ExecutionTracer {
 
@@ -20,7 +21,7 @@ export default class ExecutionTracer {
     /**
      * Key -> the unique descriptive id of the coverage objective
      */
-    private static readonly  objectiveCoverage : Map<String, TargetInfo> = new Map<>();
+    private static readonly  objectiveCoverage : Map<string, TargetInfo> = new Map<>();
 
     /**
      * A test case can be composed by 1 or more actions, eg HTTP calls.
@@ -40,14 +41,14 @@ export default class ExecutionTracer {
      * keep track during test execution.
      * We keep track of it separately for each action
      */
-    private static readonly  additionalInfoList : Array<AdditionalInfoDto> = [];
+    private static readonly  additionalInfoList : Array<AdditionalInfo> = [];
 
 
     static  reset() {
         ExecutionTracer.objectiveCoverage.clear();
         ExecutionTracer.actionIndex = 0;
         ExecutionTracer.additionalInfoList.length = 0;
-        ExecutionTracer.additionalInfoList.push(new AdditionalInfoDto());
+        ExecutionTracer.additionalInfoList.push(new AdditionalInfo());
         ExecutionTracer.inputVariables = new Set<>();
     }
 
@@ -55,7 +56,7 @@ export default class ExecutionTracer {
     static setAction(action: Action){
         if(action.getIndex() != ExecutionTracer.actionIndex) {
             ExecutionTracer.actionIndex = action.getIndex();
-            //additionalInfoList.add(new AdditionalInfo());
+            ExecutionTracer.additionalInfoList.add(new AdditionalInfo());
         }
 
         if(action.getInputVariables() && action.getInputVariables().size > 0){
@@ -94,7 +95,7 @@ export default class ExecutionTracer {
     // }
 
 
-    static  exposeAdditionalInfoList() : Array<AdditionalInfoDto>{
+    static  exposeAdditionalInfoList() : Array<AdditionalInfo>{
         return ExecutionTracer.additionalInfoList;
     }
 
@@ -112,36 +113,34 @@ export default class ExecutionTracer {
 
 
 
- static markLastExecutedStatement(lastLine:string, lastMethod: string){
-    ExecutionTracer.additionalInfoList[ExecutionTracer.actionIndex]
-        .pushLastExecutedStatement(lastLine, lastMethod);
-}
+    public static markLastExecutedStatement(lastLine:string, lastMethod: string){
+        ExecutionTracer.additionalInfoList[ExecutionTracer.actionIndex]
+            .pushLastExecutedStatement(lastLine, lastMethod);
+    }
 
-public static final String COMPLETED_LAST_EXECUTED_STATEMENT_NAME = "completedLastExecutedStatement";
-public static final String COMPLETED_LAST_EXECUTED_STATEMENT_DESCRIPTOR = "()V";
 
-public static void completedLastExecutedStatement(){
-    additionalInfoList.get(actionIndex).popLastExecutedStatement();
-}
+    public static completedLastExecutedStatement(){
+        ExecutionTracer.additionalInfoList[ExecutionTracer.actionIndex].popLastExecutedStatement();
+    }
 
-public static Map<String, TargetInfo> getInternalReferenceToObjectiveCoverage() {
-    return objectiveCoverage;
-}
+    public static  getInternalReferenceToObjectiveCoverage() : Map<String, TargetInfo> {
+        return ExecutionTracer.objectiveCoverage;
+    }
 
 /**
[…]
