+++ client-js/evomaster-client-js/src/instrumentation/AdditionalInfo.ts
@@ -1,0 +1,128 @@
+class StatementDescription {
+    readonly line: string;
+    readonly method: string;
+
+    constructor(line: string, method: string) {
+        this.line = line;
+        this.method = method;
+    }
+}
+
+/**
+ * Besides code coverage, there can be additional info that we want
+ * to collect at runtime when test cases are executed.
+ */
+export default class AdditionalInfo {
+
+    /**
+     * In REST APIs, it can happen that some query parameters do not
+     * appear in the schema if they are indirectly accessed via
+     * objects like WebRequest.
+     * But we can track at runtime when such kind of objects are used
+     * to access the query parameters
+     */
+    private queryParameters: Set<string> = new Set<string>();
+
+
+    /**
+     * In REST APIs, it can happen that some HTTP headers do not
+     * appear in the schema if they are indirectly accessed via
+     * objects like WebRequest.
+     * But we can track at runtime when such kind of objects are used
+     * to access the query parameters
+     */
+    private headers: Set<string> = new Set<string>();
+
+    /**
+     * Map from taint input name to string specializations for it
+     */
+    //private Map<String, Set<StringSpecializationInfo>> stringSpecializations = new ConcurrentHashMap<>();
+
+
+    /**
+     * Keep track of the last executed statement done in the SUT.
+     * But not in the third-party libraries, just the business logic of the SUT.
+     * The statement is represented with a descriptive unique id, like the class name and line number.
+     *
+     * We need to use a stack to handle method call invocations, as we can know when a statement
+     * starts, but not so easily when it ends.
+     */
+    private lastExecutedStatementStack: Array<StatementDescription> = new Array<StatementDescription>();
+
+    /**
+     * In case we pop all elements from stack, keep track of last one separately.
+     */
+    private noExceptionStatement: StatementDescription = null;
+
+
+// public void addSpecialization(String taintInputName, StringSpecializationInfo info){
+//     if(!ExecutionTracer.getTaintType(taintInputName).isTainted()){
+//         throw new IllegalArgumentException("No valid input name: " + taintInputName);
+//     }
+//     Objects.requireNonNull(info);
+//
+//     stringSpecializations.putIfAbsent(taintInputName, new CopyOnWriteArraySet<>());
+//     Set<StringSpecializationInfo> set = stringSpecializations.get(taintInputName);
+//     set.add(info);
+// }
+//
+// public Map<String, Set<StringSpecializationInfo>> getStringSpecializationsView(){
+//     //note: this does not prevent modifying the sets inside it
+//     return Collections.unmodifiableMap(stringSpecializations);
+// }
+
+    public addQueryParameter(param: string) {
+        if (param && param.length > 0) {
+            this.queryParameters.add(param);
+        }
+    }
+
+    public getQueryParametersView(): Set<string> {
+        return this.queryParameters
+    }
+
+    public addHeader(header: string) {
+        if (header && header.length > 0) {
+            this.headers.add(header);
+        }
+    }
+
+    public getHeadersView(): Set<string> {
+        return this.headers;
+    }
+
+    public getLastExecutedStatement(): string {
+
+        if (this.lastExecutedStatementStack.length == 0) {
+            if (!this.noExceptionStatement) {
+                return null;
+            }
+            return this.noExceptionStatement.line;
+        }
+
+        let current: StatementDescription = this.lastExecutedStatementStack[this.lastExecutedStatementStack.length - 1];
+        return current.line;
+    }
+
+    public pushLastExecutedStatement(lastLine: string, lastMethod: string) {
+
+        this.noExceptionStatement = null;
+
+        let statement = new StatementDescription(lastLine, lastMethod);
[…]
