+++ client-java/instrumentation/src/main/java/org/evomaster/client/java/instrumentation/staticstate/ObjectiveRecorder.java
@@ -182,7 +182,7 @@ public class ObjectiveRecorder {
      * @param descriptiveId of the objective/target
      * @param value         of the coverage heuristic, in [0,1]
      */
-    static void update(String descriptiveId, double value) {
+    public static void update(String descriptiveId, double value) {
 
         Objects.requireNonNull(descriptiveId);
         if (value < 0d || value > 1) {
