+++ client-js/evomaster-client-js/src/instrumentation/staticstate/ExecutionTracer.ts
@@ -5,7 +5,6 @@
 import TargetInfo from "../TargetInfo";
 import Action from "../Action";
 import ObjectiveNaming from "../ObjectiveNaming";
-import AdditionalInfoDto from "../../controller/api/dto/AdditionalInfoDto";
 import AdditionalInfo from "../AdditionalInfo";
 
 export default class ExecutionTracer {
@@ -21,30 +20,30 @@ export default class ExecutionTracer {
     /**
      * Key -> the unique descriptive id of the coverage objective
      */
-    private static readonly  objectiveCoverage : Map<string, TargetInfo> = new Map<string, TargetInfo>();
+    private static readonly objectiveCoverage: Map<string, TargetInfo> = new Map<string, TargetInfo>();
 
     /**
      * A test case can be composed by 1 or more actions, eg HTTP calls.
      * When we get the best distance for a testing target, we might
      * also want to know which action in the test led to it.
      */
-    private static actionIndex : number = 0;
+    private static actionIndex: number = 0;
 
     /**
      * A set of possible values used in the tests, needed for some kinds
      * of taint analyses
      */
-    private static inputVariables : Set<string> = new Set();
+    private static inputVariables: Set<string> = new Set();
 
     /**
      * Besides code coverage, there might be other events that we want to
      * keep track during test execution.
      * We keep track of it separately for each action
      */
-    private static readonly  additionalInfoList : Array<AdditionalInfo> = [];
+    private static readonly additionalInfoList: Array<AdditionalInfo> = [];
 
 
-    static  reset() {
+    static reset() {
         ExecutionTracer.objectiveCoverage.clear();
         ExecutionTracer.actionIndex = 0;
         ExecutionTracer.additionalInfoList.length = 0;
@@ -53,24 +52,24 @@ export default class ExecutionTracer {
     }
 
 
-    static setAction(action: Action){
-        if(action.getIndex() != ExecutionTracer.actionIndex) {
+    static setAction(action: Action) {
+        if (action.getIndex() != ExecutionTracer.actionIndex) {
             ExecutionTracer.actionIndex = action.getIndex();
             ExecutionTracer.additionalInfoList.push(new AdditionalInfo());
         }
 
-        if(action.getInputVariables() && action.getInputVariables().size > 0){
+        if (action.getInputVariables() && action.getInputVariables().size > 0) {
             ExecutionTracer.inputVariables = action.getInputVariables();
         }
     }
 
     /**
-    * Check if the given input represented a tainted value from the test cases.
-    * This could be based on static info of the input (eg, according to a precise
-    * name convention given by TaintInputName), or dynamic info given directly by
-    * the test itself (eg, the test at action can register a list of values to check
-    * for)
-    */
+     * Check if the given input represented a tainted value from the test cases.
+     * This could be based on static info of the input (eg, according to a precise
+     * name convention given by TaintInputName), or dynamic info given directly by
+     * the test itself (eg, the test at action can register a list of values to check
+     * for)
+     */
     // public static boolean isTaintInput(String input){
     //     return TaintInputName.isTaintInput(input) || inputVariables.contains(input);
     // }
@@ -95,7 +94,7 @@ export default class ExecutionTracer {
     // }
 
 
-    static  exposeAdditionalInfoList() : Array<AdditionalInfo>{
+    static exposeAdditionalInfoList(): Array<AdditionalInfo> {
         return ExecutionTracer.additionalInfoList;
     }
 
@@ -112,28 +111,27 @@ export default class ExecutionTracer {
     // }
 
 
-
-    public static markLastExecutedStatement(lastLine:string, lastMethod: string){
[…]
