### 2020-01-19 c9b574d
JS basic coverage instrumentation
