+++ client-js/evomaster-client-js/tests/unit/instrumentation/babel-plugin-evomaster-test.ts
@@ -26,11 +26,11 @@ test("simple", () => {
         
         const __EM__ = require("evomaster-client-js").InjectedFunctions;
         
-        __EM__.enteringStatement(foo, 1, 0);
+        __EM__.enteringStatement("/test.ts", 1, 0);
         
         let x = 0;
         
-        __EM__.completedStatement(foo, 1, 0);
+        __EM__.completedStatement("/test.ts", 1, 0);
     `);
 });
 
@@ -46,17 +46,17 @@ test("simple multi assignment, same line", () => {
         
         const __EM__ = require("evomaster-client-js").InjectedFunctions;
 
-        __EM__.enteringStatement(foo, 1, 0);
+        __EM__.enteringStatement("/test.ts", 1, 0);
         
         let x = 0;
         
-        __EM__.completedStatement(foo, 1, 0);
+        __EM__.completedStatement("/test.ts", 1, 0);
         
-        __EM__.enteringStatement(foo, 1, 1);
+        __EM__.enteringStatement("/test.ts", 1, 1);
         
         let y = 0;
         
-        __EM__.completedStatement(foo, 1, 1);
+        __EM__.completedStatement("/test.ts", 1, 1);
     `);
 });
 
@@ -81,34 +81,34 @@ test("simple multi lines", () => {
         
         const __EM__ = require("evomaster-client-js").InjectedFunctions;
 
-        __EM__.enteringStatement(foo, 1, 0);
+        __EM__.enteringStatement("/test.ts", 1, 0);
         
         let x = 0;
         
-        __EM__.completedStatement(foo, 1, 0);
+        __EM__.completedStatement("/test.ts", 1, 0);
         
-        __EM__.enteringStatement(foo, 3, 1);
+        __EM__.enteringStatement("/test.ts", 3, 1);
         
         let y = 1;
         
-        __EM__.completedStatement(foo, 3, 1);
+        __EM__.completedStatement("/test.ts", 3, 1);
         
-        __EM__.enteringStatement(foo, 5, 2);
+        __EM__.enteringStatement("/test.ts", 5, 2);
         
         const sum = function (a, b) {
-          __EM__.enteringStatement(foo, 6, 3);
+          __EM__.enteringStatement("/test.ts", 6, 3);
 
           return a + b;
         
-          __EM__.completedStatement(foo, 6, 3);
+          __EM__.completedStatement("/test.ts", 6, 3);
         };
     
-        __EM__.completedStatement(foo, 5, 2);
+        __EM__.completedStatement("/test.ts", 5, 2);
         
-        __EM__.enteringStatement(foo, 9, 4);
+        __EM__.enteringStatement("/test.ts", 9, 4);
         
         const k = sum(x, y);
         
-        __EM__.completedStatement(foo, 9, 4);
+        __EM__.completedStatement("/test.ts", 9, 4);
     `);
 });
