+++ client-js/evomaster-client-js/tests/unit/instrumentation/babel-code-test.ts
@@ -34,6 +34,46 @@ test("simple", () => {
     `);
 });
 
+// test("expressions", () => {
+//
+//     const code = dedent`
+//         let x = 0;
+//         x;
+//         "foo";
+//         5;
+//         bar.foo();
+//     `;
+//
+//     const res = runPlugin(code);
+//     expect(res.code).toEqual(dedent`
+//         //File instrumented with EvoMaster
+//
+//         const __EM__ = require("evomaster-client-js").InjectedFunctions;
+//
+//         __EM__.enteringStatement("/test.ts", 1, 0);
+//
+//         let x = 0;
+//
+//         __EM__.completedStatement("/test.ts", 1, 0);
+//
+//         __EM__.enteringStatement("/test.ts", 2, 1);
+//
+//         __EM__.completingStatement(x, "/test.ts", 2, 1);
+//
+//         __EM__.enteringStatement("/test.ts", 3, 2);
+//
+//         __EM__.completingStatement("foo", "/test.ts", 3, 2);
+//
+//         __EM__.enteringStatement("/test.ts", 4, 3);
+//
+//         __EM__.completingStatement(5, "/test.ts", 4, 3);
+//
+//         __EM__.enteringStatement("/test.ts", 5, 4);
+//
+//         __EM__.completingStatement(bar.foo(), "/test.ts", 5, 4);
+//     `);
+// });
+
 
 test("simple multi assignment, same line", () => {
     const code = dedent`
@@ -98,9 +138,7 @@ test("simple multi lines", () => {
         const sum = function (a, b) {
           __EM__.enteringStatement("/test.ts", 6, 3);
 
-          return a + b;
-        
-          __EM__.completedStatement("/test.ts", 6, 3);
+          return __EM__.completingStatement(a + b, "/test.ts", 6, 3);
         };
     
         __EM__.completedStatement("/test.ts", 5, 2);
