### 2020-01-20 #2001 @arcuri82
Hi,
 yes, I can provide any further documentation that is needed. But, there is already quite a bit of documentation (including a demo video of using the tool). To improve it, I will need some more info on which parts in the documentation are unclear, and know where @mado89 got stuck with the instructions
