### 2020-01-20 19dfef2
proper type info for index.ts
