+++ client-js/integration-tests/tests/snippets/babel-eval-test.js
@@ -1,5 +1,8 @@
 const {transform} = require("@babel/core");
 const dedent = require("dedent");
+const ET = require("evomaster-client-js").internal.ExecutionTracer;
+const OR = require("evomaster-client-js").internal.ObjectiveRecorder;
+
 
 function runPlugin(code) {
     const res = transform(code, {
@@ -16,9 +19,32 @@ function runPlugin(code) {
 }
 
 
-test("simple", () => {
+beforeEach(()=>{
+    ET.reset();
+    ET.actionIndex
+});
 
 
+test("simple, no side effect", () => {
+
+    let x = 0;
+
+    const code = dedent`
+        let k = 42;
+        x = k;
+    `;
+
+    const instrumented = runPlugin(code).code;
+
+    const i = eval(instrumented);
+
+    expect(x).toBe(42);
+});
+
+
+test("simple block", () => {
+
+    expect(ET)
 
     let x = 0;
 
