+++ client-java/instrumentation/src/test/java/com/foo/somedifferentpackage/examples/methodreplacement/TestabilityExcImp.java
@@ -147,4 +147,5 @@ public class TestabilityExcImp implements TestabilityExc {
         return matcher.matches();
     }
 
+
 }

[…]
