### 2020-01-20 616842b
added new classes to skipInstrumentationList.txt
