### 2020-01-20 75fae67
test cases for MongoLogger
