+++ client-java/instrumentation/src/main/java/org/evomaster/client/java/instrumentation/coverage/ThirdPartyClassVisitor.java
@@ -1,7 +1,6 @@
 package org.evomaster.client.java.instrumentation.coverage;
 
 import org.evomaster.client.java.instrumentation.Constants;
-import org.evomaster.client.java.instrumentation.mongo.MongoReplacementMethodVisitor;
 import org.evomaster.client.java.instrumentation.shared.ClassName;
 import org.objectweb.asm.ClassVisitor;
 import org.objectweb.asm.MethodVisitor;
@@ -42,8 +41,6 @@ public class ThirdPartyClassVisitor extends ClassVisitor {
         }
 
         mv = new MethodReplacementMethodVisitor(false, mv, bytecodeClassName, name, descriptor);
-
-        mv = new MongoReplacementMethodVisitor(mv, bytecodeClassName, name, descriptor);
 
         return mv;
     }
