### 2020-01-20 b79a238
test cases for MongoReplacementClass
