+++ e2e-tests/spring-rest-mongo/src/test/java/com/foo/mongo/MyMongoAppEmbeddedController.java
@@ -1,0 +1,15 @@
+package com.foo.mongo;
+
+import com.foo.spring.rest.mongo.SpringRestMongoController;
+
+public class MyMongoAppEmbeddedController extends SpringRestMongoController {
+
+    public MyMongoAppEmbeddedController() {
+        super(MyMongoApp.class);
+    }
+
+    @Override
+    public String getPackagePrefixesToCover() {
+        return "com.foo.customer";
+    }
+}
