+++ e2e-tests/spring-rest-mongo/src/main/java/com/foo/mongo/MyMongoApp.java
@@ -1,4 +1,4 @@
-package com.foo.customer;
+package com.foo.mongo;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -7,10 +7,10 @@ import springfox.documentation.swagger2.annotations.EnableSwagger2;
 
 @EnableSwagger2
 @SpringBootApplication(exclude = SecurityAutoConfiguration.class)
-public class CustomerApplication {
+public class MyMongoApp {
 
     public static void main(String[] args) {
-        SpringApplication.run(CustomerApplication.class, args);
+        SpringApplication.run(MyMongoApp.class, args);
     }
 
 }
