### 2020-01-20 527472b
moving logic for MongoDB heuristics to Core
