+++ client-java/controller-api/src/main/java/org/evomaster/client/java/controller/api/dto/ExtraHeuristicsDto.java
@@ -1,6 +1,7 @@
 package org.evomaster.client.java.controller.api.dto;
 
 import org.evomaster.client.java.controller.api.dto.database.execution.ExecutionDto;
+import org.evomaster.client.java.controller.api.dto.database.execution.MongoExecutionDto;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -19,4 +20,6 @@ public class ExtraHeuristicsDto {
     public List<HeuristicEntryDto> heuristics = new ArrayList<>();
 
     public ExecutionDto databaseExecutionDto;
+
+    public MongoExecutionDto mongoExecutionDto;
 }
