### 2020-01-20 f3f94c8
test and fixed issue for line coverage instrumentation
