+++ client-js/integration-tests/tests/snippets/babel-eval-test.js
@@ -2,7 +2,7 @@ const {transform} = require("@babel/core");
 const dedent = require("dedent");
 const ET = require("evomaster-client-js").internal.ExecutionTracer;
 const OR = require("evomaster-client-js").internal.ObjectiveRecorder;
-
+const ON = require("evomaster-client-js").internal.ObjectiveNaming;
 
 function runPlugin(code) {
     const res = transform(code, {
@@ -21,7 +21,6 @@ function runPlugin(code) {
 
 beforeEach(()=>{
     ET.reset();
-    ET.actionIndex
 });
 
 
@@ -44,7 +43,9 @@ test("simple, no side effect", () => {
 
 test("simple block", () => {
 
-    expect(ET)
+    expect(ET.getNumberOfObjectives()).toBe(0);
+    expect(ET.getNumberOfObjectives(ON.LINE)).toBe(0);
+
 
     let x = 0;
 
@@ -58,4 +59,7 @@ test("simple block", () => {
     const i = eval(instrumented);
 
     expect(x).toBe(42);
+
+    expect(ET.getNumberOfObjectives()).toBe(3); // 2 lines and 1 file
+    expect(ET.getNumberOfObjectives(ON.LINE)).toBe(2);
 });
