### 2020-01-21 80822d0
A bit of refactoring in preparation for the review.
