### 2020-01-21 d2d8086
Minor update. A bit of refactoring in preparation for the review.
