### 2020-01-21 ccab661
fix for coverage instrumentation
