### 2020-01-21 b07729b
processing mongo operations on core process
