+++ client-java/instrumentation/src/main/java/org/evomaster/client/java/instrumentation/mongo/MongoOperation.java
@@ -1,4 +1,0 @@
-package org.evomaster.client.java.instrumentation.mongo;
-
-public class MongoOperation {
-}
