+++ client-java/controller-api/src/main/java/org/evomaster/client/java/controller/api/dto/database/execution/MongoOperationDto.java
@@ -1,0 +1,11 @@
+package org.evomaster.client.java.controller.api.dto.database.execution;
+
+public class MongoOperationDto {
+
+    public enum Type {MONGO_FIND}
+
+    public Type operationType;
+
+    public String operationJsonStr;
+
+}
