+++ client-java/instrumentation-shared/src/main/java/org/evomaster/client/java/instrumentation/shared/mongo/MongoOperation.java
@@ -1,0 +1,4 @@
+package org.evomaster.client.java.instrumentation.shared.mongo;
+
+public abstract class MongoOperation {
+}
