+++ e2e-tests/spring-examples/src/main/java/com/foo/rest/examples/spring/bodytypes/BodyTypesRest.java
@@ -60,21 +60,20 @@ public class BodyTypesRest {
     }
 
 //    @PostMapping(value = "/api/bodytypes/q", consumes = "application/x-www-form-urlencoded")
-//    public int formQ(@ModelAttribute("value") int value){
-//        System.out.println("Q: Form");
+//    public int formQ(Integer value){
 //        return 8;
 //    }
 
 
-//    @PostMapping(value = "/api/bodytypes/r", consumes = "application/x-www-form-urlencoded")
-//    public int formR(@ModelAttribute("value") int value) {
-//        return 9;
-//    }
-//
-//    @PostMapping(value = "/api/bodytypes/t", consumes = "application/x-www-form-urlencoded")
-//    public int formT(@ModelAttribute("a") int a, @ModelAttribute("b") int b) {
-//        return 10;
-//    }
+    @PostMapping(value = "/api/bodytypes/r", consumes = "application/x-www-form-urlencoded")
+    public int formR(Integer value) {
+        return 9;
+    }
+
+    @PostMapping(value = "/api/bodytypes/t", consumes = "application/x-www-form-urlencoded")
+    public int formT(Integer a,  Integer b) {
+        return 10;
+    }
 
 
 }
