+++ client-java/controller/src/test/java/org/evomaster/client/java/controller/internal/db/mongo/MongoHandlerTest.java
@@ -1,105 +1,42 @@
 package org.evomaster.client.java.controller.internal.db.mongo;
 
-import com.mongodb.client.MongoCollection;
-import org.bson.BsonDocument;
-import org.bson.Document;
-import org.evomaster.client.java.controller.InstrumentedSutStarter;
-import org.evomaster.client.java.instrumentation.mongo.MongoLogger;
-import org.junit.jupiter.api.*;
+import org.evomaster.client.java.controller.api.dto.database.execution.MongoExecutionDto;
+import org.evomaster.client.java.controller.api.dto.database.execution.MongoOperationDto;
+import org.junit.jupiter.api.Test;
 
-import static org.evomaster.client.java.controller.api.ControllerConstants.BASE_PATH;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assertions.*;
 
-public class MongoHandlerTest extends MongoTestTemplate {
-
-    private static InstrumentedSutStarter starter;
-    private static MongoFakeSutController sutController;
-    private static String url;
-
-    @BeforeAll
-    public static void startController() {
-        sutController = buildSutController();
-        starter = new InstrumentedSutStarter(sutController);
-        url = start(starter);
-        url += BASE_PATH;
-    }
-
-    @BeforeEach
-    public void dropMongoDatabase() {
-        mongoClient.getDatabase("testdb").drop();
-        sutController.initMongoClient();
-    }
-
-    @AfterEach
-    public void resetMongoHandler() {
-        sutController.getMongoHandler().reset();
-    }
-
-    @AfterAll
-    public static void stopController() {
-        if (starter != null) {
-            starter.stop();
-        }
-    }
-
+public class MongoHandlerTest {
 
     @Test
-    public void testEmptyDistances() {
-        startNewTest(url);
-        assertTrue(sutController.getMongoHandler().getMongoExecutionDto().mongoOperations.isEmpty());
+    public void testIsExtractMongoExecution() {
+        MongoHandler mongoHandler = new MongoHandler();
+        assertFalse(mongoHandler.isExtractMongoExecution());
+        mongoHandler.setExtractMongoExecution(true);
+        assertTrue(mongoHandler.isExtractMongoExecution());
     }
 
     @Test
-    public void testFindNoResult() {
-        startNewTest(url);
-        assertTrue(sutController.getMongoHandler().getMongoExecutionDto().mongoOperations.isEmpty());
-
-        MongoCollection<Document> mongoCollection = mongoClient.getDatabase("testdb").getCollection("customers");
-
-        // find({}) returns no element
-        MongoLogger.getInstance().logFind(mongoCollection, new BsonDocument());
-
-        assertEquals(1, sutController.getMongoHandler().getMongoExecutionDto().mongoOperations.size());
+    public void testHandle() {
+        MongoHandler mongoHandler = new MongoHandler();
+        mongoHandler.setExtractMongoExecution(true);
+        mongoHandler.handle("{\"databaseName\":\"mydb\",\"collectionName\":\"mycollection\",\"query\":{}}\n");
+        MongoExecutionDto dto = mongoHandler.getMongoExecutionDto();
+        assertEquals(1, dto.mongoOperations.size());
+        assertEquals(MongoOperationDto.Type.MONGO_FIND, dto.mongoOperations.get(0).operationType);
+        assertEquals("{\"databaseName\":\"mydb\",\"collectionName\":\"mycollection\",\"query\":{}}\n", dto.mongoOperations.get(0).operationJsonStr);
     }
-
 
     @Test
-    public void testTwoFindOperations() {
[…]
