+++ client-js/evomaster-client-js/src/instrumentation/heuristic/HeuristicsForBooleans.ts
@@ -1,0 +1,7 @@
+
+
+export default class HeuristicsForBooleans{
+
+
+
+}
