### 2020-01-23 3b7b64f
more utility functions on distances
