+++ e2e-tests/e2e-tests-utils/src/test/java/org/evomaster/e2etests/utils/RestTestBase.java
@@ -216,7 +216,7 @@ public abstract class RestTestBase {
         PrintWriter pw = new PrintWriter(writer);
 
         TestExecutionSummary summary = JUnitTestRunner.runTestsInClass(klass);
-        summary.printFailuresTo(pw);
+        summary.printFailuresTo(pw, 100);
         String failures = writer.toString();
 
         assertTrue(summary.getContainersFoundCount() > 0);
