### 2020-01-28 a13ff22
adding more tests for branch distance. refactored && implementation
