+++ client-js/evomaster-client-js/tests/unit/instrumentation/heuristic/HeuristicsForBooleans-test.ts
@@ -1,5 +1,10 @@
 import h from "../../../../src/instrumentation/heuristic/HeuristicsForBooleans";
+import Truthness from "../../../../src/instrumentation/heuristic/Truthness";
+import HeuristicsForBooleans from "../../../../src/instrumentation/heuristic/HeuristicsForBooleans";
 
+beforeEach(() => {
+    h.clearLastEvaluation();
+});
 
 test("=== numbers", () => {
 
@@ -9,6 +14,14 @@ test("=== numbers", () => {
     expect(t.getOfFalse()).toBeLessThan(1);
 });
 
+
+test("=== strings", () => {
+
+    const t = h.compare("foo", "===", "foo");
+
+    expect(t.getOfTrue()).toBe(1);
+    expect(t.getOfFalse()).toBeLessThan(1);
+});
 
 test("=== numbers, same distance", () => {
 
@@ -38,7 +51,346 @@ test("=== numbers, gradient", () => {
     expect(c.getOfTrue()).toBeLessThan(d.getOfTrue());
 });
 
+test("=== strings, gradient", () => {
+
+    const a = h.compare("aaaaaaaaaaaaaaaaaaaaa", "===", "foo");
+    const b = h.compare("", "===", "foo");
+    const c = h.compare("a", "===", "foo");
+    const d = h.compare("f", "===", "foo");
+    const e = h.compare("fo", "===", "foo");
+
+    expect(a.getOfFalse()).toBe(1);
+    expect(b.getOfFalse()).toBe(1);
+    expect(c.getOfFalse()).toBe(1);
+    expect(d.getOfFalse()).toBe(1);
+    expect(e.getOfFalse()).toBe(1);
+
+
+    expect(a.getOfTrue()).toBeLessThan(b.getOfTrue());
+    expect(b.getOfTrue()).toBeLessThan(c.getOfTrue());
+    expect(c.getOfTrue()).toBeLessThan(d.getOfTrue());
+    expect(d.getOfTrue()).toBeLessThan(e.getOfTrue());
+});
+
+test("== and === mixed types", () =>{
+
+    const a = h.compare("42", "===", 42);
+    const b = h.compare(42, "===", "42");
+    const c = h.compare("42", "==", 42);
+    const d = h.compare(42, "==", "42");
+
+    expect(a.getOfFalse()).toBe(1);
+    expect(b.getOfFalse()).toBe(1);
+    expect(c.getOfTrue()).toBe(1);
+    expect(d.getOfTrue()).toBe(1);
+});
+
+
+test("== same types", () =>{
+
+    const a = h.compare("42", "==", "42");
+    const b = h.compare(42, "==", 42);
+
+    expect(a.getOfTrue()).toBe(1);
+    expect(b.getOfTrue()).toBe(1);
+});
+
+test("== not handled types", () =>{
+
+    const a = h.compare("42", "==", []);
+    const b = h.compare(42, "==", {});
+
+    expect(a.getOfFalse()).toBe(1);
+    expect(b.getOfFalse()).toBe(1);
+});
+
+test("!==", () =>{
+    const t = h.compare(42, "!==", 66);
+
+    expect(t.getOfTrue()).toBe(1);
+    expect(t.getOfFalse()).toBeLessThan(1);
+});
+
+test("!= same type", () =>{
+    const t = h.compare(42, "!=", 66);
+
+    expect(t.getOfTrue()).toBe(1);
+    expect(t.getOfFalse()).toBeLessThan(1);
+});
+
+
+test("!== number, same distance", () =>{
+
+    const a = h.compare(44, "!==", 42);
+    const b = h.compare(-6, "!==", -4);
+
[…]
