### 2020-01-28 454faee
fixed bug in || && branch distances
