### 2020-01-30 #2001 @mado89
@arcuri82 sorry for not getting back to you until now. I see that the official review process has not started yet. Currently, I am under a lot of stress but I hope that starting from next week things will ease a little bit. Then I will start giving your software a thorough investigation and let you know where I need help.
@gkthiruvathukal whats the current status on the review process?
