+++ client-js/evomaster-client-js/tests/unit/instrumentation/heuristic/HeuristicsForBooleans-test.ts
@@ -506,13 +506,4 @@ test("&& left true, right exception", () =>{
 });
 
 
-test("", () =>{
-
-});
-
-test("", () =>{
-
-});
-
-
 
