### 2020-01-30 2e4f106
fixed bug in handling of null blocks
