### 2020-02-04 6cef647
fixing issues found while running EM on client-js
