### 2020-02-04 2ac60ed
fixed wrong type in OpenAPI
