### 2020-02-05 16f6dd8
fixed collection of unit info
