### 2020-02-06 d1ff37d
WiP. Corrected null checks, and updated some variable naming.
