+++ client-js/evomaster-client-js/src/instrumentation/methodreplacement/MethodReplacementClass.ts
@@ -1,0 +1,7 @@
+import ReplacementFunction from "./ReplacementFunction";
+
+
+export default abstract class MethodReplacementClass{
+
+    public abstract getReplacements() : Array<ReplacementFunction>;
+}
