### 2020-02-11 #2001 @gkthiruvathukal
@markgalassi. I was looking through our list of reviewers and saw that you have software engineering as an area of interest. Would you be willing to be the second reviewer for this JOSS submission? I need two reviewers (@mado89 has already agreed) in order to proceed.
