### 2020-02-17 d51b1eb
WiP. Initial e2e test for expectations.
Includes initial (and unsuccessful) attempt at v3 test. Leaving the files there,
to update later.
