+++ e2e-tests/spring-examples/src/test/java/com/foo/rest/examples/spring/expectations/ExpectationsController.java
@@ -1,0 +1,11 @@
+package com.foo.rest.examples.spring.expectations;
+
+import com.foo.rest.examples.spring.SpringController;
+import com.foo.rest.examples.spring.escape.EscapeController;
+
+public class ExpectationsController extends SpringController {
+
+    public ExpectationsController(){
+        super(ExpectationsApplication.class);
+    }
+}
