+++ e2e-tests/spring-examples/src/main/java/com/foo/rest/examples/spring/expectations/ExpectationsRest.java
@@ -1,0 +1,19 @@
+package com.foo.rest.examples.spring.expectations;
+
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+@RequestMapping(path = "/api/expectations")
+public class ExpectationsRest {
+    @RequestMapping(
+            value = "/getExpectations/{b}",
+            method = RequestMethod.GET
+    )
+    public String expectTest(
+            @PathVariable("b") boolean succeeded
+    ) throws IllegalArgumentException{
+        if (succeeded) return "Success is True!";
+        else throw new IllegalArgumentException("Input Was False");
+    }
+
+}
