### 2020-02-18 a1fc027
WiP. Quick update to ExpectationsEMTest (to include a failing test as well).
