### 2020-02-20 36ad573
refactoring how we deal with HTTP responses in the generated tests
