+++ e2e-tests/spring-examples/src/main/java/com/foo/rest/examples/spring/expectations/ExpectationsRest.java
@@ -16,4 +16,25 @@ public class ExpectationsRest {
         else throw new IllegalArgumentException("Input Was False");
     }
 
+    @RequestMapping(
+            value = "/basicResponsesString/{b}",
+            method = RequestMethod.GET
+    )
+    public String basicResponsesString(
+            @PathVariable("b") boolean succeeded
+    ) {
+        if (succeeded) return "Success is True!";
+        else return "What do you mean no string?";
+    }
+
+    @RequestMapping(
+            value = "/basicNumbersString/{b}",
+            method = RequestMethod.GET
+    )
+    public int basicResponsesNumber(
+            @PathVariable("b") boolean succeeded
+    ) {
+        if (succeeded) return 42;
+        else return -1;
+    }
 }
