### 2020-02-21 43e9fce
fixing issues with generation of JS tests
