### 2020-02-24 c662c27
removed incorrect check
