### 2020-02-27 bd8f014
Add test for select currval
