### 2020-02-27 4863207
updating exp.py example
