+++ scripts/exp.py
@@ -139,6 +139,21 @@ else:
     print("ERROR: target folder already exists")
     exit(1)
 
+
+JDK_8 = "JDK_8"
+JDK_11 = "JDK_11"
+JS = "JS"
+
+class Sut:
+    def __init__(self, name, timeWeight, platform):
+        self.name = name
+        # the higher value, the more time it will need compared to the other SUTS
+        self.timeWeight = timeWeight
+        # Java? JS? NodeJS
+        self.platform = platform
+
+
+
 ### We need different settings based on whether we are running the
 ### scripts on cluster or locally.
 if CLUSTER:
@@ -151,13 +166,13 @@ if CLUSTER:
     # of the SUTs (eg, by commenting them out)
 
     SUTS = [
-        ("features-service", 1),
-        ("scout-api", 2),
-        ("proxyprint", 2),
-        ("rest-ncs", 2),
-        ("rest-scs", 1),
-        ("rest-news", 1),
-        ("catwatch", 1)
+        Sut("features-service", 1, JDK_8),
+        Sut("scout-api", 2, JDK_8),
+        Sut("proxyprint", 2, JDK_8),
+        Sut("rest-ncs", 2, JDK_8),
+        Sut("rest-scs", 1, JDK_8),
+        Sut("rest-news", 1, JDK_8),
+        Sut("catwatch", 1, JDK_8)
     ]
 
     HOME = os.environ['HOME']
@@ -170,8 +185,10 @@ if CLUSTER:
 else:
     # These SUTs requires Docker
     SUTS = [
-        ("ind0", 1),
-        ("ocvn-rest", 1)
+        Sut("ind0", 1, JDK_8),
+        Sut("ocvn-rest", 1, JDK_8),
+        # Sut("ncs-js", 1, JS),
+        # Sut("scs-js", 1, JS)
     ]
 
     # You will need to define environment variables on your OS
@@ -239,13 +256,21 @@ if not CLUSTER:
 
     #Due to Windows limitations (ie crappy FS), we need to copy JARs over
     for sut in SUTS:
-        sut_name = sut[0]
-        shutil.copy(os.path.join(CASESTUDY_DIR, sut_name + EM_POSTFIX), BASE_DIR)
-        shutil.copy(os.path.join(CASESTUDY_DIR, sut_name + SUT_POSTFIX), BASE_DIR)
+        if sut.platform == JDK_8 or sut.platform == JDK_11:
+            # copy jar files
+            shutil.copy(os.path.join(CASESTUDY_DIR, sut.name + EM_POSTFIX), BASE_DIR)
+            shutil.copy(os.path.join(CASESTUDY_DIR, sut.name + SUT_POSTFIX), BASE_DIR)
+        elif sut.platform == JS:
+            # copy folders, which include both SUT and EM Controller
+            shutil.copytree(os.path.join(CASESTUDY_DIR, sut.name), os.path.join(BASE_DIR, sut.name))
+
     shutil.copy(os.path.join(CASESTUDY_DIR, AGENT), BASE_DIR)
     shutil.copy(os.path.join(EVOMASTER_DIR, "evomaster.jar"), BASE_DIR)
 
 
+
+# We could end up with many scripts, up to the max number of jobs we can run in parallel, eg. 400.
+# But those scripts still need to be submitted. So, we create a script to do that.
 def createRunallScript():
     script_path = BASE_DIR + "/runall.sh"
     script = open(script_path, "w")
@@ -267,8 +292,8 @@ def createRunallScript():
 
 
 
-def writeScript(code, port, sut_name):
-    script_path = SCRIPT_DIR + "/evomaster_" + str(port) + "_" + sut_name + ".sh"
[…]
