### 2020-02-28 360a8e3
preparing numeric comparison instrumentation
