+++ client-java/instrumentation/src/main/java/org/evomaster/client/java/instrumentation/heuristic/TruthnessUtils.java
@@ -4,9 +4,29 @@ import org.evomaster.client.java.instrumentation.coverage.methodreplacement.Dist
 
 public class TruthnessUtils {
 
+    /**
+     * scales to a positive double value to the [0,1] range
+     *
+     * @param v a non-negative double value
+     * @return
+     */
+    public static double normalizeValue(double v) {
+        if (v < 0) {
+            throw new IllegalArgumentException("Negative value: " + v);
+        }
+
+        //normalization function from old ICST/STVR paper
+        double normalized = v / (v + 1d);
+
+        assert normalized >= 0 && normalized <= 1;
+
+        return normalized;
+    }
+
+
     public static Truthness getEqualityTruthness(int a, int b) {
         double distance = DistanceHelper.getDistanceToEquality(a, b);
-        double normalizedDistance = Truthness.normalizeValue(distance);
+        double normalizedDistance = normalizeValue(distance);
         return new Truthness(
                 1d - normalizedDistance,
                 a != b ? 1d : 0d
@@ -15,7 +35,7 @@ public class TruthnessUtils {
 
     public static Truthness getEqualityTruthness(long a, long b) {
         double distance = DistanceHelper.getDistanceToEquality(a, b);
-        double normalizedDistance = Truthness.normalizeValue(distance);
+        double normalizedDistance = normalizeValue(distance);
         return new Truthness(
                 1d - normalizedDistance,
                 a != b ? 1d : 0d
@@ -32,11 +52,28 @@ public class TruthnessUtils {
 
     public static Truthness getEqualityTruthness(double a, double b) {
         double distance = DistanceHelper.getDistanceToEquality(a, b);
-        double normalizedDistance = Truthness.normalizeValue(distance);
+        double normalizedDistance = normalizeValue(distance);
         return new Truthness(
                 1d - normalizedDistance,
                 a != b ? 1d : 0d
         );
     }
 
+
+    /**
+     * @param len a positive value for a length
+     * @return
+     */
+    public static Truthness getTruthnessToEmpty(int len) {
+        Truthness t;
+        if (len < 0) {
+            throw new IllegalArgumentException("lengths should always be non-negative. Invalid length " + len);
+        }
+        if (len == 0) {
+            t = new Truthness(1, 0);
+        } else {
+            t = new Truthness(1d / (1d + len), 1);
+        }
+        return t;
+    }
 }
