### 2020-03-02 a31d2a8
adding support for non-integer comparisons. started (but not finished) writing test cases for it
