+++ client-java/instrumentation/src/test/java/org/evomaster/client/java/instrumentation/example/nonintegercomparisons/NIC_ExampleInstrumentedTest.java
@@ -1,0 +1,383 @@
+package org.evomaster.client.java.instrumentation.example.nonintegercomparisons;
+
+import com.foo.somedifferentpackage.examples.branches.BranchesImp;
+import com.foo.somedifferentpackage.examples.nonintegercomparisons.NIC_ExampleImp;
+import org.evomaster.client.java.instrumentation.InstrumentingClassLoader;
+import org.evomaster.client.java.instrumentation.example.branches.Branches;
+import org.evomaster.client.java.instrumentation.shared.ObjectiveNaming;
+import org.evomaster.client.java.instrumentation.staticstate.ExecutionTracer;
+import org.evomaster.client.java.instrumentation.staticstate.ObjectiveRecorder;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Set;
+import java.util.function.Supplier;
+import java.util.stream.Collectors;
+
+import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+/**
+ * Created by arcuri82 on 02-Mar-20.
+ */
+public class NIC_ExampleInstrumentedTest {
+
+    protected NIC_Example getInstance() throws Exception {
+
+        InstrumentingClassLoader cl = new InstrumentingClassLoader("com.foo");
+
+        return (NIC_Example)
+                cl.loadClass(NIC_ExampleImp.class.getName()).getDeclaredConstructor().newInstance();
+    }
+
+    private int evalPos(long x, long y){
+
+        try {
+            return getInstance().pos(x,y);
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    private int evalNeg(long x, long y){
+
+        try {
+            return getInstance().neg(x,y);
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    private int evalEq(long x, long y){
+
+        try {
+            return getInstance().eq(x,y);
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+
+    private int evalPos(double x, double y){
+
+        try {
+            return getInstance().pos(x,y);
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    private int evalNeg(double x, double y){
+
+        try {
+            return getInstance().neg(x,y);
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    private int evalEq(double x, double y){
+
+        try {
+            return getInstance().eq(x,y);
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    private int evalPos(float x, float y){
+
+        try {
+            return getInstance().pos(x,y);
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    private int evalNeg(float x, float y){
+
+        try {
+            return getInstance().neg(x,y);
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    private int evalEq(float x, float y){
+
+        try {
+            return getInstance().eq(x,y);
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+
+
+
+    @BeforeAll
[…]
