### 2020-03-02 93346f6
WiP. Refactoring complete?
