### 2020-03-03 ff49482
more unit tests for numeric comparisons
