+++ client-java/instrumentation/src/test/java/org/evomaster/client/java/instrumentation/example/nonintegercomparisons/NIC_ExampleInstrumentedTest.java
@@ -163,11 +163,11 @@ public class NIC_ExampleInstrumentedTest {
         );
     }
 
-    private void testPosX(Supplier<Integer> firstCall,
-                         Supplier<Integer> secondCall,
-                         Supplier<Integer> thirdCall){
+    private void testPosX(Supplier<Integer> firstCall_positiveX,
+                         Supplier<Integer> secondCall_worseX,
+                         Supplier<Integer> thirdCall_betterX){
 
-        int res = firstCall.get();
+        int res = firstCall_positiveX.get();
         //first branch should had been taken
         assertEquals(0, res);
 
@@ -183,20 +183,14 @@ public class NIC_ExampleInstrumentedTest {
             assertTrue(h > 0 ); // it has been reached though
         }
 
-        List<Double> first = nonCovered.stream()
-                .sorted()
-                .map(id -> ExecutionTracer.getValue(id))
-                .collect(Collectors.toList());
+        List<Double> first = extractHeuristicsSorted(nonCovered);
 
 
-        secondCall.get(); //worse value, should have no impact
+        secondCall_worseX.get(); //worse value, should have no impact
         nonCovered = ExecutionTracer.getNonCoveredObjectives(ObjectiveNaming.NUMERIC_COMPARISON);
         assertEquals(2, nonCovered.size());
 
-        List<Double> second = nonCovered.stream()
-                .sorted()
-                .map(id -> ExecutionTracer.getValue(id))
-                .collect(Collectors.toList());
+        List<Double> second = extractHeuristicsSorted(nonCovered);
 
         for(int i=0; i<first.size(); i++) {
             //no impact, the same
@@ -204,14 +198,11 @@ public class NIC_ExampleInstrumentedTest {
         }
 
 
-        thirdCall.get(); //better value
+        thirdCall_betterX.get(); //better value
         nonCovered = ExecutionTracer.getNonCoveredObjectives(ObjectiveNaming.NUMERIC_COMPARISON);
         assertEquals(2, nonCovered.size());
 
-        List<Double> third = nonCovered.stream()
-                .sorted()
-                .map(id -> ExecutionTracer.getValue(id))
-                .collect(Collectors.toList());
+        List<Double> third = extractHeuristicsSorted(nonCovered);
 
         for(int i=0; i<first.size(); i++) {
             //better
@@ -219,165 +210,319 @@ public class NIC_ExampleInstrumentedTest {
             //but still not covered
             assertTrue(third.get(i) < 1);
         }
-
     }
 
 
 
-//    @Test
-//    public void testPosY() {
-//
-//        assertEquals(0d, ObjectiveRecorder.computeCoverage(ObjectiveNaming.BRANCH));
-//
-//        int res;
-//        res = evalPos(10, 0);
-//        assertEquals(0, res);
-//
-//        res = evalPos(-5, 4);
-//        assertEquals(1, res);
-//
-//        //seen 2 "if", but returned on the second "if"
-//        assertEquals(4, ExecutionTracer.getNumberOfObjectives(ObjectiveNaming.BRANCH));
-//        assertEquals(1, ExecutionTracer.getNumberOfNonCoveredObjectives(ObjectiveNaming.BRANCH));
-//
-//        String elseBranch = ExecutionTracer.getNonCoveredObjectives(ObjectiveNaming.BRANCH).iterator().next();
-//        assertTrue(elseBranch.contains(ObjectiveNaming.FALSE_BRANCH));
[…]
