### 2020-03-04 ba732de
Ignore select when is currval or nextval of a sequence
