+++ client-java/controller/src/test/java/org/evomaster/client/java/controller/internal/db/SqlHandlerInDBTest.java
@@ -6,12 +6,14 @@ import org.evomaster.client.java.controller.InstrumentedSutStarter;
 import org.evomaster.client.java.controller.api.dto.TestResultsDto;
 import org.evomaster.client.java.controller.api.dto.database.execution.ExecutionDto;
 import org.evomaster.client.java.controller.db.SqlScriptRunner;
+import org.jetbrains.annotations.NotNull;
 import org.junit.jupiter.api.Test;
 
 import static io.restassured.RestAssured.given;
 import static org.evomaster.client.java.controller.api.ControllerConstants.BASE_PATH;
 import static org.evomaster.client.java.controller.api.ControllerConstants.TEST_RESULTS;
 import static org.junit.jupiter.api.Assertions.*;
+import java.sql.SQLException;
 
 /**
  * Created by arcuri82 on 24-Apr-19.
@@ -39,20 +41,7 @@ public class SqlHandlerInDBTest extends DatabaseTestTemplate {
         InstrumentedSutStarter starter = getInstrumentedSutStarter();
 
         try {
-            String url = start(starter);
-            url += BASE_PATH;
-
-            startNewTest(url);
-
-            ExecutionDto dto = getSqlExecutionDto(0,url);
-
-            assertTrue(dto == null || dto.deletedData == null || dto.deletedData.isEmpty());
-
-            startNewActionInSameTest(url, 1);
-
-            SqlScriptRunner.execCommand(getConnection(), "Delete FROM Foo");
-
-            dto = getSqlExecutionDto(1,url);
+            ExecutionDto dto = executeCommand(starter, "Delete FROM Foo");
 
             assertNotNull(dto);
             assertNotNull(dto.deletedData);
@@ -64,7 +53,6 @@ public class SqlHandlerInDBTest extends DatabaseTestTemplate {
         }
     }
 
-
     @Test
     public void testInsertTable() throws Exception {
 
@@ -73,20 +61,7 @@ public class SqlHandlerInDBTest extends DatabaseTestTemplate {
         InstrumentedSutStarter starter = getInstrumentedSutStarter();
 
         try {
-            String url = start(starter);
-            url += BASE_PATH;
-
-            startNewTest(url);
-
-            ExecutionDto dto = getSqlExecutionDto(0,url);
-
-            assertTrue(dto == null || dto.insertedData == null || dto.insertedData.isEmpty());
-
-            startNewActionInSameTest(url, 1);
-
-            SqlScriptRunner.execCommand(getConnection(), "insert into Foo (x) values (42)");
-
-            dto = getSqlExecutionDto(1,url);
+            ExecutionDto dto = executeCommand(starter, "insert into Foo (x) values (42)");
 
             assertNotNull(dto);
             assertNotNull(dto.insertedData);
@@ -98,6 +73,7 @@ public class SqlHandlerInDBTest extends DatabaseTestTemplate {
         }
     }
 
+
     @Test
     public void testUpdateTable() throws Exception {
 
@@ -106,20 +82,7 @@ public class SqlHandlerInDBTest extends DatabaseTestTemplate {
         InstrumentedSutStarter starter = getInstrumentedSutStarter();
 
         try {
-            String url = start(starter);
-            url += BASE_PATH;
-
-            startNewTest(url);
-
-            ExecutionDto dto = getSqlExecutionDto(0,url);
-
-            assertTrue(dto == null || dto.updatedData == null || dto.updatedData.isEmpty());
-
-            startNewActionInSameTest(url, 1);
-
-            SqlScriptRunner.execCommand(getConnection(), "update Foo set x=42");
-
-            dto = getSqlExecutionDto(1,url);
[…]
