### 2020-03-04 e60aafe
Biggest mongo distances are computed for each find operation
