+++ e2e-tests/spring-rest-mongo/src/test/java/org/evomaster/e2etests/spring/rest/mongo/customer/MongoFooAppExtraFitnessTest.java
@@ -1,0 +1,148 @@
+package org.evomaster.e2etests.spring.rest.mongo.customer;
+
+import com.foo.mongo.MyMongoAppEmbeddedController;
+import com.google.inject.Injector;
+import com.google.inject.Key;
+import com.google.inject.TypeLiteral;
+import org.evomaster.client.java.controller.internal.db.StandardOutputTracker;
+import org.evomaster.core.Main;
+import org.evomaster.core.database.DbAction;
+import org.evomaster.core.problem.rest.*;
+import org.evomaster.core.problem.rest.auth.NoAuth;
+import org.evomaster.core.problem.rest.param.Param;
+import org.evomaster.core.search.EvaluatedIndividual;
+import org.evomaster.core.search.FitnessValue;
+import org.evomaster.core.search.service.FitnessFunction;
+import org.evomaster.e2etests.spring.rest.mongo.SpringRestMongoTestBase;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.LinkedList;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+public class MongoFooAppExtraFitnessTest extends SpringRestMongoTestBase {
+
+    private static final MyMongoAppEmbeddedController sutController = new MyMongoAppEmbeddedController();
+
+    @BeforeAll
+    public static void init() throws Exception {
+        SpringRestMongoTestBase.initClass(sutController);
+    }
+
+    @BeforeEach
+    public void turnOnTracker() {
+        StandardOutputTracker.setTracker(true, sutController);
+    }
+
+    @AfterEach
+    public void turnOffTracker() {
+        StandardOutputTracker.setTracker(false, sutController);
+    }
+
+    @Test
+    public void testNoDocumentsFound() {
+
+        String[] args = new String[]{
+                "--createTests", "true",
+                "--seed", "42",
+                "--sutControllerPort", "" + controllerPort,
+                "--maxActionEvaluations", "1",
+                "--stoppingCriterion", "FITNESS_EVALUATIONS",
+                "--heuristicsForSQL", "true",
+                "--generateSqlDataWithSearch", "true",
+                "--heuristicsForMongo", "true",
+                "--maxTestSize", "1"
+        };
+
+        Injector injector = Main.init(args);
+
+        FitnessFunction<RestIndividual> ff = injector.getInstance(Key.get(
+                new TypeLiteral<FitnessFunction<RestIndividual>>() {
+                }));
+
+
+        RestCallAction getAction = new RestCallAction("GET/api/mongobar",
+                HttpVerb.GET, new RestPath("/api/mongobar"),
+                new LinkedList<Param>(),
+                new NoAuth(),
+                false,
+                null,
+                new LinkedList<String>(),
+                new HashMap<String, String>());
+
+        RestIndividual individual = new RestIndividual(Arrays.asList(getAction),
+                SampleType.RANDOM,
+                new LinkedList<DbAction>(),
+                null,
+                null);
+
+        EvaluatedIndividual ei = ff.calculateCoverage(individual);
+        FitnessValue fv = ei.getFitness();
+
+        //as no data in database, should get distance greater than zero
+        assertTrue(fv.averageExtraDistancesToMinimize(0)>0.0);
+
+    }
+
+    @Test
+    public void testOneDocumentFound() {
+
+        String[] args = new String[]{
+                "--createTests", "true",
+                "--seed", "42",
+                "--sutControllerPort", "" + controllerPort,
[…]
