+++ client-java/controller-api/src/main/java/org/evomaster/client/java/controller/api/dto/database/execution/MongoOperationDto.java
@@ -8,4 +8,5 @@ public class MongoOperationDto {
 
     public String operationJsonStr;
 
+    public boolean hasOperationFoundAnyDocuments;
 }
