+++ e2e-tests/spring-rest-mongo/src/main/java/com/foo/mongo/MongoFooRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 
 import java.util.List;
 
-public interface MyMongoRepository extends MongoRepository<MongoFooEntity, String> {
+public interface MongoFooRepository extends MongoRepository<MongoFooEntity, String> {
 
     MongoFooEntity findByFirstName(String firstName);
 
