### 2020-03-04 ab97227
new extractMongoExecutionInfo added to config options documentation
