+++ client-java/controller-api/src/main/java/org/evomaster/client/java/controller/api/dto/database/execution/FindResultDto.java
@@ -1,0 +1,11 @@
+package org.evomaster.client.java.controller.api.dto.database.execution;
+
+public class FindResultDto {
+
+    public enum FindResultType { SUMMARY }
+
+    public FindResultType findResultType;
+
+    public boolean hasReturnedAnyDocument;
+
+}
