+++ client-java/controller-api/src/main/java/org/evomaster/client/java/controller/api/dto/database/execution/FindOperationDto.java
@@ -1,0 +1,10 @@
+package org.evomaster.client.java.controller.api.dto.database.execution;
+
+public class FindOperationDto {
+
+    public String databaseName;
+
+    public String collectionName;
+
+    public String queryJsonStr;
+}
