+++ client-java/controller-api/src/main/java/org/evomaster/client/java/controller/api/dto/database/execution/MongoExecutionDto.java
@@ -14,5 +14,5 @@ import java.util.List;
  */
 public class MongoExecutionDto {
 
-    public List<MongoOperationDto> mongoOperations = new ArrayList<>();
+    public List<ExecutedFindOperationDto> executedFindOperationDtos = new ArrayList<>();
 }
