+++ client-java/controller/src/test/java/org/evomaster/client/java/controller/internal/db/mongo/MongoHandlerIntegrationTest.java
@@ -1,15 +1,22 @@
 package org.evomaster.client.java.controller.internal.db.mongo;
 
+import com.google.gson.Gson;
 import com.mongodb.client.MongoCollection;
+import io.restassured.http.ContentType;
+import org.apache.http.HttpStatus;
 import org.bson.BsonDocument;
 import org.bson.Document;
 import org.evomaster.client.java.controller.InstrumentedSutStarter;
+import org.evomaster.client.java.controller.api.dto.WrappedResponseDto;
+import org.evomaster.client.java.controller.api.dto.database.execution.FindOperationDto;
+import org.evomaster.client.java.controller.api.dto.database.execution.FindResultDto;
 import org.evomaster.client.java.instrumentation.mongo.MongoLogger;
 import org.junit.jupiter.api.*;
 
+import static io.restassured.RestAssured.given;
 import static org.evomaster.client.java.controller.api.ControllerConstants.BASE_PATH;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.evomaster.client.java.controller.api.ControllerConstants.MONGO_COMMAND;
+import static org.junit.jupiter.api.Assertions.*;
 
 public class MongoHandlerIntegrationTest extends MongoTestTemplate {
 
@@ -47,59 +54,107 @@ public class MongoHandlerIntegrationTest extends MongoTestTemplate {
     @Test
     public void testEmptyDistances() {
         startNewTest(url);
-        assertTrue(sutController.getMongoHandler().getMongoExecutionDto().mongoOperations.isEmpty());
+        assertTrue(sutController.getMongoHandler().getMongoExecutionDto().executedFindOperationDtos.isEmpty());
     }
 
     @Test
     public void testFindNoResult() {
         startNewTest(url);
-        assertTrue(sutController.getMongoHandler().getMongoExecutionDto().mongoOperations.isEmpty());
+        assertTrue(sutController.getMongoHandler().getMongoExecutionDto().executedFindOperationDtos.isEmpty());
 
         MongoCollection<Document> mongoCollection = mongoClient.getDatabase("testdb").getCollection("customers");
 
         // find({}) returns no element
         MongoLogger.getInstance().logFind(mongoCollection, new BsonDocument(), true);
 
-        assertEquals(1, sutController.getMongoHandler().getMongoExecutionDto().mongoOperations.size());
+        assertEquals(1, sutController.getMongoHandler().getMongoExecutionDto().executedFindOperationDtos.size());
     }
 
 
     @Test
     public void testTwoFindOperations() {
         startNewTest(url);
-        assertTrue(sutController.getMongoHandler().getMongoExecutionDto().mongoOperations.isEmpty());
+        assertTrue(sutController.getMongoHandler().getMongoExecutionDto().executedFindOperationDtos.isEmpty());
 
         MongoCollection<Document> mongoCollection = mongoClient.getDatabase("testdb").getCollection("customers");
 
         // log find({}) operation
         MongoLogger.getInstance().logFind(mongoCollection, new BsonDocument(), true);
 
-        assertEquals(1, sutController.getMongoHandler().getMongoExecutionDto().mongoOperations.size());
+        assertEquals(1, sutController.getMongoHandler().getMongoExecutionDto().executedFindOperationDtos.size());
 
         // log find({}) operation
         MongoLogger.getInstance().logFind(mongoCollection, new BsonDocument(), true);
 
-        assertEquals(2, sutController.getMongoHandler().getMongoExecutionDto().mongoOperations.size());
+        assertEquals(2, sutController.getMongoHandler().getMongoExecutionDto().executedFindOperationDtos.size());
     }
 
 
     @Test
[…]
