### 2020-03-06 524f17a
RemoteController receives correct argument for computing Mongo heuristics
