+++ e2e-tests/e2e-tests-utils/src/test/java/org/evomaster/e2etests/utils/RestTestBase.java
@@ -300,7 +300,7 @@ public abstract class RestTestBase {
 
         controllerPort = embeddedStarter.getControllerServerPort();
 
-        remoteController = new RemoteController("localhost", controllerPort, true);
+        remoteController = new RemoteController("localhost", controllerPort, true, true);
         boolean started = remoteController.startSUT();
         assertTrue(started);
 

[…]
