### 2020-03-09 #2001 @jaynanavati
Dear Sir,

I am willing to review the submission mentioned in the subject.

Thank you.

*Dr. Jay Nanavati*

On Mon, Mar 9, 2020 at 1:43 AM George K. Thiruvathukal <
notifications@github.com> wrote:

> @jaynanavati <https://github.com/jaynanavati> Are you willing to
> contribute a review for this JOSS submission?
>
