### 2020-03-09 #2153 @whedon
[REVIEW]: EvoMaster: A Search-Based System Test Generation Tool
**Submitting author:** @arcuri82 (<a href="http://orcid.org/0000-0003-0799-2930">Andrea Arcuri</a>)
**Repository:** <a href="https://github.com/EMResearch/EvoMaster" target ="_blank">https://github.com/EMResearch/EvoMaster</a>
**Version:** 1.1.0

@mado89, @UTH-Tuan, @s0nata, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
