### 2020-03-09 42f3719
bugfix: MongoLogger was using incorrect OutputStream
