+++ e2e-tests/spring-rest-mongo/src/main/java/com/foo/mongo/person/AddressDto.java
@@ -1,0 +1,9 @@
+package com.foo.mongo.person;
+
+public class AddressDto {
+
+    public String streetName;
+
+    public int streetNumber;
+
+}
