+++ e2e-tests/spring-rest-mongo/src/main/java/com/foo/mongo/person/PersonDto.java
@@ -1,0 +1,8 @@
+package com.foo.mongo.person;
+
+public class PersonDto {
+
+    public String firstName;
+    public String lastName;
+    public AddressDto address;
+}

[…]
