### 2020-03-09 7d93280
queries for mongoDB repository
