+++ e2e-tests/spring-rest-mongo/src/main/java/com/foo/mongo/person/PersonDto.java
@@ -5,4 +5,5 @@ public class PersonDto {
     public String firstName;
     public String lastName;
     public AddressDto address;
+    public int age;
 }
