### 2020-03-10 0ddce14
new test case for extra fitness of mongodb
