### 2020-03-11 f399360
created MongoHeuristicCalculator for computing the heuristic distance of a document to a filter
