+++ client-java/controller-api/src/main/java/org/evomaster/client/java/controller/api/dto/database/execution/FindResultDto.java
@@ -1,11 +1,19 @@
 package org.evomaster.client.java.controller.api.dto.database.execution;
 
+import org.bson.Document;
+
 public class FindResultDto {
 
-    public enum FindResultType { SUMMARY }
+    public static final String HAS_RETURNED_ANY_DOCUMENT_FIELD_NAME = "hasReturnedAnyDocument";
+
+    public static final String DOCUMENTS_FIELD_NAME = "documents";
+
+    public enum FindResultType {SUMMARY, DETAILED}
 
     public FindResultType findResultType;
 
     public boolean hasReturnedAnyDocument;
+
+    public Document[] documents;
 
 }
