+++ e2e-tests/spring-rest-mongo/src/test/java/org/evomaster/e2etests/spring/rest/mongo/foo/ExtraFitnessMongoQueryTest.java
@@ -1,14 +1,20 @@
 package org.evomaster.e2etests.spring.rest.mongo.foo;
 
 import com.foo.mongo.MyMongoAppEmbeddedController;
+import com.foo.mongo.person.PersonDto;
 import com.google.inject.Injector;
 import com.google.inject.Key;
 import com.google.inject.TypeLiteral;
+import com.mongodb.client.FindIterable;
+import io.restassured.http.ContentType;
+import io.restassured.response.ValidatableResponse;
+import org.bson.Document;
 import org.evomaster.client.java.controller.internal.db.StandardOutputTracker;
 import org.evomaster.core.Main;
 import org.evomaster.core.database.DbAction;
 import org.evomaster.core.problem.rest.*;
 import org.evomaster.core.problem.rest.auth.NoAuth;
+import org.evomaster.core.problem.rest.param.Param;
 import org.evomaster.core.problem.rest.param.PathParam;
 import org.evomaster.core.search.EvaluatedIndividual;
 import org.evomaster.core.search.FitnessValue;
@@ -26,6 +32,8 @@ import java.util.Arrays;
 import java.util.HashMap;
 import java.util.LinkedList;
 
+import static io.restassured.RestAssured.given;
+import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 public class ExtraFitnessMongoQueryTest extends SpringRestMongoTestBase {
@@ -46,6 +54,15 @@ public class ExtraFitnessMongoQueryTest extends SpringRestMongoTestBase {
     @AfterEach
     public void turnOffTracker() {
         StandardOutputTracker.setTracker(false, sutController);
+    }
+
+    private static ValidatableResponse add(PersonDto johnDoeDto) {
+        return given()
+                .contentType(ContentType.JSON)
+                .body(johnDoeDto)
+                .post(baseUrlOfSut + "/api/mongoperson/add")
+                .then()
+                .statusCode(200);
     }
 
     @Test
@@ -103,7 +120,79 @@ public class ExtraFitnessMongoQueryTest extends SpringRestMongoTestBase {
         FitnessValue fv = ei.getFitness();
 
         //as no data in database, should get distance greater than zero
-        assertTrue(fv.averageExtraDistancesToMinimize(0) > 0.0);
+        assertTrue(fv.averageExtraDistancesToMinimize(0) == Double.MAX_VALUE);
+
+    }
+
+    @Test
+    public void testExtraFitnessWithSomeDocuments() {
+
+        String[] args = new String[]{
+                "--createTests", "true",
+                "--seed", "42",
+                "--sutControllerPort", "" + controllerPort,
+                "--maxActionEvaluations", "2",
+                "--stoppingCriterion", "FITNESS_EVALUATIONS",
+                "--heuristicsForMongo", "true",
+                "--maxTestSize", "1"
+        };
+
+        Injector injector = Main.init(args);
+
+        FitnessFunction<RestIndividual> ff = injector.getInstance(Key.get(
+                new TypeLiteral<FitnessFunction<RestIndividual>>() {
+                }));
+
+        IntegerGene ageGene = new IntegerGene("age",
+                0,
+                Integer.MIN_VALUE,
+                Integer.MAX_VALUE,
+                new IntMutationUpdate(Integer.MIN_VALUE, Integer.MAX_VALUE, 0, false));
+
+        PathParam pathParam = new PathParam("age",
+                new DisruptiveGene<IntegerGene>(
+                        "age",
+                        ageGene,
+                        1.0d
+                )
+        );
+
[…]
