+++ client-java/controller/src/test/java/org/evomaster/client/java/controller/internal/db/mongo/MongoHandlerIntegrationTest.java
@@ -1,11 +1,9 @@
 package org.evomaster.client.java.controller.internal.db.mongo;
 
-import com.google.gson.Gson;
 import com.mongodb.client.MongoCollection;
 import io.restassured.http.ContentType;
 import org.apache.http.HttpStatus;
-import org.bson.BsonDocument;
-import org.bson.Document;
+import org.bson.*;
 import org.evomaster.client.java.controller.InstrumentedSutStarter;
 import org.evomaster.client.java.controller.api.dto.WrappedResponseDto;
 import org.evomaster.client.java.controller.api.dto.database.execution.FindOperationDto;
@@ -13,12 +11,18 @@ import org.evomaster.client.java.controller.api.dto.database.execution.FindResul
 import org.evomaster.client.java.instrumentation.mongo.MongoLogger;
 import org.junit.jupiter.api.*;
 
+import java.util.List;
+import java.util.Map;
+
 import static io.restassured.RestAssured.given;
 import static org.evomaster.client.java.controller.api.ControllerConstants.BASE_PATH;
 import static org.evomaster.client.java.controller.api.ControllerConstants.MONGO_COMMAND;
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 public class MongoHandlerIntegrationTest extends MongoTestTemplate {
+
+    public static final String DATABASE_NAME = "testdb";
 
     private static InstrumentedSutStarter starter;
     private static MongoFakeSutController sutController;
@@ -34,7 +38,7 @@ public class MongoHandlerIntegrationTest extends MongoTestTemplate {
 
     @BeforeEach
     public void dropMongoDatabase() {
-        mongoClient.getDatabase("testdb").drop();
+        mongoClient.getDatabase(DATABASE_NAME).drop();
         sutController.initMongoClient();
     }
 
@@ -62,7 +66,7 @@ public class MongoHandlerIntegrationTest extends MongoTestTemplate {
         startNewTest(url);
         assertTrue(sutController.getMongoHandler().getMongoExecutionDto().executedFindOperationDtos.isEmpty());
 
-        MongoCollection<Document> mongoCollection = mongoClient.getDatabase("testdb").getCollection("customers");
+        MongoCollection<Document> mongoCollection = mongoClient.getDatabase(DATABASE_NAME).getCollection("customers");
 
         // find({}) returns no element
         MongoLogger.getInstance().logFind(mongoCollection, new BsonDocument(), true);
@@ -76,7 +80,7 @@ public class MongoHandlerIntegrationTest extends MongoTestTemplate {
         startNewTest(url);
         assertTrue(sutController.getMongoHandler().getMongoExecutionDto().executedFindOperationDtos.isEmpty());
 
-        MongoCollection<Document> mongoCollection = mongoClient.getDatabase("testdb").getCollection("customers");
+        MongoCollection<Document> mongoCollection = mongoClient.getDatabase(DATABASE_NAME).getCollection("customers");
 
         // log find({}) operation
         MongoLogger.getInstance().logFind(mongoCollection, new BsonDocument(), true);
@@ -95,7 +99,7 @@ public class MongoHandlerIntegrationTest extends MongoTestTemplate {
         startNewTest(url);
         assertTrue(sutController.getMongoHandler().getMongoExecutionDto().executedFindOperationDtos.isEmpty());
 
-        MongoCollection<Document> mongoCollection = mongoClient.getDatabase("testdb").getCollection("customers");
+        MongoCollection<Document> mongoCollection = mongoClient.getDatabase(DATABASE_NAME).getCollection("customers");
 
[…]
