### 2020-03-12 189cdfd
Working on MongoHeuristicCalculator
