### 2020-03-12 371449a
disable check on final state of database
