### 2020-03-13 0cbfb52
translating Filters to AST for queries
