### 2020-03-15 9d5c21b
support for queries observed from Spring-mongo
