+++ e2e-tests/spring-rest-mongo/src/test/java/org/evomaster/e2etests/spring/rest/mongo/foo/ExtraFitnessMongoFilterTest.java
@@ -11,7 +11,6 @@ import io.restassured.response.ValidatableResponse;
 import org.bson.Document;
 import org.evomaster.client.java.controller.internal.db.StandardOutputTracker;
 import org.evomaster.core.Main;
-import org.evomaster.core.database.DbAction;
 import org.evomaster.core.problem.rest.*;
 import org.evomaster.core.problem.rest.auth.NoAuth;
 import org.evomaster.core.problem.rest.param.Param;
@@ -19,8 +18,11 @@ import org.evomaster.core.problem.rest.param.PathParam;
 import org.evomaster.core.search.EvaluatedIndividual;
 import org.evomaster.core.search.FitnessValue;
 import org.evomaster.core.search.gene.DisruptiveGene;
+import org.evomaster.core.search.gene.GeneIndependenceInfo;
 import org.evomaster.core.search.gene.IntegerGene;
+import org.evomaster.core.search.gene.StringGene;
 import org.evomaster.core.search.service.FitnessFunction;
+import org.evomaster.core.search.service.mutator.geneMutation.ArchiveMutator;
 import org.evomaster.core.search.service.mutator.geneMutation.IntMutationUpdate;
 import org.evomaster.e2etests.spring.rest.mongo.SpringRestMongoTestBase;
 import org.junit.jupiter.api.AfterEach;
@@ -66,7 +68,7 @@ public class ExtraFitnessMongoFilterTest extends SpringRestMongoTestBase {
     }
 
     @Test
-    public void testExtraFitnessNoDocumentsFound() {
+    public void testFindByAgeNoDocuments() {
 
         String[] args = new String[]{
                 "--createTests", "true",
@@ -92,7 +94,7 @@ public class ExtraFitnessMongoFilterTest extends SpringRestMongoTestBase {
                 new IntMutationUpdate(Integer.MIN_VALUE, Integer.MAX_VALUE, 0, false));
 
         PathParam pathParam = new PathParam("age",
-                new DisruptiveGene<IntegerGene>(
+                new DisruptiveGene<>(
                         "age",
                         ageGene,
                         1.0d
@@ -107,12 +109,12 @@ public class ExtraFitnessMongoFilterTest extends SpringRestMongoTestBase {
                 new NoAuth(),
                 false,
                 null,
-                new LinkedList<String>(),
-                new HashMap<String, String>());
+                new LinkedList<>(),
+                new HashMap<>());
 
         RestIndividual individual = new RestIndividual(Arrays.asList(getAction),
                 SampleType.RANDOM,
-                new LinkedList<DbAction>(),
+                new LinkedList<>(),
                 null,
                 null);
 
@@ -125,7 +127,7 @@ public class ExtraFitnessMongoFilterTest extends SpringRestMongoTestBase {
     }
 
     @Test
-    public void testExtraFitnessWithSomeDocuments() {
+    public void testFindByAge() {
 
         String[] args = new String[]{
                 "--createTests", "true",
@@ -150,7 +152,7 @@ public class ExtraFitnessMongoFilterTest extends SpringRestMongoTestBase {
                 new IntMutationUpdate(Integer.MIN_VALUE, Integer.MAX_VALUE, 0, false));
 
         PathParam pathParam = new PathParam("age",
-                new DisruptiveGene<IntegerGene>(
+                new DisruptiveGene<>(
                         "age",
                         ageGene,
                         1.0d
@@ -159,12 +161,12 @@ public class ExtraFitnessMongoFilterTest extends SpringRestMongoTestBase {
 
         RestCallAction postAction = new RestCallAction("POST/api/mongoperson/addJoeBlack",
                 HttpVerb.POST, new RestPath("/api/mongoperson/addJoeBlack"),
-                new LinkedList<Param>(),
[…]
