### 2020-03-18 957ea95
WiP. Minor refactoring + setting up the possibility to send parameters to the Clusterer and clustering metrics.
