### 2020-03-23 bee6a57
more documentation and ICST'2020 paper
