### 2020-03-23 a621438
fixed bug in Randomness, and added checks for non-determinism
