### 2020-03-23 70227be
added test, and fixed issue when lambda throws exception in deterministic checks
