+++ e2e-tests/spring-examples/src/test/java/org/evomaster/e2etests/spring/examples/strings/StringsEMTest.java
@@ -11,6 +11,14 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 public class StringsEMTest extends StringsTestBase {
 
     @Test
+    public void testDeterminism(){
+
+        runAndCheckDeterminism(1_000, (args) -> {
+            initAndRun(args);
+        });
+    }
+
+    @Test
     public void testRunEM() throws Throwable {
 
         runTestHandlingFlakyAndCompilation(
