### 2020-03-23 e15f4a2
fixed another source of non-determinism
