+++ e2e-tests/spring-examples/src/test/java/org/evomaster/e2etests/spring/examples/strings/StringsEMTest.java
@@ -21,6 +21,8 @@ public class StringsEMTest extends StringsTestBase {
     @Test
     public void testRunEM() throws Throwable {
 
+        defaultSeed = 13;
+
         runTestHandlingFlakyAndCompilation(
                 "StringsEM",
                 "org.bar.StringsEM",
