+++ e2e-tests/spring-examples/src/test/java/org/evomaster/e2etests/spring/examples/impact/ObjectImpactEMTest.java
@@ -1,178 +1,0 @@
-package org.evomaster.e2etests.spring.examples.impact;
-
-import com.foo.rest.examples.spring.impact.ImpactRestController;
-import org.evomaster.core.problem.rest.RestIndividual;
-import org.evomaster.core.problem.rest.util.ParamUtil;
-import org.evomaster.core.search.EvaluatedIndividual;
-import org.evomaster.core.search.Individual;
-import org.evomaster.core.search.Solution;
-import org.evomaster.core.search.gene.Gene;
-import org.evomaster.core.search.gene.ObjectGene;
-import org.evomaster.core.search.impact.GeneImpact;
-import org.evomaster.core.search.impact.Impact;
-import org.evomaster.core.search.impact.GeneMutationSelectionMethod;
-import org.evomaster.core.search.impact.ImpactUtils;
-import org.evomaster.core.search.impact.value.ObjectGeneImpact;
-import org.evomaster.e2etests.spring.examples.SpringTestBase;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Test;
-
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-import java.util.*;
-import java.util.stream.Collectors;
-
-import static org.junit.jupiter.api.Assertions.*;
-import static org.junit.jupiter.api.Assertions.fail;
-
-/**
- * created by manzh on 2019-09-12
- */
-public class ObjectImpactEMTest extends SpringTestBase {
-
-    @Test
-    public void testAwayNoImpact() throws Throwable {
-        testRunEM(GeneMutationSelectionMethod.AWAY_NOIMPACT);
-    }
-
-    @Test
-    public void testImpact() throws Throwable {
-        testRunEM(GeneMutationSelectionMethod.APPROACH_IMPACT);
-    }
-
-    @Test
-    public void testLatestImpact() throws Throwable {
-        testRunEM(GeneMutationSelectionMethod.APPROACH_LATEST_IMPACT);
-    }
-
-    @Test
-    public void testLatestImprovement() throws Throwable {
-        testRunEM(GeneMutationSelectionMethod.APPROACH_LATEST_IMPROVEMENT);
-    }
-
-    @Test
-    public void testBalance() throws Throwable {
-        testRunEM(GeneMutationSelectionMethod.BALANCE_IMPACT_NOIMPACT);
-    }
-
-    public void testRunEM(GeneMutationSelectionMethod method) throws Throwable {
-
-        runTestHandlingFlakyAndCompilation(
-                "none",
-                "none",
-                1_000,
-                false,
-                (args) -> {
-
-                    args.add("--probOfArchiveMutation");
-                    args.add("1.0");
-
-                    args.add("--geneSelectionMethod");
-                    args.add(method.toString());
-
-                    args.add("--enableTrackEvaluatedIndividual");
-                    args.add("true");
-
-                    args.add("--exportImpacts");
-                    args.add("true");
-                    args.add("--impactFile");
-                    args.add("target/impactInfo/TestobjImpacts_"+method.toString()+".csv");
-
-                    args.add("--exportCoveredTarget");
-                    args.add("true");
-                    args.add("--coveredTargetFile");
-                    String path = "target/coveredTargets/testobjImpacts_CoveredTargetsBy"+method.toString()+".csv";
-                    args.add(path);
-                    args.add("--coveredTargetSortedBy");
-                    args.add("TEST");
-
-                    Solution<RestIndividual> solution = initAndRun(args);
-
-                    assertTrue(solution.getIndividuals().size() >= 1);
-
-                    assertTrue(Files.exists(Paths.get(path)));
-
-                    try {
-                        Set tests = new HashSet<String>();
[…]
