### 2020-04-02 bb6123b
updating release instructions
