+++ README.md
@@ -71,7 +71,13 @@ __Known limitations__:
 * External services (e.g., other RESTful APIs): currently there is no support for them (e.g., to automatically mock them).
   It is work in progress.
   
-* NoSQL databases (e.g., MongoDB): currently no support. It is work in progress.    
+* NoSQL databases (e.g., MongoDB): currently no support. It is work in progress. 
+
+* Failing tests: the tests generated by _EvoMaster_ should all pass, and not fail, even when they detect a fault.
+  In those cases, comments/test-names would point out that a test is revealing a possible fault, while still passing.
+  However, in some cases the generated tests might fail. This is due to the so called _flaky_ tests, e.g., when
+  a test has assertions based on the time clock (e.g., dates and timestamps). 
+  There is ongoing effort to address this problem, but it is still not fully solved.   
 
 <!--### Videos---> 
 <!-- 
