+++ client-java/controller/src/main/java/org/evomaster/client/java/controller/mongo/SummaryFindResult.java
@@ -1,6 +1,6 @@
 package org.evomaster.client.java.controller.mongo;
 
-import org.evomaster.client.java.controller.api.dto.database.execution.FindResultDto;
+import org.evomaster.client.java.controller.api.dto.mongo.FindResultDto;
 
 public class SummaryFindResult extends FindResult {
 
