+++ client-java/controller/src/main/java/org/evomaster/client/java/controller/mongo/DetailedFindResult.java
@@ -1,15 +1,18 @@
 package org.evomaster.client.java.controller.mongo;
 
 import org.bson.Document;
-import org.evomaster.client.java.controller.api.dto.database.execution.FindResultDto;
+import org.evomaster.client.java.controller.api.dto.mongo.FindResultDto;
 
 import java.util.ArrayList;
+import java.util.List;
+import java.util.Objects;
 
 public class DetailedFindResult extends FindResult {
 
-    private final ArrayList<Document> documentList = new ArrayList<>();
+    private final List<Document> documentList = new ArrayList<>();
 
     public void addDocument(Document document) {
+        Objects.requireNonNull(document);
         documentList.add(document);
     }
 
@@ -21,7 +24,7 @@ public class DetailedFindResult extends FindResult {
     public FindResultDto toDto() {
         FindResultDto dto = new FindResultDto();
         dto.findResultType = FindResultDto.FindResultType.DETAILED;
-        dto.documents = this.documentList.toArray(new Document[]{});
+        dto.documents = new ArrayList<>(this.documentList);
         dto.hasReturnedAnyDocument = !this.documentList.isEmpty();
         return dto;
     }
