### 2020-04-07 16906f8
Change parsing resource Id and change to template when parsing location fails
