+++ client-java/controller-api/src/main/java/org/evomaster/client/java/controller/api/EMTestUtils.java
@@ -37,7 +37,7 @@ public class EMTestUtils {
         try{
             locationURI = URI.create(locationHeader);
         } catch (Exception e){
-            return locationHeader;
+            return expectedTemplate;
         }
         String locationPath = locationURI.getPath();
         String[] locationTokens = locationPath.split("/");
