### 2020-04-09 ccd315d
refactoring tracker methods into method replacements
