+++ client-java/instrumentation/src/test/java/org/evomaster/client/java/instrumentation/coverage/methodreplacement/ReplacementListTest.java
@@ -1,5 +1,6 @@
 package org.evomaster.client.java.instrumentation.coverage.methodreplacement;
 
+import org.evomaster.client.java.instrumentation.coverage.methodreplacement.classes.WebRequestClassReplacement;
 import org.evomaster.client.java.instrumentation.shared.ReplacementType;
 import org.junit.jupiter.api.Test;
 
@@ -17,6 +18,9 @@ class ReplacementListTest {
 
         for (MethodReplacementClass mrc : ReplacementList.getList()) {
 
+            //make sure that during testing all third-party libraries are available
+            assertTrue(mrc.isAvailable(), "Not available: " + mrc.getClass().getName());
+
             for (Method m : mrc.getClass().getDeclaredMethods()) {
 
                 Replacement r = m.getAnnotation(Replacement.class);
@@ -33,25 +37,38 @@ class ReplacementListTest {
                 }
 
                 Class[] inputs = m.getParameterTypes();
-                assertTrue(inputs.length>0, "Should always be at least 1 parameter, eg the idTemplate");
-                assertEquals(String.class, inputs[inputs.length-1], "Last parameter should always be the idTemplate");
-
                 Class<?> targetClass = mrc.getTargetClass();
                 assertNotNull(targetClass);
 
+                if(r.type() != ReplacementType.TRACKER){
+                    assertTrue(inputs.length>0, "Should always be at least 1 parameter, eg the idTemplate");
+                    assertEquals(String.class, inputs[inputs.length-1], "Last parameter should always be the idTemplate");
+                }
+
                 if(! r.replacingStatic()){
                     //if not replacing a static method, then caller must be passed as first input
-                    assertEquals(targetClass, inputs[0]);
-                    assertTrue(inputs.length>=2);// caller and idTemplate
+                    assertTrue(inputs.length >= 1);// caller
+
+                    if(mrc instanceof ThirdPartyMethodReplacementClass) {
+                        //must always be Object when dealing with third-party library replacements
+                        assertEquals(Object.class, inputs[0]);
+                    } else {
+                        assertEquals(targetClass, inputs[0]);
+                    }
                 }
 
-
-                Class[] reducedInputs;
-                if(r.replacingStatic()){
-                    reducedInputs = Arrays.copyOfRange(inputs, 0, inputs.length-1);
-                } else {
-                    reducedInputs = Arrays.copyOfRange(inputs, 1, inputs.length-1);
+                int start = 0;
+                if(!r.replacingStatic()){
+                    start = 1;
                 }
+
+                int end = inputs.length-1;
+                if(r.type() == ReplacementType.TRACKER){
+                    //no idTemplate at the end
+                    end = inputs.length;
+                }
+
+                Class[] reducedInputs = Arrays.copyOfRange(inputs, start, end);
 
                 Method targetMethod = null;
                 try {
