+++ client-java/instrumentation/src/main/java/org/evomaster/client/java/instrumentation/coverage/MethodReplacementMethodVisitor.java
@@ -1,11 +1,12 @@
 package org.evomaster.client.java.instrumentation.coverage;
 
-import org.evomaster.client.java.instrumentation.shared.ClassName;
 import org.evomaster.client.java.instrumentation.Constants;
-import org.evomaster.client.java.instrumentation.shared.ObjectiveNaming;
 import org.evomaster.client.java.instrumentation.coverage.methodreplacement.MethodReplacementClass;
 import org.evomaster.client.java.instrumentation.coverage.methodreplacement.Replacement;
 import org.evomaster.client.java.instrumentation.coverage.methodreplacement.ReplacementList;
+import org.evomaster.client.java.instrumentation.shared.ClassName;
+import org.evomaster.client.java.instrumentation.shared.ObjectiveNaming;
+import org.evomaster.client.java.instrumentation.shared.ReplacementType;
 import org.evomaster.client.java.instrumentation.staticstate.ObjectiveRecorder;
 import org.evomaster.client.java.instrumentation.staticstate.UnitsInfoRecorder;
 import org.evomaster.client.java.utils.SimpleLogger;
@@ -24,11 +25,13 @@ public class MethodReplacementMethodVisitor extends MethodVisitor {
     private final String className;
     private final String methodName;
     private final boolean registerNewTargets;
+    private final boolean applyTrackingMethods;
 
     private int currentLine;
     private int currentIndex;
 
     public MethodReplacementMethodVisitor(boolean registerNewTargets,
+                                          boolean applyTrackingMethods,
                                           MethodVisitor mv,
                                           String className,
                                           String methodName,
@@ -38,6 +41,7 @@ public class MethodReplacementMethodVisitor extends MethodVisitor {
         this.className = className;
         this.methodName = methodName;
         this.registerNewTargets = registerNewTargets;
+        this.applyTrackingMethods = applyTrackingMethods;
         currentLine = 0;
     }
 
@@ -59,18 +63,10 @@ public class MethodReplacementMethodVisitor extends MethodVisitor {
             return;
         }
 
-        /*
-            Loading class here could have side-effects.
-            However, method replacements is only done for JDK APIs, which
-            anyway should be loaded with boot-classloader.
-            so, loading them here "hopefully" should be OK...
-
-            TODO: in future, might also target methods in Kotlin API
-         */
-        if (!owner.startsWith("java/")) {
-            super.visitMethodInsn(opcode, owner, name, desc, itf);
-            return;
-        }
+//        if (!owner.startsWith("java/")) {
+//            super.visitMethodInsn(opcode, owner, name, desc, itf);
+//            return;
+//        }
 
 
         /*
@@ -86,11 +82,14 @@ public class MethodReplacementMethodVisitor extends MethodVisitor {
          * in Guava when they have "super.()" calls.
          * For now, we just skip them.
          */
-        if(opcode == Opcodes.INVOKESPECIAL){
+        if (opcode == Opcodes.INVOKESPECIAL) {
             super.visitMethodInsn(opcode, owner, name, desc, itf);
             return;
         }
 
+        /*
+            Loading class here could have side-effects if code is executed in static initializer.
+         */
         Class<?> klass = null;
         try {
             klass = this.getClass().getClassLoader().loadClass(ClassName.get(owner).getFullNameWithDots());
@@ -113,8 +112,13 @@ public class MethodReplacementMethodVisitor extends MethodVisitor {
                 .filter(m -> m.getName().equals(name))
                 .filter(m -> {
                     Replacement br = m.getAnnotation(Replacement.class);
[…]
