+++ client-java/instrumentation/src/main/java/org/evomaster/client/java/instrumentation/coverage/methodreplacement/MethodReplacementClass.java
@@ -9,4 +9,17 @@ public interface MethodReplacementClass {
      * The target class in the JDK this class provides replacements for
      */
     Class<?> getTargetClass();
+
+
+    default String getTargetClassName(){
+        return getTargetClass().getName();
+    }
+
+    /**
+     * Depending on the SUT, third-party libraries might not be present
+     * on the classpath
+     */
+    default boolean isAvailable(){
+        return getTargetClass() != null;
+    }
 }

[…]
