### 2020-04-09 0c3a8c1
by default, RemoteController has computeSqlHeuristics, extractSqlExecutionInfo and extractMongoExecutionInfo disabled
