+++ client-java/controller-api/src/main/java/org/evomaster/client/java/controller/api/dto/SutRunDto.java
@@ -33,13 +33,13 @@ public class SutRunDto {
     public SutRunDto() {
     }
 
-    public SutRunDto(Boolean run, Boolean resetState,
+    public SutRunDto(Boolean run,
+                     Boolean resetState,
                      Boolean calculateSqlHeuristics,
                      Boolean extractSqlExecutionInfo,
                      Boolean extractMongoExecutionInfo) {
         if (calculateSqlHeuristics != null && calculateSqlHeuristics && extractSqlExecutionInfo != null && !extractSqlExecutionInfo)
             throw new IllegalArgumentException("extractSqlExecutionInfo should be enabled when calculateSqlHeuristics is enabled");
-
 
         this.run = run;
         this.resetState = resetState;
@@ -48,7 +48,11 @@ public class SutRunDto {
         this.extractMongoExecutionInfo = extractMongoExecutionInfo;
     }
 
-    public SutRunDto(Boolean run, Boolean resetState, Boolean calculateSqlHeuristics) {
-        this(run, resetState, calculateSqlHeuristics, calculateSqlHeuristics != null && calculateSqlHeuristics, false);
+    public SutRunDto(Boolean run, Boolean resetState, Boolean calculateSqlHeuristics, Boolean extractMongoExecutionInfo) {
+        this(run,
+                resetState,
+                calculateSqlHeuristics,
+                calculateSqlHeuristics != null && calculateSqlHeuristics,
+                extractMongoExecutionInfo != null && extractMongoExecutionInfo);
     }
 }
