+++ client-java/controller/src/main/java/org/evomaster/client/java/controller/internal/EMController.java
@@ -184,7 +184,7 @@ public class EMController {
             boolean mongoExecution = dto.extractMongoExecutionInfo != null && dto.extractMongoExecutionInfo;
 
             sutController.enableComputeSqlHeuristicsOrExtractExecution(sqlHeuristics, sqlExecution);
-            sutController.enableExtractMongoExecution(mongoExecution);
+            sutController.enableExtractMongoExecutionInfo(mongoExecution);
 
             boolean doReset = dto.resetState != null && dto.resetState;
 
