### 2020-04-14 #2153 @arfon
@arcuri82 - could you advise if there are any major changes between `v1.0.0` and `v1.0.2`?
