### 2020-04-14 #2153 @arcuri82
the project is an active development (as it has been for the last 3.5 years).
The latest release is `v1.0.1` (see [release page](https://github.com/EMResearch/EvoMaster/releases)), and it was made 2 weeks ago. It is just bug fixes over `v1.0.0`.

The version `1.0.2-SNAPSHOT` is simply the working version in the master branch.

Which one to review? Both would be equivalent, as all new features under development are deactivated by default. It all depends if you want to download an already packaged jar file from the release page, or build the jar directly
