### 2020-04-24 EMResearch/EvoMaster#169 @mado89
Confused about ports in EmbeddedEvoMasterController
Hi,
I have a question about the EmbeddedEvoMasterController. What is the meaning of the port parameter? It seems to be ignored.
Take https://github.com/EMResearch/EMB/blob/master/em/embedded/rest/news/src/main/java/em/embedded/org/tsdes/EmbeddedEvoMasterController.java for example.
* In line 29 a local variable port is declared but seems to be unused.
* Since in Line 52 0 is passed as parameter. https://github.com/EMResearch/EMB/blob/9f769ffab5d5c3e8836c1bae6354863adee316ca/em/embedded/rest/ncs/src/main/java/em/embedded/org/restncs/EmbeddedEvoMasterController.java#L52

Am I missing something here?

https://github.com/openjournals/joss-reviews/issues/2153
