### 2020-04-24 EMResearch/EvoMaster#170 @mado89
Qualitity of the Paper
Hi,

while (for now) I do not see much which would hinder an acceptance of the software I'd like to have some major improvements on the paper itself.

Main points (so far)
* Referenced literature consists of self citations only
* Statement of need / Addressed research question does not come through (for me)
* No related work / state of field at all (see first point)
* From what I can see you did quite some benchmarking. I'd like a summary of your results to be featured in the paper. Something along the lines of "When I let it run for the suggested 24h (black box testing), can I use my computer meanwhile or do I have to run it on a different machine?".

https://github.com/openjournals/joss-reviews/issues/2153
