### 2020-04-24 EMResearch/EvoMaster#169 @arcuri82
Hi,
 it is not ignored :) Somehow, this is briefly discussed in the [documentation](https://github.com/EMResearch/EvoMaster/blob/master/docs/write_driver.md): "When starting the SUT, there are at least two configurations that you want to change: the binding port, as you want to use 0 for ephemeral ports (to avoid port conflicts)".

You can learn more about it by googling for "ephemeral port" and "tcp port 0".

Also from the documentation:

```
The actual chosen port can then be extracted with:

protected int getSutPort() {
        return (Integer) ((Map) ctx.getEnvironment()
                .getPropertySources().get("server.ports").getSource())
                .get("local.server.port");
}
```

Do you think we should add some more info on what ephemeral ports are?
