### 2020-04-24 42a4fda
fix issue for class name handling
