### 2020-04-24 EMResearch/EvoMaster#169 @arcuri82
I guess here also a possible source of confusion is that there are 2 ports involved:
- the one of EvoMaster driver itself (default 40100, but that can be changed from command-line)
- the one of the SUT (System Under Test), which should be ephemeral
