### 2020-04-24 EMResearch/EvoMaster#169 @mado89
Yes that was the confusion
