### 2020-04-24 EMResearch/EvoMaster#169 @arcuri82
ok, fair enough :) I ll add some more documentation about it
