### 2020-04-24 EMResearch/EvoMaster#169 @arcuri82
hi,
 do the documentation changes in #171 clarify it? Or should I add some more info?
