+++ docs/joss-paper/paper.bib
@@ -4,7 +4,8 @@
   booktitle="IEEE International Conference on Software Testing, Verification and Validation (ICST)",
   pages={394-397},
   year={2018},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/icst.2018.00046}
 }
 
 
@@ -13,7 +14,8 @@
     title = {{Many Independent Objective (MIO) Algorithm for Test Suite Generation}},
     booktitle = "International Symposium on Search Based Software Engineering (SSBSE)",
     pages={3--17},
-    year = {2017}
+    year = {2017},
+    doi={10.1007/978-3-319-66299-2_1}
 }
 
 @article{arcuri2018test,
@@ -23,7 +25,7 @@ journal={Information and Software Technology},
 volume={104},
 pages={195--206},
 year={2018},
-publisher={Elsevier}
+publisher={Elsevier},
 }
 
 @inproceedings{arcuri2017restful,
@@ -32,7 +34,8 @@ publisher={Elsevier}
   booktitle="IEEE International Conference on Software Quality, Reliability and Security (QRS)",
   pages={9--20},
   year={2017},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/qrs.2017.11}
 }
 
 @article{arcuri2019restful,
@@ -52,7 +55,8 @@ author={Zhang, Man and Marculescu, Bogdan and Arcuri, Andrea},
 booktitle="Genetic and Evolutionary Computation Conference (GECCO)",
 pages={1426--1434},
 year={2019},
-organization={ACM}
+organization={ACM},
+doi={10.1145/3321707.3321815}
 }
 
 
@@ -62,7 +66,8 @@ author={Arcuri, Andrea and Galeotti, Juan P},
 booktitle="Genetic and Evolutionary Computation Conference (GECCO)",
 pages={1390--1398},
 year={2019},
-organization={ACM}
+organization={ACM},
+doi={10.1145/3321707.3321732}
 }
 
 @inproceedings{arcuri2020testability,
@@ -72,4 +77,80 @@ booktitle="IEEE International Conference on Software Testing, Verification and V
 pages={},
 year={2020},
 organization={IEEE}
+}
+
+
+@misc{tricentis2017,
+    title = {{Software Fail Watch 5th Edition}},
+    howpublished = {https://www.tricentis.com/resources/software-fail-watch-5th-edition/}
+}
+
+
+@article{harman2012search,
+  title={Search-based software engineering: Trends, techniques and applications},
+  author={M. Harman and S. A. Mansouri and Y. Zhang},
+  journal={ACM Computing Surveys (CSUR)},
+  volume={45},
+  number={1},
+  pages={11},
+  year={2012},
+  publisher={ACM}
+}
+
+
+
+@book{allamaraju2010restful,
+  title={Restful web services cookbook: solutions for improving scalability and simplicity},
+  author={Allamaraju, Subbu},
+  year={2010},
+  publisher={" O'Reilly Media, Inc."}
+}
+
+@phdthesis{fielding2000architectural,
+  title={Architectural styles and the design of network-based software architectures},
+  author={Fielding, Roy Thomas},
+  year={2000},
+  school={University of California, Irvine}
+}
+
+@inproceedings{restlerICSE2019,
+author = {Atlidakis, Vaggelis and Godefroid, Patrice and Polishchuk, Marina},
+title = {RESTler: Stateful REST API Fuzzing},
+year = {2019},
+publisher = {IEEE Press},
+url = {https://doi.org/10.1109/ICSE.2019.00083},
+doi = {10.1109/ICSE.2019.00083},
[…]
