+++ docs/joss-paper/paper.bib
@@ -81,13 +81,14 @@ organization={IEEE}
 
 
 @misc{tricentis2017,
+    author = {Tricentis}
     title = {{Software Fail Watch 5th Edition}},
     howpublished = {https://www.tricentis.com/resources/software-fail-watch-5th-edition/}
 }
 
 
 @article{harman2012search,
-  title={Search-based software engineering: Trends, techniques and applications},
+  title={{Search-based software engineering: Trends, techniques and applications}},
   author={M. Harman and S. A. Mansouri and Y. Zhang},
   journal={ACM Computing Surveys (CSUR)},
   volume={45},
@@ -115,7 +116,7 @@ organization={IEEE}
 
 @inproceedings{restlerICSE2019,
 author = {Atlidakis, Vaggelis and Godefroid, Patrice and Polishchuk, Marina},
-title = {RESTler: Stateful REST API Fuzzing},
+title = {{RESTler: Stateful REST API Fuzzing}},
 year = {2019},
 publisher = {IEEE Press},
 url = {https://doi.org/10.1109/ICSE.2019.00083},
@@ -129,18 +130,18 @@ series = {ICSE ’19}
 
 
 @inproceedings{karlsson2020QuickREST,
-title={QuickREST: Property-based Test Generation of OpenAPI Described RESTful APIs},
+title={{QuickREST: Property-based Test Generation of OpenAPI Described RESTful APIs}},
 author={Karlsson, Stefan and Causevic, Adnan and Sundmark, Daniel},
-booktitle=ICST,
+booktitle="IEEE International Conference on Software Testing, Verification and Validation (ICST)",
 pages={},
 year={2020},
 organization={IEEE}
 }
 
 @inproceedings{viglianisi2020resttestgen,
-title={RESTTESTGEN: Automated Black-Box Testing of RESTful APIs},
+title={{RESTTESTGEN: Automated Black-Box Testing of RESTful APIs}},
 author={Viglianisi, Emanuele and Dallago, Michael and Ceccato, Mariano},
-booktitle=ICST,
+booktitle="IEEE International Conference on Software Testing, Verification and Validation (ICST)",
 pages={},
 year={2020},
 organization={IEEE}
@@ -149,7 +150,7 @@ organization={IEEE}
 @inproceedings{eddouibi2018automatic,
 	author={Ed-douibi, Hamza and Cánovas Izquierdo, Javier Luis and Cabot, Jordi },
 	booktitle={2018 IEEE 22nd International Enterprise Distributed Object Computing Conference (EDOC)},
-	title={Automatic Generation of Test Cases for REST APIs: A Specification-Based Approach},
+	title={{Automatic Generation of Test Cases for REST APIs: A Specification-Based Approach}},
 	year={2018},
 	volume={}, number={},
 	pages={181-190}
