+++ docs/joss-paper/paper.bib
@@ -81,7 +81,7 @@ organization={IEEE}
 
 
 @misc{tricentis2017,
-    author = {Tricentis}
+    author = {Tricentis},
     title = {{Software Fail Watch 5th Edition}},
     howpublished = {https://www.tricentis.com/resources/software-fail-watch-5th-edition/}
 }
