+++ docs/joss-paper/paper.bib
@@ -83,7 +83,8 @@ organization={IEEE}
 @misc{tricentis2017,
     author = {Tricentis},
     title = {{Software Fail Watch 5th Edition}},
-    howpublished = {https://www.tricentis.com/resources/software-fail-watch-5th-edition/}
+    howpublished = {https://www.tricentis.com/resources/software-fail-watch-5th-edition/},
+    year={2017}
 }
 
 
