+++ docs/joss-paper/paper.bib
@@ -1,7 +1,7 @@
 @inproceedings{arcuri2018evomaster,
   title={{EvoMaster: Evolutionary Multi-context Automated System Test Generation}},
   author={Arcuri, Andrea},
-  booktitle="IEEE International Conference on Software Testing, Verification and Validation (ICST)",
+  booktitle={{IEEE International Conference on Software Testing, Verification and Validation (ICST)}},
   pages={394-397},
   year={2018},
   organization={IEEE},
@@ -12,7 +12,7 @@
 @INPROCEEDINGS{mio2017,
     author = {Arcuri, Andrea},
     title = {{Many Independent Objective (MIO) Algorithm for Test Suite Generation}},
-    booktitle = "International Symposium on Search Based Software Engineering (SSBSE)",
+    booktitle = {{International Symposium on Search Based Software Engineering (SSBSE)}},
     pages={3--17},
     year = {2017},
     doi={10.1007/978-3-319-66299-2_1}
@@ -21,7 +21,7 @@
 @article{arcuri2018test,
 title={{Test suite generation with the Many Independent Objective (MIO) algorithm}},
 author={Arcuri, Andrea},
-journal={Information and Software Technology},
+journal={{Information and Software Technology}},
 volume={104},
 pages={195--206},
 year={2018},
@@ -31,7 +31,7 @@ publisher={Elsevier},
 @inproceedings{arcuri2017restful,
   title={{RESTful API Automated Test Case Generation}},
   author={Arcuri, Andrea},
-  booktitle="IEEE International Conference on Software Quality, Reliability and Security (QRS)",
+  booktitle={{IEEE International Conference on Software Quality, Reliability and Security (QRS)}},
   pages={9--20},
   year={2017},
   organization={IEEE},
@@ -41,7 +41,7 @@ publisher={Elsevier},
 @article{arcuri2019restful,
 title={{RESTful API Automated Test Case Generation with EvoMaster}},
 author={Arcuri, Andrea},
-journal="ACM Transactions on Software Engineering and Methodology (TOSEM)",
+journal={{ACM Transactions on Software Engineering and Methodology (TOSEM)}},
 volume={28},
 number={1},
 pages={3},
@@ -52,7 +52,7 @@ publisher={ACM}
 @inproceedings{zhang2019resource,
 title={{Resource-based test case generation for RESTful web services}},
 author={Zhang, Man and Marculescu, Bogdan and Arcuri, Andrea},
-booktitle="Genetic and Evolutionary Computation Conference (GECCO)",
+booktitle={{Genetic and Evolutionary Computation Conference (GECCO)}},
 pages={1426--1434},
 year={2019},
 organization={ACM},
@@ -63,7 +63,7 @@ doi={10.1145/3321707.3321815}
 @inproceedings{arcuri2019sql,
 title={{SQL data generation to enhance Search-Based System Testing}},
 author={Arcuri, Andrea and Galeotti, Juan P},
-booktitle="Genetic and Evolutionary Computation Conference (GECCO)",
+booktitle={{Genetic and Evolutionary Computation Conference (GECCO)}},
 pages={1390--1398},
 year={2019},
 organization={ACM},
@@ -73,7 +73,7 @@ doi={10.1145/3321707.3321732}
 @inproceedings{arcuri2020testability,
 title={{Testability Transformations For Existing APIs}},
 author={Arcuri, Andrea and Galeotti, Juan P.},
-booktitle="IEEE International Conference on Software Testing, Verification and Validation (ICST)",
+booktitle={{IEEE International Conference on Software Testing, Verification and Validation (ICST)}},
 pages={},
 year={2020},
 organization={IEEE}
[…]
