### 2020-04-24 EMResearch/EvoMaster#170 @arcuri82
Hi,
 thanks. I updated the paper, addressing your first 3 points. However, I am not sure about your last one: "When I let it run for the suggested 24h (black box testing), can I use my computer meanwhile or do I have to run it on a different machine?"
That all depends on the hardware: ie, how many cores and how much RAM is available (and network connection in case of black-box testing of a remote service). It also  depends on the system under test (e.g., does it use a database?).
We currently do not have any published data on timing in respect to the hardware involved (all experiments run so far were based on number of HTTP calls, to abstract from all the issues of doing analyses that are time and hardware dependent).
