### 2020-04-24 #2153 @arcuri82
thanks @mado89 for your feedback. i made changes and updated the paper. Let me know if you require further changes

Question to @arfon: is @jaynanavati active? I was looking at his/her GitHub profile today, and it does not seem there is any activity whatsoever on GitHub:  https://github.com/jaynanavati?tab=overview&from=2020-04-01&to=2020-04-24
Or did I miss something?  (I apologize in that case)
