### 2020-04-25 #2153 @arfon
> Question to @arfon: is @jaynanavati active? I was looking at his/her GitHub profile today, and it does not seem there is any activity whatsoever on GitHub: https://github.com/jaynanavati?tab=overview&from=2020-04-01&to=2020-04-24

Not clear. I'd like us to wait longer before considering changing out reviewers, especially as so many people's lives are severely disrupted by COVID-19.
