### 2020-04-25 e71078f
candidate fix for LinkageError
