+++ client-java/instrumentation/src/main/java/org/evomaster/client/java/instrumentation/Instrumentator.java
@@ -65,21 +65,9 @@ public class Instrumentator {
         reader.accept(cn, readFlags);
 
         if(canInstrumentForCoverage(className)){
-
             cv = new CoverageClassVisitor(cv, className);
-
-            /*
-                this should be done after coverage instrumentation, as
-                we don't want these extra methods added as part of
-                targets to cover
-             */
-            //cv = new TrackerClassVisitor(cv, className);
-
         } else {
-
             cv = new ThirdPartyClassVisitor(cv, className);
-
-            //reader.accept(cv, readFlags);
         }
 
         cn.accept(cv);
