### 2020-04-28 #2153 @arfon
👋 @feliperodri & @s0nata - we're looking for a second reviewer of the [EvoMaster](https://github.com/EMResearch/EvoMaster) package for JOSS.

Would either of you be willing to review this submission for us? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
