### 2020-04-28 6b4f86d
Dto to OpenApi schema converter
