+++ client-java/instrumentation/src/test/java/org/evomaster/client/java/instrumentation/object/ClassToSchemaTest.java
@@ -1,0 +1,144 @@
+package org.evomaster.client.java.instrumentation.object;
+
+import com.google.gson.Gson;
+import com.google.gson.JsonObject;
+import com.google.gson.JsonParser;
+import org.evomaster.client.java.instrumentation.object.dtos.*;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+public class ClassToSchemaTest {
+
+    private static final Gson GSON = new Gson();
+
+    private JsonObject parse(String schema){
+        return JsonParser.parseString("{" + schema + "}").getAsJsonObject();
+    }
+
+    @Test
+    public void testBase(){
+
+        String schema = ClassToSchema.getOrDeriveSchema(DtoBase.class);
+        JsonObject json = parse(schema);
+
+        JsonObject obj = json.get(DtoBase.class.getName()).getAsJsonObject();
+        assertNotNull(obj);
+        assertNotNull(obj.get("type"));
+        assertNotNull(obj.get("properties"));
+        assertEquals(2, obj.entrySet().size());
+        assertEquals("object", obj.get("type").getAsString());
+
+        assertEquals(1, obj.get("properties").getAsJsonObject().entrySet().size());
+        assertEquals("string", obj.get("properties").getAsJsonObject()
+                .get("foo").getAsJsonObject().get("type").getAsString());
+    }
+
+
+    private void verifyTypeOfFieldInProperties(JsonObject obj, String expected, String fieldName){
+        assertEquals(expected, obj.get("properties").getAsJsonObject()
+                .get(fieldName).getAsJsonObject().get("type").getAsString());
+    }
+
+    private void verifyTypeAndFormatOfFieldInProperties(JsonObject obj, String type, String format, String fieldName){
+        JsonObject field = obj.get("properties").getAsJsonObject().get(fieldName).getAsJsonObject();
+        assertEquals(type, field.get("type").getAsString());
+        assertEquals(format, field.get("format").getAsString());
+    }
+
+
+
+    @Test
+    public void testNumeric(){
+
+        String schema = ClassToSchema.getOrDeriveSchema(DtoNumeric.class);
+        JsonObject json = parse(schema);
+
+        JsonObject obj = json.get(DtoNumeric.class.getName()).getAsJsonObject();
+        assertNotNull(obj);
+        assertNotNull(obj.get("type"));
+        assertNotNull(obj.get("properties"));
+        assertEquals(2, obj.entrySet().size());
+        assertEquals("object", obj.get("type").getAsString());
+
+        assertEquals(12, obj.get("properties").getAsJsonObject().entrySet().size());
+        verifyTypeAndFormatOfFieldInProperties(obj, "integer", "int8", "byte_p");
+        verifyTypeAndFormatOfFieldInProperties(obj, "integer", "int8", "byte_w");
+        verifyTypeAndFormatOfFieldInProperties(obj, "integer", "int16", "short_p");
+        verifyTypeAndFormatOfFieldInProperties(obj, "integer", "int16", "short_w");
+        verifyTypeAndFormatOfFieldInProperties(obj, "integer", "int32", "integer_p");
+        verifyTypeAndFormatOfFieldInProperties(obj, "integer", "int32", "integer_w");
+        verifyTypeAndFormatOfFieldInProperties(obj, "integer", "int64", "long_p");
+        verifyTypeAndFormatOfFieldInProperties(obj, "integer", "int64", "long_w");
+        verifyTypeAndFormatOfFieldInProperties(obj, "number", "float", "float_p");
+        verifyTypeAndFormatOfFieldInProperties(obj, "number", "float", "float_w");
[…]
