### 2020-05-03 105a8d3
fixing handling of GSON
