+++ client-java/controller/src/main/java/org/evomaster/client/java/controller/internal/SutController.java
@@ -401,6 +401,7 @@ public abstract class SutController implements SutHandler {
         dto.numberOfReplacedMethodsInThirdParty = recorder.getNumberOfReplacedMethodsInThirdParty();
         dto.numberOfTrackedMethods = recorder.getNumberOfTrackedMethods();
         dto.unitNames = recorder.getUnitNames();
+        dto.parsedDtos = recorder.getParsedDtos();
         return dto;
     }
 }
