### 2020-05-03 98eab17
fixing handling of GSON
