+++ client-java/instrumentation/src/main/java/org/evomaster/client/java/instrumentation/coverage/methodreplacement/classes/GsonClassReplacement.java
@@ -2,8 +2,11 @@ package org.evomaster.client.java.instrumentation.coverage.methodreplacement.cla
 
 import org.evomaster.client.java.instrumentation.coverage.methodreplacement.Replacement;
 import org.evomaster.client.java.instrumentation.coverage.methodreplacement.ThirdPartyMethodReplacementClass;
+import org.evomaster.client.java.instrumentation.coverage.methodreplacement.UsageFilter;
 import org.evomaster.client.java.instrumentation.object.ClassToSchema;
 import org.evomaster.client.java.instrumentation.shared.ReplacementType;
+import org.evomaster.client.java.instrumentation.staticstate.ExecutionTracer;
+import org.evomaster.client.java.instrumentation.staticstate.UnitsInfoRecorder;
 
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
@@ -19,13 +22,17 @@ public class GsonClassReplacement extends ThirdPartyMethodReplacementClass {
 
     // TODO all versions of fromJson
 
-    @Replacement(replacingStatic = false, type = ReplacementType.TRACKER, id = "fromJson_string_class")
+    @Replacement(replacingStatic = false,
+            type = ReplacementType.TRACKER,
+            id = "fromJson_string_class",
+            usageFilter = UsageFilter.ONLY_SUT)
     public static Object fromJson(Object caller, String json, Class<?> classOfT){
 
         if(classOfT != null) {
+            String name = classOfT.getName();
             String schema = ClassToSchema.getOrDeriveSchema(classOfT);
-
-            //TODO
+            UnitsInfoRecorder.registerNewParsedDto(name, schema);
+            ExecutionTracer.addParsedDtoName(name);
         }
 
         Method original = getOriginal(singleton, "fromJson_string_class");
