### 2020-05-06 63ec757
replaced Gson with Jackson / Third-party replacemente for MongoDB
