### 2020-05-07 bdfe0da
Fix tests and add special method to print snapshots
