### 2020-05-07 EMResearch/EvoMaster#170 @s0nata
Hi, I'll join this discussion with remarks on the paper quality on my end:

Statement of Need section:
- [ ] paragraph 2: "... techniques aimed at automating several different tasks in **SE**" --> did you mean "in testing"?
- [ ] do you think you could reformulate the sentence about the potential research users? To me it seems too vague for its current length (just shortening is fine).

Tool Summary section, paragraph 2:
- [ ] could you maybe list a couple examples of the challenges you refer to?
- [ ] it is not clear to me from the text if `EvoMaster` incorporates the techniques listed, or was used as an auxiliary tool in  experiments. It is also a bit odd to read about publications in the section that describes tool functionalities. If there was an explicit list of basic and additional functionalities of the tool, I would have found it helpful both in understanding what the tool does and how it compares to the competition.

References:
- [ ] could you please consistently provide DOIs/links to cited publications?

[openjournals/joss-reviews#2153](https://github.com/openjournals/joss-reviews/issues/2153)
