### 2020-05-07 #2153 @s0nata
Hi @arcuri82, I have just added my comments on the paper text to the issue [#170](https://github.com/EMResearch/EvoMaster/issues/170)
