### 2020-05-07 130b8e9
removed mongoDB dependency from instrumentation-shared
