+++ client-java/instrumentation/src/main/java/org/evomaster/client/java/instrumentation/mongo/MongoLogger.java
@@ -8,9 +8,19 @@ import java.io.PrintStream;
 import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
+import java.util.Objects;
 
 public class MongoLogger {
 
+    public static final String GET_NAMESPACE = "getNamespace";
+    public static final String GET_DATABASE_NAME = "getDatabaseName";
+    public static final String GET_COLLECTION_NAME = "getCollectionName";
+    public static final String GET_CODEC_REGISTRY = "getCodecRegistry";
+    public static final String TO_BSON_DOCUMENT = "toBsonDocument";
+    public static final String BUILDER = "builder";
+    public static final String BUILD = "build";
+    public static final String AS_BSON_READER = "asBsonReader";
+    public static final String DECODE = "decode";
     private static MongoLogger instance = null;
 
     private PrintStream outputStream = null;
@@ -36,80 +46,83 @@ public class MongoLogger {
 
     public static final String PREFIX = "MONGO_LOGGER";
 
+    private static Object invokeMethod(Class<?> receiverClass, Object receiver, String methodName)
+            throws NoSuchMethodException, InvocationTargetException, IllegalAccessException {
+        return invokeMethod(receiverClass, receiver, methodName, new Class[]{}, new Object[]{});
+    }
+
+    private static Object invokeMethod(Class<?> receiverClass, Object receiver, String methodName, Class[] argumentTypes, Object[] arguments)
+            throws NoSuchMethodException, InvocationTargetException, IllegalAccessException {
+
+        Objects.requireNonNull(receiverClass);
+        Objects.requireNonNull(methodName);
+        Objects.requireNonNull(argumentTypes);
+        Objects.requireNonNull(arguments);
+        if (argumentTypes.length != arguments.length) {
+            throw new IllegalArgumentException("Mismatch number of argument types and concrete arguments. Argument types "
+                    + argumentTypes.length + " while concrete arguments were "
+                    + arguments.length);
+        }
+
+        Method getNamespaceMethod = receiverClass.getMethod(methodName, argumentTypes);
+        boolean getNamespaceMethodIsAccessible = getNamespaceMethod.isAccessible();
+        getNamespaceMethod.setAccessible(true);
+        Object retVal = getNamespaceMethod.invoke(receiver, arguments);
+        getNamespaceMethod.setAccessible(getNamespaceMethodIsAccessible);
+        return retVal;
+    }
+
+    private static Object invokeConstructor(Class targetClass)
+            throws NoSuchMethodException, InvocationTargetException, IllegalAccessException, InstantiationException {
+        return invokeConstructor(targetClass, new Class[]{}, new Object[]{});
+    }
+
+    private static Object invokeConstructor(Class targetClass, Class[] argumentTypes, Object[] arguments)
+            throws NoSuchMethodException, InvocationTargetException, IllegalAccessException, InstantiationException {
+
+        Objects.requireNonNull(targetClass);
+        Objects.requireNonNull(argumentTypes);
+        Objects.requireNonNull(arguments);
+        if (argumentTypes.length != arguments.length) {
+            throw new IllegalArgumentException("Mismatch number of argument types and concrete arguments. Argument types "
+                    + argumentTypes.length + " while concrete arguments were "
+                    + arguments.length);
+        }
+
+        Constructor<?> documentCodecConstructor = targetClass.getConstructor(argumentTypes);
+        boolean documentCodecConstructorIsAccessible = documentCodecConstructor.isAccessible();
[…]
