### 2020-05-08 EMResearch/EvoMaster#177 @mado89
Whitebox Testing
Hi,

I wanted to follow https://github.com/EMResearch/EvoMaster/blob/master/docs/write_driver.md to create a driver to test your framework using an application I have written.

Some comments / questions (I will add more once they arise):

- [x] Is it correct that I have to download a different software than "evomaster.jar" from github? Using "evomaster.jar" the package "org.evomaster.client.java.controller" is missing.
- [ ] authentication: Could you please provide a link or some information on how to implement 'getInfoForAuthentication'

Thanks!

openjournals/joss-reviews#2153
