### 2020-05-08 EMResearch/EvoMaster#177 @arcuri82
hi,
 thx. you are indeed correct. but u do not download it from GitHub, you download it from Maven Central with Maven/Gradle.
I added a clarification to `write_driver.md`
