### 2020-05-08 EMResearch/EvoMaster#170 @arcuri82
@s0nata: thanks for the feedback. I tried to address all of your concerns, and updated the paper accordingly. Let me know if some parts are still unclear, and/or if you require further clarifications and modifications.
