### 2020-05-08 e8f70d3
refactor to remove dependency of instrumentation to Bson
