+++ client-java/controller-api/src/main/java/org/evomaster/client/java/controller/api/dto/mongo/FindResultDto.java
@@ -1,7 +1,5 @@
 package org.evomaster.client.java.controller.api.dto.mongo;
 
-import org.bson.Document;
-
 import java.util.List;
 
 public class FindResultDto {
@@ -16,6 +14,6 @@ public class FindResultDto {
 
     public boolean hasReturnedAnyDocument;
 
-    public List<Document> documents;
+    public List<DocumentDto> documents;
 
 }
