+++ client-java/instrumentation/src/main/java/org/evomaster/client/java/instrumentation/mongo/MongoLogger.java
@@ -2,10 +2,8 @@ package org.evomaster.client.java.instrumentation.mongo;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.bson.Document;
 
 import java.io.PrintStream;
-import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.util.Objects;
@@ -15,12 +13,7 @@ public class MongoLogger {
     public static final String GET_NAMESPACE = "getNamespace";
     public static final String GET_DATABASE_NAME = "getDatabaseName";
     public static final String GET_COLLECTION_NAME = "getCollectionName";
-    public static final String GET_CODEC_REGISTRY = "getCodecRegistry";
-    public static final String TO_BSON_DOCUMENT = "toBsonDocument";
-    public static final String BUILDER = "builder";
-    public static final String BUILD = "build";
-    public static final String AS_BSON_READER = "asBsonReader";
-    public static final String DECODE = "decode";
+
     private static MongoLogger instance = null;
 
     private PrintStream outputStream = null;
@@ -51,7 +44,11 @@ public class MongoLogger {
         return invokeMethod(receiverClass, receiver, methodName, new Class[]{}, new Object[]{});
     }
 
-    private static Object invokeMethod(Class<?> receiverClass, Object receiver, String methodName, Class[] argumentTypes, Object[] arguments)
+    private static Object invokeMethod(Class<?> receiverClass,
+                                       Object receiver,
+                                       String methodName,
+                                       Class<?>[] argumentTypes,
+                                       Object[] arguments)
             throws NoSuchMethodException, InvocationTargetException, IllegalAccessException {
 
         Objects.requireNonNull(receiverClass);
@@ -72,66 +69,49 @@ public class MongoLogger {
         return retVal;
     }
 
-    private static Object invokeConstructor(Class targetClass)
-            throws NoSuchMethodException, InvocationTargetException, IllegalAccessException, InstantiationException {
-        return invokeConstructor(targetClass, new Class[]{}, new Object[]{});
-    }
-
-    private static Object invokeConstructor(Class targetClass, Class[] argumentTypes, Object[] arguments)
-            throws NoSuchMethodException, InvocationTargetException, IllegalAccessException, InstantiationException {
-
-        Objects.requireNonNull(targetClass);
-        Objects.requireNonNull(argumentTypes);
-        Objects.requireNonNull(arguments);
-        if (argumentTypes.length != arguments.length) {
-            throw new IllegalArgumentException("Mismatch number of argument types and concrete arguments. Argument types "
-                    + argumentTypes.length + " while concrete arguments were "
-                    + arguments.length);
-        }
-
-        Constructor<?> documentCodecConstructor = targetClass.getConstructor(argumentTypes);
-        boolean documentCodecConstructorIsAccessible = documentCodecConstructor.isAccessible();
-        documentCodecConstructor.setAccessible(true);
-        Object newInstance = documentCodecConstructor.newInstance(arguments);
-        documentCodecConstructor.setAccessible(documentCodecConstructorIsAccessible);
-        return newInstance;
-    }
-
 
     private static String getDatabaseName(Object mongoCollection) throws ClassNotFoundException, NoSuchMethodException, IllegalAccessException, InvocationTargetException {
-        final Class<?> mongoNamespaceClass = Class.forName("com.mongodb.MongoNamespace");
[…]
