+++ client-java/controller-api/src/main/java/org/evomaster/client/java/controller/api/dto/mongo/DocumentDto.java
@@ -1,0 +1,8 @@
+package org.evomaster.client.java.controller.api.dto.mongo;
+
+public class DocumentDto {
+
+    public String documentAsJsonString;
+
+
+}
