+++ client-java/controller-api/src/main/java/org/evomaster/client/java/controller/api/dto/mongo/FindOperationDto.java
@@ -6,5 +6,5 @@ public class FindOperationDto {
 
     public String collectionName;
 
-    public String queryJsonStr;
+    public DocumentDto queryDocumentDto;
 }

[…]
