### 2020-05-13 12bacac
testfix: missing --extractMongoExecutionInfo option
