### 2020-05-14 d51e47d
bugfix: redundant debug code is removed.
