### 2020-05-18 EMResearch/EvoMaster#170 @s0nata
@arcuri82 looks great, thanks! One last minor point: "Statement of need" section, paragraph 2 sentence 2: "To address this problem, in SE ... in SE" - the second "in SE" reference in the same sentence seems redundant. Could you please take it out?
