### 2020-05-18 EMResearch/EvoMaster#178 @mado89
baseUrl must not be null
Ciao Andrea,

When trying to run a whitebox test using the instructions the program ends up in an illegal state

`java.lang.IllegalStateException: baseUrl must not be null
	at org.evomaster.core.problem.rest.service.AbstractRestFitness.getBaseUrl(AbstractRestFitness.kt:277)
	at org.evomaster.core.problem.rest.service.AbstractRestFitness.getCookies(AbstractRestFitness.kt:566)
	at org.evomaster.core.problem.rest.service.RestFitness.doCalculateCoverage(RestFitness.kt:38)
	at org.evomaster.core.problem.rest.service.RestFitness.doCalculateCoverage(RestFitness.kt:22)
	at org.evomaster.core.search.service.FitnessFunction.calculateCoverage(FitnessFunction.kt:46)
	at org.evomaster.core.search.algorithms.MioAlgorithm.search(MioAlgorithm.kt:41)
	at org.evomaster.core.Main$Companion.run(Main.kt:280)
	at org.evomaster.core.Main$Companion.initAndRun(Main.kt:148)
	at org.evomaster.core.Main$Companion.main(Main.kt:72)
	at org.evomaster.core.Main.main(Main.kt)`

I assume that there is something wrong with my configuration. But how can I find out my mistake?

Thanks
