### 2020-05-18 EMResearch/EvoMaster#177 @mado89
Thanks. The instructions are much clearer now.
I guess what is missing now is some information on how to provide the authentication. I am currently testing your framework against an application of mine. Most end points are secured using oauth. Hence, it would be nice to have this information.
