### 2020-05-21 EMResearch/EvoMaster#170 @arcuri82
@s0nata: thanks for spotting that issue. it is fixed now
