### 2020-05-21 EMResearch/EvoMaster#178 @arcuri82
even if it is a misconfiguration on your side, we should give a better error message :)  i ll fix it

are you doing Black-box or White-box testing? (the configurations to fix it would be different...)
