### 2020-05-21 EMResearch/EvoMaster#178 @arcuri82
i added some better error message.
however, it is weird, as it should not be possible to have "null" there.
how exactly did you run EvoMaster? eg, if you are doing White-Box testing, how is the implementation of your startSut()?
