### 2020-05-21 EMResearch/EvoMaster#177 @arcuri82
we have basic support for some kinds of authentication. OAuth is unfortunately not among them :(  we will support it in the future, but it will take several months before we have time to add it (it is not in the current priorities)

anyway, I will add some documentation on how security is currently handled (eg, clarify that OAuth is not supported at the moment)

btw, is you are application open-source? if yes, it would be an interesting case study for when we will support OAuth
