+++ docs/write_driver.md
@@ -12,6 +12,7 @@ _EvoMaster_ Java client library. For example, in Maven:
 <dependency>
    <groupId>org.evomaster</groupId>
    <artifactId>evomaster-client-java-controller</artifactId>
+   <scope>test</scope>
    <version>LATEST</version>
 </dependency>
 ```
@@ -22,7 +23,7 @@ If you are compiling directly from the _EvoMaster_ source code, make sure to use
 install the snapshot version of the Java client into your local Maven repository 
 (e.g., under *~/.m2*). 
 
-Note: the core application `evomaster.jar` is independent from the driver library, and it contains none of 
+Note: the core application `evomaster.jar` is independent of the driver library, and it contains none of 
 the driver's classes.
 
 
@@ -38,7 +39,7 @@ The easiest approach (which we recommend) is to use the *embedded* version, espe
 frameworks like Spring and DropWizard. 
 However, when the presence of the _EvoMaster_ client library gives side-effects (although 
 its third-party libraries are shaded, side-effects might still happen),
-or when it is not possible (or too complicate) to start the SUT directly (e.g., JEE),
+or when it is not possible (or too complicated) to start the SUT directly (e.g., JEE),
 it is better to use the *external* version.
 The requirement is that there should be a single, self-executable uber/fat jar for the SUT 
 (e.g., Wildfly Swarm).
@@ -103,7 +104,7 @@ When writing an _EvoMaster_ driver, there are 2 TCP ports that you need to consi
   for this (e.g., by using the value 0, and then read back in the driver which port was actually 
   assigned to the server).   
 
-## SpringBoot Example
+## Starting The Application
 
 How to start/reset/stop the SUT depends on the chosen framework used to implement the SUT. 
 To implement an _EvoMaster Driver_ class, you need check the JavaDocs of the extended super class,
@@ -127,7 +128,7 @@ When starting the SUT, there are at least two configurations that you want to ch
 * the binding port, as you want to use 0 for ephemeral ports (to avoid port conflicts).
 * if the SUT is using a SQL database, you MUST wrap the SQL driver with _P6Spy_. 
   This is as simple as adding `:p6spy` in the connecting datasource URL and change the `driver-class-name`.
-  This is needed by _EvoMaster_ to be able intercept and analyse all the interactions with the database.
+  This is needed by _EvoMaster_ to be able to intercept and analyse all the interactions with the database.
   
 For a SUT like `features-service`, this can be done with:
 
@@ -154,6 +155,9 @@ protected int getSutPort() {
 
 Finally, the `startSut()` method must return the URL of where the SUT is listening on.
 When running tests locally, this is as simple as returning `"http://localhost:" + getSutPort()`.
+
+
+## SQL Databases
 
 If the application is using a SQL database, you must configure `getConnection()` and `getDatabaseDriverName()`,
 instead of leaving their returned values as `null`.
@@ -189,6 +193,8 @@ If your application uses some caches, those might be reset at each test executio
 However, an easier approach could be to just start the SUT without the caches, for example using
 the option `--spring.cache.type=NONE`.
 
+
+## Code Coverage  
  
 When _EvoMaster_ evolves test cases, it tries to maximize code coverage in the SUT.
 But there is no much point in trying to maximize code coverage of the third-party libraries,
@@ -198,22 +204,48 @@ business logic.
 In the case of the `features-service` SUT, this was `org.javiermf.features`.
 
 
+## REST OpenApi/Swagger Schema
+
 To test a RESTful API, in the the `getProblemInfo()`, you need to return an instance of the
 `RestProblem` class.
[…]
