### 2020-05-21 EMResearch/EvoMaster#177 @arcuri82
I now added more info on security at [https://github.com/EMResearch/EvoMaster/blob/master/docs/write_driver.md]()
