### 2020-05-21 a8f8fa3
bugfix: wrong getter name was making Jackson miss a boolean field
