+++ client-java/controller/src/main/java/org/evomaster/client/java/controller/internal/db/mongo/MongoHandler.java
@@ -77,7 +77,7 @@ public class MongoHandler {
 
         FindResultDto findResultDto = new FindResultDto();
         findResultDto.findResultType = FindResultDto.FindResultType.SUMMARY;
-        findResultDto.hasReturnedAnyDocument = loggedExecutedFindOperation.hasReturnedAnyDocument();
+        findResultDto.hasReturnedAnyDocument = loggedExecutedFindOperation.getHasReturnedAnyDocument();
 
         ExecutedFindOperationDto dto = new ExecutedFindOperationDto();
         dto.findOperationDto = findOperationDto;
