### 2020-05-26 4a57b1e
splitting run of E2E on Travis
