+++ e2e-tests/e2e-tests-utils/src/test/java/org/evomaster/e2etests/utils/RestTestBase.java
@@ -436,7 +436,8 @@ public abstract class RestTestBase {
         boolean ok = solution.getIndividuals().stream().anyMatch(
                 ind -> hasAtLeastOne(ind, verb, expectedStatusCode, path, inResponse));
 
-        String errorMsg = "Missing " + expectedStatusCode + " " + verb + " " + path + " " + inResponse + "\n";
+        String errorMsg = "Seed " + (defaultSeed-1)+". ";
+        errorMsg += "Missing " + expectedStatusCode + " " + verb + " " + path + " " + inResponse + "\n";
 
         assertTrue(ok, errorMsg + restActions(solution));
     }
