### 2020-05-26 807eeff
restoring DbAut*Test to investigate issue on CircleCI
