### 2020-05-26 c7bb218
trying JDK 11 in CircleCI to see if the JDK was the issue with flakiness
