+++ e2e-tests/spring-examples/src/test/java/org/evomaster/e2etests/spring/examples/taintMulti/TaintMultiEMTest.java
@@ -24,6 +24,8 @@ public class TaintMultiEMTest extends SpringTestBase {
     @Test
     public void testDeterminism(){
 
+        defaultSeed = 15;
+
         runAndCheckDeterminism(5_000, (args) -> {
             initAndRun(args);
         });
