### 2020-05-26 8e55186
trying to get Travis to work with JDK switcher
