### 2020-05-27 a5161ce
fix for NPE
