### 2020-05-28 de6f3f0
SutController does not have an explicit dependency to a given MongoDB version
