+++ e2e-tests/spring-rest-mongo/src/test/java/com/foo/mongo/MyMongoAppEmbeddedController.java
@@ -12,4 +12,6 @@ public class MyMongoAppEmbeddedController extends SpringRestMongoController {
     public String getPackagePrefixesToCover() {
         return "com.foo.mongo";
     }
+
+
 }
