+++ client-java/controller/src/test/java/org/evomaster/client/java/controller/internal/db/mongo/MongoFakeSutController.java
@@ -73,8 +73,5 @@ public class MongoFakeSutController extends EmbeddedSutController {
         return this.mongoHandler;
     }
 
-    @Override
-    public MongoClient getMongoClient() {
-        return this.mongoClient;
-    }
+
 }
