### 2020-05-28 88c744c
testfix: missing implementation of executeMongoFindOperation() in MongoFakeSutController
