+++ client-java/controller/src/test/java/org/evomaster/client/java/controller/internal/db/mongo/MongoFakeSutController.java
@@ -1,15 +1,23 @@
 package org.evomaster.client.java.controller.internal.db.mongo;
 
 import com.mongodb.MongoClient;
+import com.mongodb.client.FindIterable;
+import com.mongodb.client.MongoCollection;
+import com.mongodb.client.MongoDatabase;
+import org.bson.Document;
 import org.evomaster.client.java.controller.EmbeddedSutController;
 import org.evomaster.client.java.controller.api.dto.AuthenticationDto;
 import org.evomaster.client.java.controller.api.dto.SutInfoDto;
+import org.evomaster.client.java.controller.api.dto.mongo.FindOperationDto;
+import org.evomaster.client.java.controller.mongo.DetailedFindResult;
+import org.evomaster.client.java.controller.mongo.FindOperation;
 import org.evomaster.client.java.controller.problem.ProblemInfo;
 import org.evomaster.client.java.controller.problem.RestProblem;
 
 import java.sql.Connection;
 import java.util.List;
 import java.util.Objects;
+import java.util.stream.StreamSupport;
 
 public class MongoFakeSutController extends EmbeddedSutController {
     private final MongoClient mongoClient;
@@ -74,4 +82,13 @@ public class MongoFakeSutController extends EmbeddedSutController {
     }
 
 
+    public DetailedFindResult executeMongoFindOperation(FindOperationDto dto) {
+        FindOperation findOperation = FindOperation.fromDto(dto);
+        MongoDatabase database = this.mongoClient.getDatabase(findOperation.getDatabaseName());
+        MongoCollection collection = database.getCollection(findOperation.getCollectionName());
+        FindIterable<Document> findIterable = collection.find(findOperation.getQuery());
+        DetailedFindResult findResult = new DetailedFindResult();
+        StreamSupport.stream(findIterable.spliterator(), false).forEach(findResult::addDocument);
+        return findResult;
+    }
 }
