### 2020-05-29 1a79853
docs on how to setup JDK environment
