+++ client-java/controller/src/main/java/org/evomaster/client/java/controller/internal/db/HeuristicsCalculator.java
@@ -9,6 +9,9 @@ import net.sf.jsqlparser.statement.Statement;
 import org.evomaster.client.java.controller.db.DataRow;
 import org.evomaster.client.java.controller.db.QueryResult;
 import org.evomaster.client.java.instrumentation.coverage.methodreplacement.DistanceHelper;
+import org.evomaster.client.java.instrumentation.shared.StringSpecialization;
+import org.evomaster.client.java.instrumentation.shared.StringSpecializationInfo;
+import org.evomaster.client.java.instrumentation.staticstate.ExecutionTracer;
 import org.evomaster.client.java.utils.SimpleLogger;
 
 import java.sql.Timestamp;
@@ -396,7 +399,10 @@ public class HeuristicsCalculator {
     private double computeComparison(String a, String b, ComparisonOperator exp) {
 
         if (exp instanceof EqualsTo) {
+
+            ExecutionTracer.handleTaintForStringEquals(a, b, false);
             return DistanceHelper.getLeftAlignmentDistance(a, b);
+
         } else if (exp instanceof NotEqualsTo) {
             if (a.equals(b)) {
                 return Double.MAX_VALUE;
