### 2020-06-02 137c7e0
fix updated parameter of EMConfig
