+++ e2e-tests/spring-examples/src/test/java/org/evomaster/e2etests/spring/examples/impact/ImpactEMTest.java
@@ -64,7 +64,7 @@ public class ImpactEMTest extends SpringTestBase {
                     args.add("--probOfArchiveMutation");
                     args.add("0.75");
 
-                    args.add("--geneSelectionMethod");
+                    args.add("--adaptiveGeneSelectionMethod");
                     args.add(method.toString());
 
                     args.add("--enableTrackEvaluatedIndividual");
