+++ e2e-tests/spring-examples/src/test/java/org/evomaster/e2etests/spring/examples/impact/ObjectImpactEMTest.java
@@ -68,7 +68,7 @@ public class ObjectImpactEMTest extends SpringTestBase {
                     args.add("--probOfArchiveMutation");
                     args.add("1.0");
 
-                    args.add("--geneSelectionMethod");
+                    args.add("--adaptiveGeneSelectionMethod");
                     args.add(method.toString());
 
                     args.add("--enableTrackEvaluatedIndividual");
