### 2020-06-03 5b9b88b
merge with weight-based solution
