### 2020-06-04 6bdf1b1
check whether collect enough for performing adaptive mutation fix a bug for handling impacts of genes which are not mutated
