### 2020-06-05 c535e50
adding info for databases in Docker
