### 2020-06-05 f1dd804
improved default exp script
