### 2020-06-08 fd51c61
disable ObjectImpactEMTest (enable when archive-based solution is merged)
