### 2020-06-08 f71cd43
Revert "disable ObjectImpactEMTest (enable when archive-based solution is merged)"
This reverts commit fd51c61c
