+++ e2e-tests/spring-examples/src/test/java/org/evomaster/e2etests/spring/examples/impactXYZ/CollectImpactXYZInfoTest.java
@@ -1,0 +1,133 @@
+package org.evomaster.e2etests.spring.examples.impactXYZ;
+
+import com.foo.rest.examples.spring.impactXYZ.ImpactXYZRestController;
+import org.evomaster.core.problem.rest.RestIndividual;
+import org.evomaster.core.problem.rest.util.ParamUtil;
+import org.evomaster.core.search.EvaluatedIndividual;
+import org.evomaster.core.search.Individual;
+import org.evomaster.core.search.Solution;
+import org.evomaster.core.search.gene.Gene;
+import org.evomaster.core.search.impact.GeneImpact;
+import org.evomaster.core.search.impact.GeneMutationSelectionMethod;
+import org.evomaster.core.search.impact.ImpactUtils;
+import org.evomaster.e2etests.spring.examples.SpringTestBase;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
+
+import java.util.*;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+
+public class CollectImpactXYZInfoTest extends SpringTestBase {
+
+    @Test
+    public void testOnlyCollectImpact() throws Throwable {
+        testRunEM(GeneMutationSelectionMethod.NONE);
+    }
+
+    public void testRunEM(GeneMutationSelectionMethod method) throws Throwable {
+
+        runTestHandlingFlakyAndCompilation(
+                "none",
+                "none",
+                500,
+                false,
+                (args) -> {
+
+                    args.add("--probOfArchiveMutation");
+                    args.add("1.0");
+
+                    args.add("--adaptiveGeneSelectionMethod");
+                    args.add(method.toString());
+
+                    //since there only exist one endpoint, we set the population for each target 3
+                    args.add("--archiveTargetLimit");
+                    args.add("3");
+
+                    args.add("--enableTrackEvaluatedIndividual");
+                    args.add("true");
+
+                    args.add("--focusedSearchActivationTime");
+                    args.add("0.0");
+
+                    // only for the test
+                    args.add("--saveImpactAfterMutationFile");
+                    args.add("target/impactXYZ/impactInfo/Impacts_ImpactXYZ_"+method.toString()+".csv");
+
+                    // only for the test
+                    args.add("--saveMutatedGeneFile");
+                    args.add("target/impactXYZ/mutatedGeneInfo/MutatedGenes_ImpactXYZ_"+method.toString()+".csv");
+
+                    // only for the test
+                    args.add("--saveArchiveAfterMutationFile");
+                    args.add("target/impactXYZ/archiveInfo/ArchiveNotCoveredSnapshot_ImpactXYZ_"+method.toString()+".csv");
+
+                    args.add("--exportCoveredTarget");
+                    args.add("true");
+
+                    args.add("--coveredTargetFile");
+                    String path = "target/impactXYZ/coveredTargets/testImpacts_CoveredTargetsBy"+method.toString()+".csv";
+                    args.add(path);
+
+
+                    Solution<RestIndividual> solution = initAndRun(args);
+
+                    assertTrue(solution.getIndividuals().size() >= 1);
+
+                    solution.getIndividuals().stream().allMatch(
+                            s -> s.anyImpactInfo() && checkImpactOfxyz(s)
+                    );
+
+                }, 3);
+    }
+
+    private String getGeneIdByName(String geneName, EvaluatedIndividual<RestIndividual> ind){
+
[…]
