### 2020-06-10 8e87da2
updating example exp.py script to handle multi-JDK in local experiments
