### 2020-06-11 8142e73
bit of docs restructuring, plus info on Windows TPC issues
