### 2020-06-16 4ec9ac5
archive gene mutation for integer gene and string gene with unit tests and e2e tests
