+++ e2e-tests/spring-examples/src/test/java/org/evomaster/e2etests/spring/examples/impact/ImpactEMTest.java
@@ -7,9 +7,9 @@ import org.evomaster.core.search.EvaluatedIndividual;
 import org.evomaster.core.search.Individual;
 import org.evomaster.core.search.Solution;
 import org.evomaster.core.search.gene.Gene;
-import org.evomaster.core.search.impact.GeneImpact;
-import org.evomaster.core.search.impact.GeneMutationSelectionMethod;
-import org.evomaster.core.search.impact.ImpactUtils;
+import org.evomaster.core.search.impact.impactInfoCollection.GeneImpact;
+import org.evomaster.core.search.impact.impactInfoCollection.GeneMutationSelectionMethod;
+import org.evomaster.core.search.impact.impactInfoCollection.ImpactUtils;
 import org.evomaster.e2etests.spring.examples.SpringTestBase;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Disabled;
