### 2020-06-16 de6e6a9
remove 'enablePrioritizeTargetsByImpact' for the moment
