+++ e2e-tests/spring-examples/src/test/java/org/evomaster/e2etests/spring/examples/db/directintwithsql/DbDirectIntWithSqlEMTest.java
@@ -103,7 +103,7 @@ public class DbDirectIntWithSqlEMTest extends DbDirectIntWithSqlTestBase {
         FitnessFunction<RestIndividual> ff = injector.getInstance(Key.get(
                 new TypeLiteral<FitnessFunction<RestIndividual>>() {
                 }));
-        EvaluatedIndividual ei = ff.calculateCoverage(ind, new HashSet<>());
+        EvaluatedIndividual ei = ff.calculateCoverage(ind);
         assertNotNull(ei);
 
         FitnessValue noDataFV = ei.getFitness();
@@ -146,7 +146,7 @@ public class DbDirectIntWithSqlEMTest extends DbDirectIntWithSqlTestBase {
 
         RestIndividual withSQL = new RestIndividual(ind.seeActions(), ind.getSampleType(), insertions, null, null);
 
-        ei = ff.calculateCoverage(withSQL, new HashSet<>());
+        ei = ff.calculateCoverage(withSQL);
         assertNotNull(ei);
 
         //should have better heuristic
@@ -177,7 +177,7 @@ public class DbDirectIntWithSqlEMTest extends DbDirectIntWithSqlTestBase {
                     }
                 });
 
-        ei = ff.calculateCoverage(withSQL, new HashSet<>());
+        ei = ff.calculateCoverage(withSQL);
         assertNotNull(ei);
 
         //As SQL data is returned, we get no heuristic, and so worst value
