### 2020-06-16 8b0c970
modify mutation weight tests that shows different performance in gene selection fix bugs that fails tests
