### 2020-06-18 4dd35a5
fix bug in handling of SQL timestamps
