### 2020-06-18 677b126
WiP. Updates as a result of feedback.
Fixed a bug in the OpenAPI handling (paths were not valid due to an added "/api" prefix.
