+++ e2e-tests/spring-examples/src/test/java/org/evomaster/e2etests/spring/examples/splitter/TestSuiteSplitterTest.java
@@ -65,6 +65,8 @@ public class TestSuiteSplitterTest extends SplitterTestBase {
                 (args) -> {
                     args.add("--testSuiteSplitType");
                     args.add("" + splitType);
+                    args.add("--expectationsActive");
+                    args.add("" + true);
                     Solution<RestIndividual> solution = initAndRun(args);
                     assertTrue(solution.getIndividuals().size() >= 1);
                     SplitResult splits = TestSuiteSplitter.INSTANCE.split(solution, em);
@@ -84,6 +86,8 @@ public class TestSuiteSplitterTest extends SplitterTestBase {
                 (args) -> {
                     args.add("--testSuiteSplitType");
                     args.add("" + splitType);
+                    args.add("--expectationsActive");
+                    args.add("" + true);
 
                     Solution<RestIndividual> solution = initAndRun(args);
 
