### 2020-06-21 c03e357
fix tests lead by refactored resource sampler
