### 2020-06-21 d1e2576
fix e2e test with refactored evaluated individual
