### 2020-06-23 0cba91e
option to focus search on a single endpoint
