### 2020-06-25 436d307
improving Travis config
