+++ .travis.yml
@@ -22,6 +22,14 @@ jdk:
   - openjdk8
 #  - openjdk11
 
+# By default, Travis does a "mvn install" during its onw 'install' phase.
+# This is annoying, and unnecessarily time consuming when using multiple run-scripts.
+# So, we override this behavior.
+# We just force the downloading of dependencies here to reduce the clutter in the logs.
+install: mvn dependency:resolve
+
+
+# As Travis has a 50 minute limit per script, we need to split the build in different stages.
 jobs:
   include:
     - stage: init
