### 2020-06-25 7c8c2ce
refactor Mutator with updated mutated genes specification
