+++ e2e-tests/spring-examples/src/test/java/org/evomaster/e2etests/spring/examples/testability/TestabilityEMTest.java
@@ -37,6 +37,14 @@ public class TestabilityEMTest extends SpringTestBase {
 
                     args.add("--baseTaintAnalysisProbability");
                     args.add("0.9");
+//
+//                    args.add("--mutationTargetsSelectionStrategy");
+//                    args.add("UPDATED_NOT_COVERED_TARGET");
+
+
+                    args.add("--saveMutatedGeneFile");
+                    args.add("target/testability/targetsFirst.csv");
+
 
                     Solution<RestIndividual> solution = initAndRun(args);
 
