### 2020-06-25 e70a261
fix restore value before it is mutated
