### 2020-06-26 5a5069e
two genes are bounded with taint analysis
