### 2020-06-26 5681e1c
testability with 3 targets strategy
