+++ e2e-tests/spring-examples/src/test/java/org/evomaster/e2etests/spring/examples/testability/TestabilityEMTest.java
@@ -38,6 +38,9 @@ public class TestabilityEMTest extends SpringTestBase {
                     args.add("--baseTaintAnalysisProbability");
                     args.add("0.9");
 
+                    args.add("--saveMutatedGeneFile");
+                    args.add("target/testability/targetsFirst.csv");
+
                     Solution<RestIndividual> solution = initAndRun(args);
 
                     assertTrue(solution.getIndividuals().size() >= 1);
@@ -49,7 +52,7 @@ public class TestabilityEMTest extends SpringTestBase {
                 10);
     }
 
-    @Disabled("check testability binding")
+    //@Disabled("check testability binding")
     @Test
     public void testRunEMWithUpdatedTargets() throws Throwable {
 
@@ -70,6 +73,38 @@ public class TestabilityEMTest extends SpringTestBase {
 
                     args.add("--saveMutatedGeneFile");
                     args.add("target/testability/targetsUpdate.csv");
+
+                    Solution<RestIndividual> solution = initAndRun(args);
+
+                    assertTrue(solution.getIndividuals().size() >= 1);
+
+                    assertHasAtLeastOne(solution, HttpVerb.GET, 500, "/api/testability/{date}/{number}/{setting}", null);
+                    assertHasAtLeastOne(solution, HttpVerb.GET, 200, "/api/testability/{date}/{number}/{setting}", "ERROR");
+                    assertHasAtLeastOne(solution, HttpVerb.GET, 200, "/api/testability/{date}/{number}/{setting}", "OK");
+                },
+                10);
+    }
+
+    @Test
+    public void testRunEMWithExpandedUpdatedTargets() throws Throwable {
+
+        CIUtils.skipIfOnCircleCI();
+
+        runTestHandlingFlakyAndCompilation(
+                "TestabilityEM",
+                "org.bar.TestabilityEM",
+                15_000,
+                true,
+                (args) -> {
+
+                    args.add("--baseTaintAnalysisProbability");
+                    args.add("0.9");
+
+                    args.add("--mutationTargetsSelectionStrategy");
+                    args.add("EXPANDED_UPDATED_NOT_COVERED_TARGET");
+
+                    args.add("--saveMutatedGeneFile");
+                    args.add("target/testability/targetsExpand.csv");
 
                     Solution<RestIndividual> solution = initAndRun(args);
 
