+++ client-java/instrumentation/src/main/java/org/evomaster/client/java/instrumentation/staticstate/ExecutionTracer.java
@@ -104,6 +104,15 @@ public class ExecutionTracer {
                 return;
             }
 
+            /*
+                We consider binding only for base versions of taint, ie we ignore
+                the special strings provided by the Core, as it would lead to nasty
+                side-effects
+             */
+            if(! TaintInputName.isTaintInput(left) || !TaintInputName.isTaintInput(right)){
+                return;
+            }
+
             //TODO could have EQUAL_IGNORE_CASE
             String id = left + "___" + right;
             addStringSpecialization(left, new StringSpecializationInfo(StringSpecialization.EQUAL, id));
