### 2020-06-26 3e0eec2
set seed for testabilityEMTest regarding updated targets
