### 2020-06-30 2825e2f
fix tests with different fullClassName
