### 2020-06-30 ac6e4f8
investigating process hanging on Windows
