### 2020-07-01 8836d65
refactor archive gene mutator from archive gene selector
Merge branch 'fitness-check' into archive-mutation

# Conflicts:
#	core/src/main/kotlin/org/evomaster/core/problem/rest/service/RestFitness.kt
#	core/src/main/kotlin/org/evomaster/core/problem/rest/service/RestResourceFitness.kt
#	core/src/main/kotlin/org/evomaster/core/problem/rest/service/RestStructureMutator.kt
#	core/src/main/kotlin/org/evomaster/core/search/EvaluatedIndividual.kt
#	core/src/main/kotlin/org/evomaster/core/search/Individual.kt
#	core/src/main/kotlin/org/evomaster/core/search/impact/impactInfoCollection/ImpactUtils.kt

[…]
