+++ e2e-tests/spring-examples/src/test/java/org/evomaster/e2etests/spring/examples/testability/MutatorTestWithTestabilityRest.java
@@ -44,7 +44,7 @@ public class MutatorTestWithTestabilityRest extends SpringTestBase {
         runTestHandlingFlakyAndCompilation(
                 "EM",
                 "EM",
-                15_000,
+                1_100,
                 false,
                 (args) -> {
 
@@ -55,6 +55,8 @@ public class MutatorTestWithTestabilityRest extends SpringTestBase {
                     args.add("1");
 
                     Injector injector = init(args);
+                    initPartialOracles(injector);
+
                     StandardMutator<RestIndividual> mutator = injector.getInstance(Key.get(
                            new TypeLiteral<StandardMutator<RestIndividual>> () {}));
                     assertNotNull(mutator);
