### 2020-07-01 3073dd0
Adding a means of modifying epsilon values for the Distance Metrics from EMConfig.
Bringing it in line with the latest master. Since some fiddling was needed with the EMConfig, the updated options.md is included in this commit.
