### 2020-07-01 c996ef0
modify the saveMutatedInfo
