### 2020-07-03 46359d2
modify gene impacts, and archive-gene mutator for stringGene
