+++ .travis.yml
@@ -38,7 +38,6 @@ install: mvn de.qaware.maven:go-offline-maven-plugin:1.2.5:resolve-dependencies
 env:
   global:
    -  CI_env: Travis
-   -  POSTGRES_HOST_AUTH_METHOD: trust
 
 
 # As Travis has a 50 minute limit per script, we need to split the build in different stages.
