### 2020-07-05 01a7a40
increasing timeout, and trying again on Travis
