+++ e2e-tests/spring-examples/src/main/java/com/foo/rest/examples/spring/ttpaper/TTPaperSql.java
@@ -1,0 +1,7 @@
+package com.foo.rest.examples.spring.ttpaper;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class TTPaperSql {
+}
