### 2020-07-06 bd6f198
Changes from PR feedback. Adding a means of modifying epsilon values for the Distance Metrics from EMConfig.
