+++ e2e-tests/spring-examples/src/test/java/org/evomaster/e2etests/spring/examples/impact/ObjectImpactEMTest.java
@@ -8,11 +8,11 @@ import org.evomaster.core.search.Individual;
 import org.evomaster.core.search.Solution;
 import org.evomaster.core.search.gene.Gene;
 import org.evomaster.core.search.gene.ObjectGene;
-import org.evomaster.core.search.impact.impactInfoCollection.GeneImpact;
-import org.evomaster.core.search.impact.impactInfoCollection.Impact;
-import org.evomaster.core.search.impact.impactInfoCollection.GeneMutationSelectionMethod;
-import org.evomaster.core.search.impact.impactInfoCollection.ImpactUtils;
-import org.evomaster.core.search.impact.impactInfoCollection.value.ObjectGeneImpact;
+import org.evomaster.core.search.impact.impactinfocollection.GeneImpact;
+import org.evomaster.core.search.impact.impactinfocollection.Impact;
+import org.evomaster.core.search.impact.impactinfocollection.GeneMutationSelectionMethod;
+import org.evomaster.core.search.impact.impactinfocollection.ImpactUtils;
+import org.evomaster.core.search.impact.impactinfocollection.value.ObjectGeneImpact;
 import org.evomaster.e2etests.spring.examples.SpringTestBase;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Disabled;
