### 2020-07-07 9c1c90c
update impact reward with double type by considering hyper mutation
