### 2020-07-07 6599301
fix bugs by modifying impact reward with double
