+++ e2e-tests/spring-examples/src/main/java/com/foo/rest/examples/spring/ttpaper/TTPaperSql.java
@@ -1,7 +1,18 @@
 package com.foo.rest.examples.spring.ttpaper;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RestController;
+
+import javax.persistence.EntityManager;
 
 @RestController
 public class TTPaperSql {
+
+    @Autowired
+    private EntityManager em;
+
+
+    public void get(){
+        //TODO
+    }
 }
