### 2020-07-08 #2153 @UTH-Tuan
> @UTH-Tuan and @shanamatthews, would either or both of you be willing to contribute a review for this JOSS submission?

@gkthiruvathukal I guess I can step in light of the situation. I'll admit that this is a new way of doing reviews for manuscripts so I may need some direction and guidance.
