### 2020-07-10 835bbbc
fix bugs count inner impacts for regex genes
