### 2020-07-12 #2153 @gkthiruvathukal
@UTH-Tuan I will be able to answer any questions about the review process.  Just so you know, we have reviewer guidelines that take you through the process at https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. I'm going to add you as a reviewer now!
