### 2020-07-12 #2153 @mado89
> @arcuri82 So sorry about this. I am still hopeful we will be able to see this through.
>
> @mado89, @s0nata, and @feliperodri, can you provide any update on when you will be able to complete your review? For some reason I thought we had reviews from at least one of you (@mado89) but we have none. I really don't want to keep @arcuri82 waiting much longer.

I am sorry for not posting my review more explicitily:
In my opinion this tool needs some more work. Although investing quite some time into this review. I am still unable to perform a thorough review of this tool for the following reasons:

* Lack of gradle support in documentation (I am aware that most people use Maven. Maybe my issues are related to some other aspects)
* Lack of user documentation: For "white box" testing there is a document explaining a lot of technical details. However, in my opinion it lacks a "Step-by-step" like guide helping novice users of the tool performing their first test. Hence, I was unable to perform the test (also I just received a cryptic failure message, which I reported back to the author. There the review stopped).
* Lack of OAuth support: I had a brief discussion with the author about this. However, we are in a disagreement about the importance of OAuth. If mentioned somewhere more prominent that OAuth is not supported than this could be ok for me. I discovered this draw back only after investing quite some time into the review and trying to set up a test for an application.

Unfortunately, this comments are somewhat hard to add to the standard review form.
When these issues are somewhat addressed I am willing to again spend some time on this review (I spent several days trying to get the tool running).
