### 2020-07-12 #2153 @arcuri82
I am here going to reply to the 3 points raised by @mado89 in the previous post

1. In the 6 months that @mado89 has been a reviewer, 4 issues ([https://github.com/EMResearch/EvoMaster/issues](https://github.com/EMResearch/EvoMaster/issues)) were open: 169, 170, 177 and 178. Unless I missed it, nothing about Gradle was ever mentioned.  Adding an example of Gradle import of a library is pretty trivial, and I ll do it after writing this response.

2. I guess this is referring to  [177](https://github.com/EMResearch/EvoMaster/issues/177), where @mado89 wrote `The instructions are much clearer now`. If there are parts that are still not clear, this does not seem reported in [177](https://github.com/EMResearch/EvoMaster/issues/177) (besides OAuth, discussed next).
Furthermore, regarding `(also I just received a cryptic failure message, which I reported back to the author. There the review stopped)`: I assume @mado89 is referring to [178](https://github.com/EMResearch/EvoMaster/issues/178). If that is the case, I am still waiting for my questions of 21st of May. It is impossible to debug such issue if no more info is provided.

 3. On this point, I am in disagreement with @mado89. As such communication was done via email, I will post here my answer:

> On this, I do disagree:
> - of all open-source REST API that I found (eg, then ones in https://github.com/EMResearch/EMB), none use OAuth
> - none of the industrial partners I collaborate with use OAuth (and they have SpringBoot REST APIs)
> - on API directories (eg https://www.programmableweb.com/  and https://www.programmableweb.com/) there are literally hundreds (if not thousands) of APIs not using OAuth. If you want to look at some numbers, check > https://www.programmableweb.com/news/spotting-api-security-trends-programmablewebs-api-directory/research/2018/01/02
> - no other test generation tool support OAuth
>
> We are going to support OAuth in the (hopefully not far) future, but it has not be a priority so far, as it is not so popular yet. There are other more important features to add first
