### 2020-07-12 #2153 @arcuri82
Gradle example has now been added to the [documentation](https://github.com/EMResearch/EvoMaster/blob/master/docs/write_driver.md).
