### 2020-07-12 #2153 @arcuri82
To clarify regarding OAuth: `If mentioned somewhere more prominent that OAuth is not supported than this could be ok for me`. As I wrote in [177](https://github.com/EMResearch/EvoMaster/issues/177), the fact that OAuth is not supported has been added to the documentation, in the section regarding Security. As I did not get any reply in 177, I thought it was OK. But I can add it now to the main readme.
