+++ README.md
@@ -55,7 +55,8 @@ __Key features__:
   as long as an OpenAPI schema is provided. However, results will be worse than whitebox testing (e.g., due
   to lack of bytecode analysis).
 
-
+* _Authentication_: we support auth based on authentication headers and cookies. 
+   However, at the moment we do not support OAuth.
 
 __Known limitations__:
 
