### 2020-07-18 #2153 @UTH-Tuan
@arcuri82 I have taken a look at the paper to understand the scope of the problem that the software addresses. I don't see any discussion relating to what others have done and how this tool differentiates itself.
