### 2020-07-19 #2153 @arcuri82
@UTH-Tuan, from the [paper](https://github.com/openjournals/joss-papers/blob/joss.02153/joss.02153/10.21105.joss.02153.pdf):

> Related Work
>In the recent years, different techniques have been proposed for black-box testing of RESTful APIs (e.g., (Atlidakis, Godefroid, & Polishchuk, 2019)(Karlsson, Causevic, & Sundmark, 2020) (Viglianisi, Dallago, & Ceccato, 2020)(Ed-douibi, Cánovas Izquierdo, & Cabot, 2018)).
>However, at the time of this writing, EvoMaster appears to be the only one can do both black-box and white-box testing, and that is also released as open-source

Are you saying that you want to see that (very brief) section extended?
