### 2020-07-21 #2153 @arcuri82
@UTH-Tuan thanks. I have added some more info in the Related Work section. Right now the *paper.md* file has 974 words, where JOSS has the limit of [250-1000 words](https://joss.readthedocs.io/en/latest/submitting.html).
