### 2020-07-21 #2153 @UTH-Tuan
@arcuri82 Thanks. I am also encountering issues in running the example. Either that or is it not clear how to run the example (based off what was presented in the Example doc). I get the error "Cannot communicate with remote EvoMaster Driver for the system under test at localhost:40100 Make sure the EvoMaster Driver for the system under test is running correctly". Please clarify and/or make the example more straightforward.
