### 2020-07-21 #2153 @arcuri82
@UTH-Tuan in the main readme of the documentation: *A short video (5 minutes) shows the use of EvoMaster on one of the case studies in EMB.* [Link](https://www.youtube.com/watch?v=3mYxjgnhLEo&feature=youtu.be).
This should help better see how things are connected, besides the description in the [documentation](https://github.com/EMResearch/EvoMaster/blob/master/docs/whitebox.md).

Btw, could be better to create new GitHub issues in the [EvoMaster repository](https://github.com/EMResearch/EvoMaster/issues), instead of having the review here (see [JOSS documentation](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)).
