### 2020-07-28 EMResearch/EvoMaster#199 @mado89
Whitebox testing step by step
Hi,

I am again trying to get EvoMaster running for testing an application but failing miserably again :(
Here is what I did so far:
* Created a new project using spring initializr
* Loaded the project into eclipse
* Added one test controller
* Tried to implement an EmbeddedSutController

What fails:
* In eclipse I can't get it to locate EvoMaster i.e. it cannot find the package
* From the commandline I get the error message below (running `gradle testClasses`).

> martin@MAILUFTERL:~/projects/evomastertest$` gradle testClasses
> > Task :compileTestJava FAILED
>
> FAILURE: Build failed with an exception.
>
> * What went wrong:
> Execution failed for task ':compileTestJava'.
> > Could not resolve all files for configuration ':testCompileClasspath'.
>    > Could not find org.evomaster:evomaster-client-java-controller:LATEST.
>      Required by:
>          project :
>
> * Try:
> Run with --stacktrace option to get the stack trace. Run with --info or --debug option to get more log output. Run with --scan to get full insights.
>
> * Get more help at https://help.gradle.org
>
> Deprecated Gradle features were used in this build, making it incompatible with Gradle 7.0.
> Use '--warning-mode all' to show the individual deprecation warnings.
> See https://docs.gradle.org/6.5.1/userguide/command_line_interface.html#sec:command_line_warnings
>
> BUILD FAILED in 3s
> 3 actionable tasks: 3 executed
>

The project is located under: https://github.com/mado89/EvoMasterTest

openjournals/joss-reviews#2153
