### 2020-07-29 EMResearch/EvoMaster#199 @arcuri82
hi,
  well, that is expected. From documentation:

>In Gradle, it would be:
>
>testCompile('org.evomaster:evomaster-client-java-controller:LATEST').
>
>For the latest version, check Maven Central Repository. The latest version number should also appear at the top of the main readme page. If you are compiling directly from the EvoMaster source code, make sure to use mvn install to install the snapshot version of the Java client into your local Maven repository (e.g., under ~/.m2).

I guess I should make it explicit that the placeholder `LATEST` must be replaced with the actual version number, e.g. either `1.0.1` (most recent release) or `1.0.2-SNAPSHOT` (work-in-progress version)
