### 2020-07-29 EMResearch/EvoMaster#199 @mado89
Yest that would be good. (Afaik 'LATEST' could in theory work). '1.0.2-SNAPSHOT' did not work for me. However, '1.0.1' did work. Not I can start implementing the controller
