### 2020-07-29 EMResearch/EvoMaster#199 @arcuri82
`LATEST` as special keyword did work in Maven 2, but was removed in Maven 3.
regarding `1.0.2-SNAPSHOT`, did you do the following?

>  make sure to use mvn install to install the snapshot version of the Java client into your local Maven repository (e.g., under ~/.m2).
