### 2020-07-29 ccbfd89
clarification on LATEST placeholder
