### 2020-07-29 EMResearch/EvoMaster#199 @arcuri82
clarification added to the [documentation](https://github.com/EMResearch/EvoMaster/blob/master/docs/write_driver.md)
