### 2020-07-29 EMResearch/EvoMaster#199 @arcuri82
ah, forgot you are using Gradle... should check what is the equivalent of `mvn install` in Gradle.
