### 2020-07-29 EMResearch/EvoMaster#199 @mado89
> ah, forgot you are using Gradle... should check what is the equivalent of `mvn install` in Gradle.

Yes, that's why I did not notice this requirement. For now I guess it is fine. At least I can properly reference EvoMaster. The review is for 1.0.0 / 1.0.1 anyway
