### 2020-07-29 EMResearch/EvoMaster#199 @arcuri82
I added now a clarification about Gradle and `mvn install` in the documentation
