### 2020-08-04 EMResearch/EvoMaster#199 @mado89
That clarification looks good. I haven't tested the gradle way of using mvn install since I think this is not the typical usecase of your software (right?)
However, I'd like to have a better documentation of the EmbeddedSutController methods. It would be nice to have a list of all methods which need to be implemented and what they are supposed to do. Please integrate this in the documentation. References to other code is fine but users (and me) might / are lost when they do not have much idea what to do
