### 2020-08-05 EMResearch/EvoMaster#199 @arcuri82
All methods that needs to be implemented are `abstract`. Failure to implement them leads to compilation errors.

Each method provides documentation as Javadocs. How to read them depends on your IDE.
They can also be read online [here](https://javadoc.io/doc/org.evomaster/evomaster-client-java-controller).
I added this to the documentation.

However, it might well be that some of those methods need more clarifications and more text. Is there any specific method you find unclear? or all of them need more text?
