### 2020-08-10 EMResearch/EvoMaster#199 @mado89
@arcuri82 Yes in my opinion they need much more documentation. Probably not documentation but rather a guidance HOW to implement them. By the way not all methods are documented. For example getAdditionalInfoList is not documented at all (if I am not mistaken).

I'd suggest to add information on the intended behaviour to both the java-documentation but also to the whitebox testing instructions at least for the most important ones. Keep in mind people have not used your software and have no clue about e.g. the nomenclature of it
