### 2020-08-16 EMResearch/EvoMaster#199 @mado89
@arcuri82 Just wanted to check whether you received my last message?

I think that all methods need an explanation of how to implemented them or more clarification about their intended behaviour. I guess in most cases you could copy examples from your examples directory?
