### 2020-08-20 #2153 @s0nata
Dear everybody, thank you for your patience with me. Got significantly affected at work due to covid situation, and only recently was able to return to a schedule that premits time for the review.

For my review the only missing part is the evaluation, and @arcuri82 , by any chance is there and example project I can set up and run EvoMaster on? I saw the links in the docs for black- and white-box testing setups, they do describe some possible use cases, but I would find it extremely helpful if there was a follow-up script (as in theater play script) for "Hey, to see how our tool works on a sample project, build it, then use this small toy project to evaluate it on, you should get following results for scenarios A, B, and C". Having a docker image with a demo would be probably too resource-consuming to produce, so not asking for it.
