### 2020-08-21 #2153 @arcuri82
Hi @s0nata,

   having a demo project would indeed be useful. Although these weeks I am quite busy with the starting of teaching semester, so might take a while before I have time to prepare it.

Anyway, for the current time, you can look at the following: from the [main readme](https://github.com/EMResearch/EvoMaster), there is a link to the [EMB](https://github.com/EMResearch/EMB) repository, which contains a collection of REST APIs with also drivers for white-box testing. From the main readme, there is also a link to a [YouTube video](https://www.youtube.com/watch?v=3mYxjgnhLEo&feature=youtu.be) showing how to run EM on one of such projects.
