### 2020-08-23 EMResearch/EvoMaster#199 @arcuri82
More JavaDocs have now been added
