### 2020-08-24 7bef7b8
info on libraries needed by the generated tests
