### 2020-08-25 #2153 @arcuri82
In the end, it took less than I thought.
I created a small example of API, available at [rest-api-example](https://github.com/EMResearch/rest-api-example).
I also made a [video](https://youtu.be/ORxZoYw7LnM) showing how to implement an`EMDriver` to enable white-box testing with EvoMaster.
Video is now linked from the main Readme documentation.
