### 2020-09-10 #2153 @mado89
The video looks good. I took some notes while watchin it and will share them with you. Maybe they could be helpful for potential users of EvoMaster.
Using the video I think I will be able to do my own testing using Evo Master.

Just one thing: Right at the end of the video either the audio cuts off or you stop speaking. Either way it is a bit hard to follow what you are doing (but I think I might be able to reproduce it using your other video about blackbox testing).
