### 2020-09-30 40ab8b6
gRPC Ind_1Server example
