### 2020-10-01 c9c7fc5
gRPC tests for helloworld and ind_1
