### 2020-10-01 089e48d
ProtoBuf parser tests for e2e ind_1.proto
