+++ e2e-tests/grpc-java-examples/src/test/java/org/evomaster/protobuf/ProtoParserTest.java
@@ -1,0 +1,730 @@
+package org.evomaster.protobuf;
+
+
+import com.squareup.wire.schema.Field;
+import com.squareup.wire.schema.Location;
+import com.squareup.wire.schema.internal.parser.*;
+import org.junit.Test;
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
+
+public class ProtoParserTest {
+
+    @Test
+    public void testParserHelloWorld() {
+
+        Location location = Location.get("file.proto");
+        String data = "syntax = \"proto3\";\n" +
+                "\n" +
+                "option java_multiple_files = true;\n" +
+                "option java_package = \"io.grpc.examples.helloworld\";\n" +
+                "option java_outer_classname = \"HelloWorldProto\";\n" +
+                "option objc_class_prefix = \"HLW\";\n" +
+                "\n" +
+                "package helloworld;\n" +
+                "\n" +
+                "// The greeting service definition.\n" +
+                "service Greeter {\n" +
+                "  // Sends a greeting\n" +
+                "  rpc SayHello (HelloRequest) returns (HelloReply) {}\n" +
+                "}\n" +
+                "\n" +
+                "// The request message containing the user's name.\n" +
+                "message HelloRequest {\n" +
+                "  string name = 1;\n" +
+                "}\n" +
+                "\n" +
+                "// The response message containing the greetings\n" +
+                "message HelloReply {\n" +
+                "  string message = 1;\n" +
+                "}";
+        ProtoFileElement parsedFile = ProtoParser.Companion.parse(location, data);
+
+        assertEquals("helloworld", parsedFile.getPackageName());
+        assertEquals(2, parsedFile.getTypes().size());
+        assertEquals(1, parsedFile.getServices().size());
+
+        ServiceElement service = parsedFile.getServices().get(0);
+        assertEquals("Greeter", service.getName());
+        assertEquals(1, service.getRpcs().size());
+
+        RpcElement rpc = service.getRpcs().get(0);
+        assertEquals("SayHello", rpc.getName());
+
+        String requestType = rpc.getRequestType();
+        assertEquals("HelloRequest", requestType);
+
+        String responseType = rpc.getResponseType();
+        assertEquals("HelloReply", responseType);
+        {
+            assertTrue(
+                    parsedFile.getTypes()
+                            .stream()
+                            .anyMatch(typeElement -> typeElement.getName().equals("HelloReply")));
+
+            parsedFile.getTypes()
+                    .stream()
+                    .filter(typeElement -> typeElement.getName().equals("HelloReply"))
+                    .findFirst()
+                    .ifPresent(
+                            helloReply -> {
+                                assertEquals("HelloReply", helloReply.getName());
+                                assertEquals(0, helloReply.getOptions().size());
+                                assertEquals(0, helloReply.getNestedTypes().size());
+
+                                assertTrue(helloReply instanceof MessageElement);
+                                MessageElement helloReplyAsMessageElement = (MessageElement) helloReply;
+
+                                assertEquals(1, helloReplyAsMessageElement.getFields().size());
+                                FieldElement messageFileld = helloReplyAsMessageElement.getFields().get(0);
+
+
+                                assertEquals("message", messageFileld.getName());
+                                assertEquals("string", messageFileld.getType());
+                            }
+                    );
+
+        }
+        {
+            assertTrue(
+                    parsedFile.getTypes()
+                            .stream()
+                            .anyMatch(typeElement -> typeElement.getName().equals("HelloRequest")));
+
+            parsedFile.getTypes()
+                    .stream()
[…]
