+++ client-python/evomaster-client-python/client/app.py
@@ -1,0 +1,6 @@
+from flask import Flask, Blueprint
+from controller import controller
+
+
+app = Flask(__name__)
+app.register_blueprint(controller)
