+++ client-python/evomaster-client-python/requirements.txt
@@ -1,0 +1,19 @@
+aniso8601==8.0.0
+attrs==20.2.0
+certifi==2020.6.20
+chardet==3.0.4
+click==7.1.2
+Flask==1.1.2
+flask-restx==0.2.0
+idna==2.10
+itsdangerous==1.1.0
+Jinja2==2.11.2
+jsonschema==3.2.0
+MarkupSafe==1.1.1
+pyrsistent==0.17.3
+python-dotenv==0.14.0
+pytz==2020.1
+requests==2.24.0
+six==1.15.0
+urllib3==1.25.10
+Werkzeug==1.0.1
