+++ client-python/evomaster-client-python/instrumentation/execution_tracer.py
@@ -1,0 +1,241 @@
+class ExecutionTracer:
+    def __init__(self) -> None:
+        self.reset()
+
+    def reset(self) -> None:
+        self.objective_coverage = {}
+        self.action_index = 0
+        self.input_variables = set()
+        self.additional_info_list = []
+        self.additional_info_list.append(AdditionalInfo())
+
+    # static setAction(action: Action) {
+    #     if (action.getIndex() != ExecutionTracer.actionIndex) {
+    #         ExecutionTracer.actionIndex = action.getIndex();
+    #         ExecutionTracer.additionalInfoList.push(new AdditionalInfo());
+    #     }
+
+    #     if (action.getInputVariables() && action.getInputVariables().size > 0) {
+    #         ExecutionTracer.inputVariables = action.getInputVariables();
+    #     }
+    # }
+
+    # /**
+    #  * Check if the given input represented a tainted value from the test cases.
+    #  * This could be based on static info of the input (eg, according to a precise
+    #  * name convention given by TaintInputName), or dynamic info given directly by
+    #  * the test itself (eg, the test at action can register a list of values to check
+    #  * for)
+    #  */
+    # // public static boolean isTaintInput(String input){
+    # //     return TaintInputName.isTaintInput(input) || inputVariables.contains(input);
+    # // }
+
+
+    # // public static TaintType getTaintType(String input){
+    # //
+    # //     if(input == null){
+    # //         return TaintType.NONE;
+    # //     }
+    # //
+    # //     if(isTaintInput(input)){
+    # //         return TaintType.FULL_MATCH;
+    # //     }
+    # //
+    # //     if(TaintInputName.includesTaintInput(input)
+    # //         || inputVariables.stream().anyMatch(v -> input.contains(v))){
+    # //         return TaintType.PARTIAL_MATCH;
+    # //     }
+    # //
+    # //     return TaintType.NONE;
+    # // }
+
+
+    # static exposeAdditionalInfoList(): Array<AdditionalInfo> {
+    #     return ExecutionTracer.additionalInfoList;
+    # }
+
+    # // public static void addQueryParameter(String param){
+    # //     additionalInfoList.get(actionIndex).addQueryParameter(param);
+    # // }
+    # //
+    # // public static void addHeader(String header){
+    # //     additionalInfoList.get(actionIndex).addHeader(header);
+    # // }
+    # //
+    # // public static void addStringSpecialization(String taintInputName, StringSpecializationInfo info){
+    # //     additionalInfoList.get(actionIndex).addSpecialization(taintInputName, info);
+    # // }
+
+
+    # public static markLastExecutedStatement(lastLine: string) {
+
+    #     /*
+    #         There is a possible issue here: when there is an exception, there
+    #         is no pop of the stmt. So, the "call-stack" until the exception will still
+    #         be saved in this stack, even if computation continues (eg after a try/catch).
+    #         This is possibly a memory leak
+    #      */
+
+    #     ExecutionTracer.additionalInfoList[ExecutionTracer.actionIndex]
+    #         .pushLastExecutedStatement(lastLine);
+    # }
+
+
+    # public static completedLastExecutedStatement(lastLine: string) {
+    #     const stmt = ExecutionTracer.additionalInfoList[ExecutionTracer.actionIndex].popLastExecutedStatement();
+    #     if (stmt !== lastLine) {
+    #         /*
+    #             actually we cannot have such check. We might end in such situation:
+
+    #             X calls F in non-instrumented framework, which then call Y (both X and Y being of SUT).
[…]
