+++ e2e-tests/spring-examples/src/test/java/org/evomaster/e2etests/spring/examples/impactXYZ/ImpactXYZTestBase.java
@@ -4,6 +4,8 @@ import com.foo.rest.examples.spring.impactXYZ.ImpactXYZRestController;
 import org.evomaster.e2etests.spring.examples.SpringTestBase;
 import org.junit.jupiter.api.BeforeAll;
 
+import java.util.Arrays;
+
 /**
  * created by manzh on 2020-06-16
  */
@@ -11,6 +13,6 @@ public class ImpactXYZTestBase extends SpringTestBase  {
 
     @BeforeAll
     public static void initClass() throws Exception {
-        SpringTestBase.initClass(new ImpactXYZRestController());
+        SpringTestBase.initClass(new ImpactXYZRestController(Arrays.asList("/api/intImpact/{name}")));
     }
 }
