### 2020-10-06 6fe74b3
fixing issue in Tomcat sending Connection Close
