+++ client-java/instrumentation/src/main/java/org/evomaster/client/java/instrumentation/coverage/methodreplacement/ReplacementList.java
@@ -3,9 +3,7 @@ package org.evomaster.client.java.instrumentation.coverage.methodreplacement;
 import org.evomaster.client.java.instrumentation.coverage.methodreplacement.classes.*;
 import org.evomaster.client.java.instrumentation.shared.ClassName;
 
-import java.util.Arrays;
-import java.util.List;
-import java.util.Objects;
+import java.util.*;
 import java.util.stream.Collectors;
 
 public class ReplacementList {
@@ -46,7 +44,7 @@ public class ReplacementList {
         final String targetClassName = ClassName.get(target).getFullNameWithDots();
 
         return getList().stream()
-                .filter(t -> t.isAvailable())
+                //.filter(t -> t.isAvailable()) // bad idea to load 3rd classes at this point...
                 .filter(t -> {
                     /*
                         TODO: this is tricky, due to how "super" calls are
@@ -61,9 +59,13 @@ public class ReplacementList {
 
 //                            boolean jdk = targetClassName.startsWith("java.");
                             //TODO based on actual packages used in the list
-                            boolean jdk = targetClassName.startsWith("java.lang.") ||
-                                    targetClassName.startsWith("java.util.") ||
-                                    targetClassName.startsWith("java.time.");
+                            Set<String> prefixes = new HashSet<>();
+                            prefixes.add("java.lang.");
+                            prefixes.add("java.util.");
+                            prefixes.add("java.time.");
+
+                            boolean jdk = prefixes.stream().anyMatch(k -> targetClassName.startsWith(k)) &&
+                                        prefixes.stream().anyMatch(k -> t.getTargetClassName().startsWith(k));
 
                             if (jdk) {
                                 Class<?> klass;
