+++ client-java/instrumentation/src/main/java/org/evomaster/client/java/instrumentation/coverage/methodreplacement/ThirdPartyMethodReplacementClass.java
@@ -37,17 +37,20 @@ public abstract class ThirdPartyMethodReplacementClass implements MethodReplacem
 
     protected ThirdPartyMethodReplacementClass(){
 
-        if(! isAvailable()){
-            //nothing to initialize
-            return;
-        }
+//        if(! isAvailable()){
+//            //nothing to initialize
+//            return;
+//        }
+        //initMethods();
+    }
 
-        /*
-            Use reflection to load all methods that were replaced.
-            This is essential to simplify the writing of the replacement, as those
-            must still call the original, but only via reflection (as original third-party
-            library must not included in EvoMaster)
-         */
+    private  void initMethods() {
+    /*
+        Use reflection to load all methods that were replaced.
+        This is essential to simplify the writing of the replacement, as those
+        must still call the original, but only via reflection (as original third-party
+        library must not included in EvoMaster)
+     */
         Class<? extends ThirdPartyMethodReplacementClass> subclass = this.getClass();
 
         for (Method m : subclass.getDeclaredMethods()) {
@@ -75,7 +78,7 @@ public abstract class ThirdPartyMethodReplacementClass implements MethodReplacem
             Method targetMethod;
             try {
                 //this will not return private methods
-                targetMethod = targetClass.getMethod(m.getName(), reducedInputs);
+                targetMethod = getTargetClass().getMethod(m.getName(), reducedInputs);
             } catch (NoSuchMethodException e) {
                 try {
                     //this would return private methods, but not public in superclasses
@@ -107,6 +110,9 @@ public abstract class ThirdPartyMethodReplacementClass implements MethodReplacem
     public static Method getOriginal(ThirdPartyMethodReplacementClass singleton, String id){
         if(id == null || id.isEmpty()){
             throw new IllegalArgumentException("Invalid empty id");
+        }
+        if(singleton.methods.isEmpty()){
+            singleton.initMethods();
         }
         Method original = singleton.methods.get(id);
         if(original == null){
