### 2020-10-06 EMResearch/EvoMaster#199 @mado89
I am almost there. I hope there is just one last issue. When I start the driver and run evomaster the application starts but the `v3/api-docs` endpoint does not work. The application is started and I can access its endpoints but evomaster fails due to `v3/api-docs` not being available. Can you please help?

I receive the following output from evomaster:
`java -jar ~/Downloads/evomaster.jar --maxTime 20s --outputFolder backboxTest/
*
 _____          ___  ___          _
|  ___|         |  \/  |         | |
| |____   _____ | .  . | __ _ ___| |_ ___ _ __
|  __\ \ / / _ \| |\/| |/ _` / __| __/ _ \ '__|
| |___\ V / (_) | |  | | (_| \__ \ ||  __/ |
\____/ \_/ \___/\_|  |_/\__,_|___/\__\___|_|

* EvoMaster version: 1.0.1
WARNING: An illegal reflective access operation has occurred
WARNING: Illegal reflective access by com.google.inject.internal.cglib.core.$ReflectUtils$1 (file:/home/martin/Downloads/evomaster.jar) to method java.lang.ClassLoader.defineClass(java.lang.String,byte[],int,int,java.security.ProtectionDomain)
WARNING: Please consider reporting this to the maintainers of com.google.inject.internal.cglib.core.$ReflectUtils$1
WARNING: Use --illegal-access=warn to enable warnings of further illegal reflective access operations
WARNING: All illegal access operations will be denied in a future release
* [ERROR] ERROR in the Remote EvoMaster Driver: Cannot retrieve Swagger JSON data from http://localhost:8080/v3/api-docs , status=404
  Look at the logs of the EvoMaster Driver to help debugging this problem.
`
And the following from my driver:
`14:35:09.221 [main] DEBUG shaded.org.eclipse.jetty.util.log - Logging to Logger[shaded.org.eclipse.jetty.util.log] via shaded.org.eclipse.jetty.util.log.Slf4jLog
14:35:09.252 [main] INFO shaded.org.eclipse.jetty.util.log - Logging initialized @1161ms to shaded.org.eclipse.jetty.util.log.Slf4jLog
14:35:09.335 [main] DEBUG shaded.org.eclipse.jetty.util.component.ContainerLifeCycle - Server@59e5ddf{STOPPED}[9.4.z-SNAPSHOT] added {QueuedThreadPool[qtp1316864772]@4e7dc304{STOPPED,8<=0<=200,i=0,q=0}[shaded.org.eclipse.jetty.util.thread.TryExecutor$$Lambda$35/762227630@80503],AUTO}
14:35:09.533 [main] DEBUG shaded.org.eclipse.jetty.util.component.ContainerLifeCycle - HttpConnectionFactory@51c8530f[HTTP/1.1] added {HttpConfiguration@7403c468{32768/8192,8192/8192,https://:0,[]},POJO}
14:35:09.554 [main] DEBUG shaded.org.eclipse.jetty.util.component.ContainerLifeCycle - ServerConnector@12c8a2c0{null,[]}{0.0.0.0:0} added {Server@59e5ddf{STOPPED}[9.4.z-SNAPSHOT],UNMANAGED}
14:35:09.556 [main] DEBUG shaded.org.eclipse.jetty.util.component.ContainerLifeCycle - ServerConnector@12c8a2c0{null,[]}{0.0.0.0:0} added {QueuedThreadPool[qtp1316864772]@4e7dc304{STOPPED,8<=0<=200,i=0,q=0}[shaded.org.eclipse.jetty.util.thread.TryExecutor$$Lambda$35/762227630@80503],AUTO}
14:35:09.557 [main] DEBUG shaded.org.eclipse.jetty.util.component.ContainerLifeCycle - ServerConnector@12c8a2c0{null,[]}{0.0.0.0:0} added {ScheduledExecutorScheduler@22eeefeb{STOPPED},AUTO}
14:35:09.557 [main] DEBUG shaded.org.eclipse.jetty.util.component.ContainerLifeCycle - ServerConnector@12c8a2c0{null,[]}{0.0.0.0:0} added {shaded.org.eclipse.jetty.io.ArrayByteBufferPool@7e0e6aa2,POJO}
14:35:09.559 [main] DEBUG shaded.org.eclipse.jetty.util.component.ContainerLifeCycle - ServerConnector@12c8a2c0{null,[http/1.1]}{0.0.0.0:0} added {HttpConnectionFactory@51c8530f[HTTP/1.1],AUTO}
14:35:09.560 [main] DEBUG shaded.org.eclipse.jetty.server.AbstractConnector - ServerConnector@12c8a2c0{HTTP/1.1,[http/1.1]}{0.0.0.0:0} added HttpConnectionFactory@51c8530f[HTTP/1.1]
14:35:09.575 [main] DEBUG shaded.org.eclipse.jetty.util.component.ContainerLifeCycle - ServerConnector@12c8a2c0{HTTP/1.1,[http/1.1]}{0.0.0.0:0} added {SelectorManager@ServerConnector@12c8a2c0{HTTP/1.1,[http/1.1]}{0.0.0.0:0},MANAGED}
14:35:09.576 [main] DEBUG shaded.org.eclipse.jetty.util.component.ContainerLifeCycle - Server@59e5ddf{STOPPED}[9.4.z-SNAPSHOT] added {ServerConnector@12c8a2c0{HTTP/1.1,[http/1.1]}{localhost:40100},AUTO}
14:35:09.590 [main] DEBUG shaded.org.eclipse.jetty.util.component.ContainerLifeCycle - Server@59e5ddf{STOPPED}[9.4.z-SNAPSHOT] added {ErrorHandler@67b467e9{STOPPED},AUTO}
14:35:09.853 [main] DEBUG shaded.org.eclipse.jetty.util.DecoratedObjectFactory - Adding Decorator: shaded.org.eclipse.jetty.util.DeprecationWarning@5cee5251
14:35:09.854 [main] WARN shaded.org.eclipse.jetty.server.handler.ContextHandler - s.o.e.j.s.ServletContextHandler@433d61fb{/,null,UNAVAILABLE} contextPath ends with /*
14:35:09.867 [main] DEBUG shaded.org.eclipse.jetty.util.component.ContainerLifeCycle - Server@59e5ddf{STOPPED}[9.4.z-SNAPSHOT] added {s.o.e.j.s.ServletContextHandler@433d61fb{/controller/api,null,UNAVAILABLE},MANAGED}
14:35:09.878 [main] DEBUG shaded.org.eclipse.jetty.util.component.ContainerLifeCycle - s.o.e.j.s.ServletContextHandler@433d61fb{/controller/api,null,UNAVAILABLE} added {ServletHandler@460d0a57{STOPPED},MANAGED}
14:35:09.881 [main] DEBUG shaded.org.eclipse.jetty.util.component.ContainerLifeCycle - ServletHandler@460d0a57{STOPPED} added {shaded.org.glassfish.jersey.servlet.ServletContainer-50de0926@45b718ee==shaded.org.glassfish.jersey.servlet.ServletContainer,jsp=null,order=-1,inst=true,AUTO}
14:35:09.883 [main] DEBUG shaded.org.eclipse.jetty.util.component.ContainerLifeCycle - ServletHandler@460d0a57{STOPPED} added {[/*]=>shaded.org.glassfish.jersey.servlet.ServletContainer-50de0926,POJO}
14:35:09.883 [main] DEBUG shaded.org.eclipse.jetty.util.component.AbstractLifeCycle - starting Server@59e5ddf{STOPPED}[9.4.z-SNAPSHOT]
14:35:09.888 [main] INFO shaded.org.eclipse.jetty.server.Server - jetty-9.4.z-SNAPSHOT; built: 2018-08-30T13:59:14.071Z; git: 27208684755d94a92186989f695db2d7b21ebc51; jvm 1.8.0_265-8u265-b01-0ubuntu2~20.04-b01
14:35:09.933 [main] DEBUG shaded.org.eclipse.jetty.server.handler.AbstractHandler - starting Server@59e5ddf{STARTING}[9.4.z-SNAPSHOT]
14:35:09.933 [main] DEBUG shaded.org.eclipse.jetty.util.component.AbstractLifeCycle - starting QueuedThreadPool[qtp1316864772]@4e7dc304{STOPPED,8<=0<=200,i=0,q=0}[shaded.org.eclipse.jetty.util.thread.TryExecutor$$Lambda$35/762227630@80503]
14:35:09.938 [main] DEBUG shaded.org.eclipse.jetty.util.thread.ReservedThreadExecutor - ReservedThreadExecutor@418e7838{s=0/4,p=0}
14:35:09.938 [main] DEBUG shaded.org.eclipse.jetty.util.component.ContainerLifeCycle - QueuedThreadPool[qtp1316864772]@4e7dc304{STARTING,8<=0<=200,i=0,q=0}[ReservedThreadExecutor@418e7838{s=0/4,p=0}] added {ReservedThreadExecutor@418e7838{s=0/4,p=0},AUTO}
14:35:09.939 [main] DEBUG shaded.org.eclipse.jetty.util.component.AbstractLifeCycle - starting ReservedThreadExecutor@418e7838{s=0/4,p=0}
14:35:09.945 [main] DEBUG shaded.org.eclipse.jetty.util.component.AbstractLifeCycle - STARTED @1857ms ReservedThreadExecutor@418e7838{s=0/4,p=0}
14:35:09.948 [main] DEBUG shaded.org.eclipse.jetty.util.component.AbstractLifeCycle - STARTED @1859ms QueuedThreadPool[qtp1316864772]@4e7dc304{STARTED,8<=8<=200,i=6,q=0}[ReservedThreadExecutor@418e7838{s=0/4,p=0}]
14:35:09.948 [main] DEBUG shaded.org.eclipse.jetty.util.component.AbstractLifeCycle - starting ErrorHandler@67b467e9{STOPPED}
14:35:09.948 [main] DEBUG shaded.org.eclipse.jetty.server.handler.AbstractHandler - starting ErrorHandler@67b467e9{STARTING}
14:35:09.949 [main] DEBUG shaded.org.eclipse.jetty.util.component.AbstractLifeCycle - STARTED @1861ms ErrorHandler@67b467e9{STARTED}
14:35:09.949 [main] DEBUG shaded.org.eclipse.jetty.util.component.AbstractLifeCycle - starting s.o.e.j.s.ServletContextHandler@433d61fb{/controller/api,null,UNAVAILABLE}
14:35:09.964 [main] DEBUG shaded.org.eclipse.jetty.http.PreEncodedHttpField - HttpField encoders loaded: [shaded.org.eclipse.jetty.http.Http1FieldPreEncoder@3daa422a]
14:35:09.978 [main] DEBUG shaded.org.eclipse.jetty.server.handler.AbstractHandler - starting s.o.e.j.s.ServletContextHandler@433d61fb{/controller/api,null,STARTING}
14:35:09.978 [main] DEBUG shaded.org.eclipse.jetty.util.component.AbstractLifeCycle - starting ServletHandler@460d0a57{STOPPED}
14:35:09.989 [main] DEBUG shaded.org.eclipse.jetty.servlet.ServletHandler - Path=/*[EMBEDDED:null] mapped to servlet=shaded.org.glassfish.jersey.servlet.ServletContainer-50de0926[EMBEDDED:null]
14:35:09.994 [main] DEBUG shaded.org.eclipse.jetty.http.pathmap.PathMappings - Added MappedResource[pathSpec=ServletPathSpec["/*",pathDepth=1,group=PREFIX_GLOB],resource=shaded.org.glassfish.jersey.servlet.ServletContainer-50de0926@45b718ee==shaded.org.glassfish.jersey.servlet.ServletContainer,jsp=null,order=-1,inst=true] to PathMappings[size=1]
14:35:09.995 [main] DEBUG shaded.org.eclipse.jetty.servlet.ServletHandler - filterNameMap={}
14:35:09.995 [main] DEBUG shaded.org.eclipse.jetty.servlet.ServletHandler - pathFilters=null
14:35:09.995 [main] DEBUG shaded.org.eclipse.jetty.servlet.ServletHandler - servletFilterMap=null
14:35:09.996 [main] DEBUG shaded.org.eclipse.jetty.servlet.ServletHandler - servletPathMap=PathMappings[size=1]
14:35:09.996 [main] DEBUG shaded.org.eclipse.jetty.servlet.ServletHandler - servletNameMap={shaded.org.glassfish.jersey.servlet.ServletContainer-50de0926=shaded.org.glassfish.jersey.servlet.ServletContainer-50de0926@45b718ee==shaded.org.glassfish.jersey.servlet.ServletContainer,jsp=null,order=-1,inst=true}
14:35:09.996 [main] DEBUG shaded.org.eclipse.jetty.servlet.ServletHandler - Adding Default404Servlet to ServletHandler@460d0a57{STARTING}
14:35:09.997 [main] DEBUG shaded.org.eclipse.jetty.util.component.ContainerLifeCycle - ServletHandler@460d0a57{STARTING} added {shaded.org.eclipse.jetty.servlet.ServletHandler$Default404Servlet-20ce78ec@ea03695c==shaded.org.eclipse.jetty.servlet.ServletHandler$Default404Servlet,jsp=null,order=-1,inst=false,AUTO}
14:35:09.998 [main] DEBUG shaded.org.eclipse.jetty.util.component.ContainerLifeCycle - ServletHandler@460d0a57{STARTING} added {[/]=>shaded.org.eclipse.jetty.servlet.ServletHandler$Default404Servlet-20ce78ec,POJO}
14:35:09.998 [main] DEBUG shaded.org.eclipse.jetty.servlet.ServletHandler - Path=/*[EMBEDDED:null] mapped to servlet=shaded.org.glassfish.jersey.servlet.ServletContainer-50de0926[EMBEDDED:null]
14:35:09.998 [main] DEBUG shaded.org.eclipse.jetty.http.pathmap.PathMappings - Added MappedResource[pathSpec=ServletPathSpec["/*",pathDepth=1,group=PREFIX_GLOB],resource=shaded.org.glassfish.jersey.servlet.ServletContainer-50de0926@45b718ee==shaded.org.glassfish.jersey.servlet.ServletContainer,jsp=null,order=-1,inst=true] to PathMappings[size=1]
14:35:09.999 [main] DEBUG shaded.org.eclipse.jetty.servlet.ServletHandler - Path=/[EMBEDDED:null] mapped to servlet=shaded.org.eclipse.jetty.servlet.ServletHandler$Default404Servlet-20ce78ec[EMBEDDED:null]
14:35:09.999 [main] DEBUG shaded.org.eclipse.jetty.http.pathmap.PathMappings - Added MappedResource[pathSpec=ServletPathSpec["/",pathDepth=-1,group=DEFAULT],resource=shaded.org.eclipse.jetty.servlet.ServletHandler$Default404Servlet-20ce78ec@ea03695c==shaded.org.eclipse.jetty.servlet.ServletHandler$Default404Servlet,jsp=null,order=-1,inst=false] to PathMappings[size=2]
14:35:09.999 [main] DEBUG shaded.org.eclipse.jetty.servlet.ServletHandler - filterNameMap={}
14:35:09.999 [main] DEBUG shaded.org.eclipse.jetty.servlet.ServletHandler - pathFilters=null
14:35:09.999 [main] DEBUG shaded.org.eclipse.jetty.servlet.ServletHandler - servletFilterMap=null
14:35:10.000 [main] DEBUG shaded.org.eclipse.jetty.servlet.ServletHandler - servletPathMap=PathMappings[size=2]
14:35:10.000 [main] DEBUG shaded.org.eclipse.jetty.servlet.ServletHandler - servletNameMap={shaded.org.glassfish.jersey.servlet.ServletContainer-50de0926=shaded.org.glassfish.jersey.servlet.ServletContainer-50de0926@45b718ee==shaded.org.glassfish.jersey.servlet.ServletContainer,jsp=null,order=-1,inst=true, shaded.org.eclipse.jetty.servlet.ServletHandler$Default404Servlet-20ce78ec=shaded.org.eclipse.jetty.servlet.ServletHandler$Default404Servlet-20ce78ec@ea03695c==shaded.org.eclipse.jetty.servlet.ServletHandler$Default404Servlet,jsp=null,order=-1,inst=false}
14:35:10.001 [main] DEBUG shaded.org.eclipse.jetty.server.handler.AbstractHandler - starting ServletHandler@460d0a57{STARTING}
14:35:10.001 [main] DEBUG shaded.org.eclipse.jetty.util.component.AbstractLifeCycle - STARTED @1913ms ServletHandler@460d0a57{STARTED}
14:35:10.002 [main] DEBUG shaded.org.eclipse.jetty.util.component.AbstractLifeCycle - starting shaded.org.eclipse.jetty.servlet.ServletHandler$Default404Servlet-20ce78ec@ea03695c==shaded.org.eclipse.jetty.servlet.ServletHandler$Default404Servlet,jsp=null,order=-1,inst=false
14:35:10.004 [main] DEBUG shaded.org.eclipse.jetty.util.component.AbstractLifeCycle - STARTED @1916ms shaded.org.eclipse.jetty.servlet.ServletHandler$Default404Servlet-20ce78ec@ea03695c==shaded.org.eclipse.jetty.servlet.ServletHandler$Default404Servlet,jsp=null,order=-1,inst=false
14:35:10.005 [main] DEBUG shaded.org.eclipse.jetty.util.component.AbstractLifeCycle - starting shaded.org.glassfish.jersey.servlet.ServletContainer-50de0926@45b718ee==shaded.org.glassfish.jersey.servlet.ServletContainer,jsp=null,order=-1,inst=true
14:35:10.005 [main] DEBUG shaded.org.eclipse.jetty.util.component.AbstractLifeCycle - STARTED @1917ms shaded.org.glassfish.jersey.servlet.ServletContainer-50de0926@45b718ee==shaded.org.glassfish.jersey.servlet.ServletContainer,jsp=null,order=-1,inst=true
14:35:10.010 [main] DEBUG shaded.org.eclipse.jetty.servlet.ServletHolder - Servlet.init shaded.org.glassfish.jersey.servlet.ServletContainer@479d31f3 for shaded.org.glassfish.jersey.servlet.ServletContainer-50de0926
Oct 06, 2020 2:35:11 PM shaded.org.glassfish.jersey.internal.inject.Providers checkProviderRuntime
WARNING: A provider org.evomaster.client.java.controller.internal.EMController registered in SERVER runtime does not implement any provider interfaces applicable in the SERVER runtime. Due to constraint configuration problems the provider org.evomaster.client.java.controller.internal.EMController will be ignored.
14:35:12.210 [main] INFO shaded.org.eclipse.jetty.server.handler.ContextHandler - Started s.o.e.j.s.ServletContextHandler@433d61fb{/controller/api,null,AVAILABLE}
14:35:12.211 [main] DEBUG shaded.org.eclipse.jetty.util.component.AbstractLifeCycle - STARTED @4123ms s.o.e.j.s.ServletContextHandler@433d61fb{/controller/api,null,AVAILABLE}
14:35:12.211 [main] DEBUG shaded.org.eclipse.jetty.util.component.AbstractLifeCycle - starting ServerConnector@12c8a2c0{HTTP/1.1,[http/1.1]}{localhost:40100}
14:35:12.219 [main] DEBUG shaded.org.eclipse.jetty.util.component.ContainerLifeCycle - ServerConnector@12c8a2c0{HTTP/1.1,[http/1.1]}{localhost:40100} added {sun.nio.ch.ServerSocketChannelImpl[/127.0.0.1:40100],POJO}
14:35:12.221 [main] DEBUG shaded.org.eclipse.jetty.util.component.AbstractLifeCycle - starting ScheduledExecutorScheduler@22eeefeb{STOPPED}
14:35:12.222 [main] DEBUG shaded.org.eclipse.jetty.util.component.AbstractLifeCycle - STARTED @4134ms ScheduledExecutorScheduler@22eeefeb{STARTED}
14:35:12.223 [main] DEBUG shaded.org.eclipse.jetty.util.component.AbstractLifeCycle - starting HttpConnectionFactory@51c8530f[HTTP/1.1]
14:35:12.223 [main] DEBUG shaded.org.eclipse.jetty.util.component.AbstractLifeCycle - STARTED @4135ms HttpConnectionFactory@51c8530f[HTTP/1.1]
14:35:12.223 [main] DEBUG shaded.org.eclipse.jetty.util.component.AbstractLifeCycle - starting SelectorManager@ServerConnector@12c8a2c0{HTTP/1.1,[http/1.1]}{localhost:40100}
14:35:12.297 [main] DEBUG shaded.org.eclipse.jetty.util.component.ContainerLifeCycle - EatWhatYouKill@1c98290c/SelectorProducer@3624da92/IDLE/p=false/QueuedThreadPool[qtp1316864772]@4e7dc304{STARTED,8<=8<=200,i=8,q=0}[ReservedThreadExecutor@418e7838{s=0/4,p=0}][pc=0,pic=0,pec=0,epc=0]@2020-10-06T14:35:12.289+02:00 added {SelectorProducer@3624da92,POJO}
14:35:12.298 [main] DEBUG shaded.org.eclipse.jetty.util.component.ContainerLifeCycle - EatWhatYouKill@1c98290c/SelectorProducer@3624da92/IDLE/p=false/QueuedThreadPool[qtp1316864772]@4e7dc304{STARTED,8<=8<=200,i=8,q=0}[ReservedThreadExecutor@418e7838{s=0/4,p=0}][pc=0,pic=0,pec=0,epc=0]@2020-10-06T14:35:12.298+02:00 added {QueuedThreadPool[qtp1316864772]@4e7dc304{STARTED,8<=8<=200,i=8,q=0}[ReservedThreadExecutor@418e7838{s=0/4,p=0}],UNMANAGED}
14:35:12.299 [main] DEBUG shaded.org.eclipse.jetty.util.thread.strategy.EatWhatYouKill - EatWhatYouKill@1c98290c/SelectorProducer@3624da92/IDLE/p=false/QueuedThreadPool[qtp1316864772]@4e7dc304{STARTED,8<=8<=200,i=8,q=0}[ReservedThreadExecutor@418e7838{s=0/4,p=0}][pc=0,pic=0,pec=0,epc=0]@2020-10-06T14:35:12.299+02:00 created
14:35:12.299 [main] DEBUG shaded.org.eclipse.jetty.util.component.ContainerLifeCycle - ManagedSelector@27a5328c{STOPPED} id=0 keys=-1 selected=-1 updates=0 added {EatWhatYouKill@1c98290c/SelectorProducer@3624da92/IDLE/p=false/QueuedThreadPool[qtp1316864772]@4e7dc304{STARTED,8<=8<=200,i=8,q=0}[ReservedThreadExecutor@418e7838{s=0/4,p=0}][pc=0,pic=0,pec=0,epc=0]@2020-10-06T14:35:12.299+02:00,MANAGED}
14:35:12.300 [main] DEBUG shaded.org.eclipse.jetty.util.component.ContainerLifeCycle - SelectorManager@ServerConnector@12c8a2c0{HTTP/1.1,[http/1.1]}{localhost:40100} added {ManagedSelector@27a5328c{STOPPED} id=0 keys=-1 selected=-1 updates=0,AUTO}
14:35:12.301 [main] DEBUG shaded.org.eclipse.jetty.util.component.ContainerLifeCycle - EatWhatYouKill@1e5f4170/SelectorProducer@6c345c5f/IDLE/p=false/QueuedThreadPool[qtp1316864772]@4e7dc304{STARTED,8<=8<=200,i=8,q=0}[ReservedThreadExecutor@418e7838{s=0/4,p=0}][pc=0,pic=0,pec=0,epc=0]@2020-10-06T14:35:12.301+02:00 added {SelectorProducer@6c345c5f,POJO}
14:35:12.302 [main] DEBUG shaded.org.eclipse.jetty.util.component.ContainerLifeCycle - EatWhatYouKill@1e5f4170/SelectorProducer@6c345c5f/IDLE/p=false/QueuedThreadPool[qtp1316864772]@4e7dc304{STARTED,8<=8<=200,i=8,q=0}[ReservedThreadExecutor@418e7838{s=0/4,p=0}][pc=0,pic=0,pec=0,epc=0]@2020-10-06T14:35:12.301+02:00 added {QueuedThreadPool[qtp1316864772]@4e7dc304{STARTED,8<=8<=200,i=8,q=0}[ReservedThreadExecutor@418e7838{s=0/4,p=0}],UNMANAGED}
14:35:12.302 [main] DEBUG shaded.org.eclipse.jetty.util.thread.strategy.EatWhatYouKill - EatWhatYouKill@1e5f4170/SelectorProducer@6c345c5f/IDLE/p=false/QueuedThreadPool[qtp1316864772]@4e7dc304{STARTED,8<=8<=200,i=8,q=0}[ReservedThreadExecutor@418e7838{s=0/4,p=0}][pc=0,pic=0,pec=0,epc=0]@2020-10-06T14:35:12.302+02:00 created
14:35:12.303 [main] DEBUG shaded.org.eclipse.jetty.util.component.ContainerLifeCycle - ManagedSelector@6b5966e1{STOPPED} id=1 keys=-1 selected=-1 updates=0 added {EatWhatYouKill@1e5f4170/SelectorProducer@6c345c5f/IDLE/p=false/QueuedThreadPool[qtp1316864772]@4e7dc304{STARTED,8<=8<=200,i=8,q=0}[ReservedThreadExecutor@418e7838{s=0/4,p=0}][pc=0,pic=0,pec=0,epc=0]@2020-10-06T14:35:12.303+02:00,MANAGED}
14:35:12.303 [main] DEBUG shaded.org.eclipse.jetty.util.component.ContainerLifeCycle - SelectorManager@ServerConnector@12c8a2c0{HTTP/1.1,[http/1.1]}{localhost:40100} added {ManagedSelector@6b5966e1{STOPPED} id=1 keys=-1 selected=-1 updates=0,AUTO}
14:35:12.303 [main] DEBUG shaded.org.eclipse.jetty.util.component.AbstractLifeCycle - starting ManagedSelector@27a5328c{STOPPED} id=0 keys=-1 selected=-1 updates=0
14:35:12.304 [main] DEBUG shaded.org.eclipse.jetty.util.component.AbstractLifeCycle - starting EatWhatYouKill@1c98290c/SelectorProducer@3624da92/IDLE/p=false/QueuedThreadPool[qtp1316864772]@4e7dc304{STARTED,8<=8<=200,i=8,q=0}[ReservedThreadExecutor@418e7838{s=0/4,p=0}][pc=0,pic=0,pec=0,epc=0]@2020-10-06T14:35:12.304+02:00
14:35:12.304 [main] DEBUG shaded.org.eclipse.jetty.util.component.AbstractLifeCycle - STARTED @4216ms EatWhatYouKill@1c98290c/SelectorProducer@3624da92/IDLE/p=false/QueuedThreadPool[qtp1316864772]@4e7dc304{STARTED,8<=8<=200,i=8,q=0}[ReservedThreadExecutor@418e7838{s=0/4,p=0}][pc=0,pic=0,pec=0,epc=0]@2020-10-06T14:35:12.304+02:00
14:35:12.308 [main] DEBUG shaded.org.eclipse.jetty.util.thread.QueuedThreadPool - queue shaded.org.eclipse.jetty.io.ManagedSelector$$Lambda$191/22446425@4fcee388
[…]
