### 2020-10-06 EMResearch/EvoMaster#199 @mado89
I assume this is due to some missing command annotation for spring docs. The endpoint works when I run `gradle bootRun`
