### 2020-10-06 EMResearch/EvoMaster#199 @arcuri82
are you using and configuring SpringFox or SpringDoc to generate the OpenAPI schema?
