### 2020-10-06 EMResearch/EvoMaster#199 @arcuri82
https://github.com/EMResearch/EvoMaster/blob/master/docs/openapi.md
