### 2020-10-06 EMResearch/EvoMaster#199 @mado89
SpringDoc (https://github.com/springdoc/springdoc-openapi). I did not configure something but just added the library to my gradle file
