### 2020-10-06 EMResearch/EvoMaster#199 @mado89
Also I am starting the driver from eclipse. Do I need to somehow add SpringDoc as a reference/library there?
