### 2020-10-06 EMResearch/EvoMaster#199 @arcuri82
adding it to Gradle should be enough
