### 2020-10-06 #2153 @kthyng
@gkthiruvathukal ah! I see. Ok, I will remove the one reviewer and edit the checklist to be for @UTH-Tuan instead. When reviewer changes happen after starting the review issue, some of this needs to happen by hand. Please let me know if anything remains incorrect after I'm done.
