### 2020-10-06 EMResearch/EvoMaster#199 @mado89
For reference, this is my gradle file:

	plugins {
		id 'org.springframework.boot' version '2.3.2.RELEASE'
		id 'io.spring.dependency-management' version '1.0.9.RELEASE'
		id 'java'
	}

	group = 'com.example'
	version = '0.0.1-SNAPSHOT'
	sourceCompatibility = '1.8'
	targetCompatibility = '1.8'

	configurations {
		compileOnly {
			extendsFrom annotationProcessor
		}
	}

	repositories {
		mavenCentral()
	}

	dependencies {
		implementation 'org.springframework.boot:spring-boot-starter-web'

		compileOnly 'org.projectlombok:lombok'

		annotationProcessor 'org.projectlombok:lombok'

		testImplementation('org.springframework.boot:spring-boot-starter-test') {
			exclude group: 'org.junit.vintage', module: 'junit-vintage-engine'
		}

		implementation group: 'org.springdoc', name: 'springdoc-openapi-ui', version: '1.4.8'

		testCompile('org.evomaster:evomaster-client-java-controller:1.0.1')
	}

	test {
		useJUnitPlatform()
	}
