### 2020-10-06 #2153 @UTH-Tuan
@gkthiruvathukal @kthyng how do I generate a checklist? I am sorry but this my first time doing this type of review without any direction  (I am more accustomed to the traditional paper review).
