### 2020-10-06 #2153 @kthyng
@UTH-Tuan You now have a checklist above to use. So the review is checklist-driven, as above, but as issues come up, please open issues on the software repo to describe and discuss, and link to this review issue so that we can track related issues. Let me or @gkthiruvathukal know if you have any questions. Thanks!
