### 2020-10-06 EMResearch/EvoMaster#199 @mado89
No. When I start the application it is available under `http://localhost:8080/v3/api-docs`.
The ProblemInfo method:

	@Override
	public ProblemInfo getProblemInfo() {
		return new RestProblem("http://localhost:8080/v3/api-docs", null);
	}

The start method:

	@Override
	public String startSut() {
		ctx = SpringApplication.run(DemoApplication.class, new String[] { });
		return "http://localhost:8080";
	}
