### 2020-10-06 EMResearch/EvoMaster#199 @arcuri82
> When I start the application it is available under http://localhost:8080/v3/api-docs.

Do you mean that, once the driver is up and running and did at least one run, then typing `http://localhost:8080/v3/api-docs` in browser (for example) actually gives you the schema? but during the search EM gives a 404?
