### 2020-10-06 EMResearch/EvoMaster#199 @mado89
Sorry. No I meant when I run the application (not the test driver) manually then it gives me the schema but not when I run the test driver.
However, when I run the test driver I can see the other endpoints.
Sorry for the confusion!
