### 2020-10-06 ac28458
fix e2etests modify the EMConfig
