+++ e2e-tests/spring-examples/src/test/java/org/evomaster/e2etests/spring/examples/testability/TestabilityEMTestWithAdaptiveMutation.java
@@ -18,7 +18,6 @@ public class TestabilityEMTestWithAdaptiveMutation extends SpringTestBase {
 
     @BeforeAll
     public static void initClass() throws Exception {
-
         SpringTestBase.initClass(new TestabilityController());
     }
 
