+++ e2e-tests/spring-examples/src/test/java/org/evomaster/e2etests/spring/examples/impactXYZ/WeightBasedMutationXYZTest.java
@@ -28,7 +28,7 @@ public class WeightBasedMutationXYZTest extends SpringTestBase {
     public void testRunEM() throws Throwable {
 
         runTestHandlingFlakyAndCompilation(
-                "ImpactXYZ_TestWM",
+                "ImpactWM",
                 "org.bar.ImpactWM",
                 1000,
                 true,
