+++ e2e-tests/spring-examples/src/test/java/org/evomaster/e2etests/spring/examples/testability/TestabilityEMTestWithAdaptiveMutation.java
@@ -28,7 +28,7 @@ public class TestabilityEMTestWithAdaptiveMutation extends SpringTestBase {
 
         runTestHandlingFlakyAndCompilation(
                 "TestabilityEM",
-                "org.bar.TestabilityEMTestWithAdaptiveMutation",
+                "org.bar.TestabilityEMAGM",
                 8_000,
                 true,
                 (args) -> {
