### 2020-10-07 EMResearch/EvoMaster#199 @arcuri82
that's very weird... but it looks more like an issue with SpringDoc, or something in your project that is misconfigured somehow
