### 2020-10-07 EMResearch/EvoMaster#199 @mado89
Thats my impression as well. But for me it seems to be in connection with evomaster. Maybe we can arrange a quick screenshare that we can check together? (I know a lot to ask ...) Or is there some guess on your side in which direction I should investigate for the misconfiguration?
