### 2020-10-07 EMResearch/EvoMaster#199 @arcuri82
if you do not manage to get SpringDoc work, you can try SpringFox, or provide the schema manually
