### 2020-10-08 EMResearch/EvoMaster#199 @mado89
Same behaviour with SpringFox.
The "application" is located at https://github.com/mado89/EvoMasterTest please take a look if you can spot the problem.
