### 2020-10-08 EMResearch/EvoMaster#199 @arcuri82
I cloned it, and, with SpringDoc (commented out SpringFox), I did not have any issue using EvoMaster on it... (started the TestEvoMaster in IntelliJ)
