### 2020-10-09 EMResearch/EvoMaster#199 @mado89
Ok then this is some weird IDE (eclipse) issue in connection with EvoMaster or undocumented behaviour of springdoc. I tried running it from the commandline with the same result.
Maybe you could provide a sample gradle task for running a test driver? This would also improve the tools integration into fully automated CI processes.
