### 2020-10-09 EMResearch/EvoMaster#199 @mado89
Finally!
I found a work around and was able to run a whitebox test
