### 2020-10-09 #2153 @UTH-Tuan
@kthyng I get the following message "Sorry, we couldn't find that repository invitation. It is possible that the invitation was revoked or that you are not logged into the invited account."
