### 2020-10-10 d442612
add inject hook, refactor dir structure
