+++ client-python/tests/dummymodule.py
@@ -1,0 +1,15 @@
+"""Module docstring."""
+
+def dummy_function(x, y):
+    print('start')
+    if x > y:
+        print(x)
+    else:
+        print(y)
+    print('end')
+
+def dummy_print(x):
+    print(x)
+
+def dummy_empty(x):
+    pass
