+++ client-python/evomaster_client/controller/app.py
@@ -1,0 +1,5 @@
+from flask import Flask, Blueprint
+from controller import controller
+
+app = Flask(__name__)
+app.register_blueprint(controller)

[…]
