+++ client-python/evomaster-client-python/controller/app.py
@@ -1,6 +1,0 @@
-from flask import Flask, Blueprint
-from controller import controller
-
-
-app = Flask(__name__)
-app.register_blueprint(controller)
