### 2020-10-10 56529dc
Load em_app (proxy) using inject_hook
