+++ client-python/evomaster_client/controller/em_app.py
@@ -1,5 +1,5 @@
 from flask import Flask, Blueprint
-from controller import controller
+from evomaster_client.controller.em_controller import controller
 
 app = Flask(__name__)
 app.register_blueprint(controller)
