### 2020-10-11 955b919
Clean up command to run Instrumented Flask app
