+++ client-python/tests/test_importhook.py
@@ -22,11 +22,9 @@ def dummymodule():
 
     sys.path.insert(0, str(this_dir))
     try:
-        with install_import_hook('dummymodule', tracer=ExecutionTracer()):
-            with warnings.catch_warnings():
-                warnings.filterwarnings('error', module='typeguard')
-                module = import_module('dummymodule')
-                return module
+        with install_import_hook(['dummymodule'], tracer=ExecutionTracer()):
+            module = import_module('dummymodule')
+            return module
     finally:
         sys.path.remove(str(this_dir))
 
