### 2020-10-12 17e531f
handling same gene name in one rest action
