### 2020-10-12 940dc9f
Added console options and updated docs
