+++ client-python/evomaster_client/cli.py
@@ -15,9 +15,12 @@ def evomaster():
 
 
 @evomaster.command()
-@click.option('--package-prefix', '-p', multiple=True, help='package prefix to measure code coverage')
-@click.option('--flask-module', '-m', required=True, help='module name where the flask application is defined')
-@click.option('--flask-app', '-a', default='app', help='flask app defined in flask-module')
+@click.option('--package-prefix', '-p', multiple=True,
+              help='package prefix to measure code coverage')
+@click.option('--flask-module', '-m', required=True,
+              help='module name where the flask application is defined')
+@click.option('--flask-app', '-a', default='app',
+              help='flask app defined in flask-module')
 def run(package_prefix, flask_module, flask_app):
     print(f'package_prefix={package_prefix}')
     print(f'flask_module={flask_module}')
