+++ client-python/evomaster_client/instrumentation/execution_tracer.py
@@ -2,6 +2,7 @@ class AdditionalInfo:
     def __init__(self) -> None:
         self.info = []
 
+
 class ExecutionTracer:
     def __init__(self) -> None:
         self.reset()
