+++ client-python/evomaster_client/instrumentation/execution_tracer.py
@@ -1,6 +1,38 @@
+from typing import Sequence, Set
+
+from evomaster_client.instrumentation import objective_naming
+
+
 class AdditionalInfo:
     def __init__(self) -> None:
-        self.info = []
+        self.query_parameters = set()
+        self.headers = set()
+        self.last_executed_statement_stack = []
+        self.no_exception_statement = None
+
+    def push_last_executed_statement(self, last_line: str) -> None:
+        self.no_exception_statement = None
+        self.last_executed_statement_stack.append(last_line)
+
+    def pop_last_executed_statement(self) -> str:
+        statement = self.last_executed_statement_stack.pop()
+        if not self.last_executed_statement_stack:
+            self.no_exception_statement = statement
+        return statement
+
+
+class Action:
+    def __init__(self, index: int, input_variables: Sequence[str]) -> None:
+        self.index = index
+        self.input_variables = input_variables
+
+
+class TargetInfo:
+    def __init__(self, mapped_id: int, descriptive_id: str, value: float, action_index: int):
+        self.mapped_id = mapped_id
+        self.descriptive_id = descriptive_id
+        self.value = value
+        self.action_index = action_index
 
 
 class ExecutionTracer:
@@ -14,233 +46,57 @@ class ExecutionTracer:
         self.additional_info_list = []
         self.additional_info_list.append(AdditionalInfo())
 
-    # static setAction(action: Action) {
-    #     if (action.getIndex() != ExecutionTracer.actionIndex) {
-    #         ExecutionTracer.actionIndex = action.getIndex();
-    #         ExecutionTracer.additionalInfoList.push(new AdditionalInfo());
-    #     }
+    def set_action(self, action: Action) -> None:
+        if action.index != self.action_index:
+            self.action_index = action.index
+            self.additional_info_list.append(AdditionalInfo())
+        if action.input_variables:
+            self.input_variables = action.input_variables
 
-    #     if (action.getInputVariables() && action.getInputVariables().size > 0) {
-    #         ExecutionTracer.inputVariables = action.getInputVariables();
-    #     }
-    # }
+    def mark_last_executed_statement(self, last_line: str) -> None:
+        self.additional_info_list[self.action_index].push_last_executed_statement(last_line)
 
-    # /**
-    #  * Check if the given input represented a tainted value from the test cases.
-    #  * This could be based on static info of the input (eg, according to a precise
-    #  * name convention given by TaintInputName), or dynamic info given directly by
-    #  * the test itself (eg, the test at action can register a list of values to check
-    #  * for)
-    #  */
-    # // public static boolean isTaintInput(String input){
-    # //     return TaintInputName.isTaintInput(input) || inputVariables.contains(input);
-    # // }
+    def completed_last_executed_statement(self, last_line: str) -> None:
+        self.additional_info_list[self.action_index].pop_last_executed_statement()
 
+    def get_number_of_objectives(self, prefix: str = '') -> int:
+        return len([k for k in self.objective_coverage.keys() if k.startsWith(prefix)])
 
-    # // public static TaintType getTaintType(String input){
-    # //
-    # //     if(input == null){
-    # //         return TaintType.NONE;
-    # //     }
-    # //
-    # //     if(isTaintInput(input)){
-    # //         return TaintType.FULL_MATCH;
-    # //     }
-    # //
-    # //     if(TaintInputName.includesTaintInput(input)
-    # //         || inputVariables.stream().anyMatch(v -> input.contains(v))){
-    # //         return TaintType.PARTIAL_MATCH;
-    # //     }
-    # //
[…]
