### 2020-10-12 023cca3
Inject Execution Tracer as a singleton instance
