+++ client-python/evomaster_client/instrumentation/execution_tracer.py
@@ -35,8 +35,22 @@ class TargetInfo:
         self.action_index = action_index
 
 
-class ExecutionTracer:
-    def __init__(self) -> None:
+class Singleton(object):
+    _instances = {}
+
+    def __new__(cls, *args, **kwargs):
+        if cls not in cls._instances:
+            s = super(Singleton, cls).__new__(cls, *args, **kwargs)
+            s.initialize()
+            cls._instances[cls] = s
+        return cls._instances[cls]
+
+    def initialize(self):
+        pass
+
+
+class ExecutionTracer(Singleton):
+    def initialize(self):
         self.reset()
 
     def reset(self) -> None:
