+++ client-python/tests/dummymodule.py
@@ -1,5 +1,6 @@
 """Module docstring."""
 
+
 def dummy_function(x, y):
     print('start')
     if x > y:
@@ -8,8 +9,10 @@ def dummy_function(x, y):
         print(y)
     print('end')
 
+
 def dummy_print(x):
     print(x)
+
 
 def dummy_empty(x):
     pass
