+++ client-python/evomaster_client/instrumentation/ast_transformer.py
@@ -2,8 +2,7 @@ import ast
 
 
 class AstTransformer(ast.NodeTransformer):
-    def __init__(self, module):
-        self.targets = set()  # TODO: move to ExecutionTracer
+    def __init__(self, module: str):
         self.module = module
         self.statement_counter = 0
 
@@ -18,12 +17,10 @@ class AstTransformer(ast.NodeTransformer):
     def visit_Statement(self, node):
         self.generic_visit(node)  # visit child nodes
         print("Visited node of type: ", node.__class__.__name__, " - line no:", node.lineno)
+
         if hasattr(node, 'body'):
             print("isBlockStatement. no point in instrumenting it. Recall, we still instrument its content anyway.")
             return node
-        self.targets.add(f"LINE_{self.module}_{node.lineno}")
-        self.statement_counter += 1
-        self.targets.add(f"STATEMENT_{self.module}_{node.lineno}_{self.statement_counter}")
 
         # For nodes that were part of a collection of statements (that applies to all statement nodes),
         # the visitor may also return a list of nodes rather than just a single node.
