### 2020-10-12 3249471
fixes, register module names instead of paths
