### 2020-10-12 3644990
Separate proxy from real EM controller
