+++ client-python/README.md
@@ -10,7 +10,7 @@ pip install -r requirements-dev.txt
 Running the instrumented version of EvoMaster Proxy controller:
 
 ```python
-python -m evomaster_client.cli run -p 'evomaster_client.controller' -m 'evomaster_client.controller.em_app'
+python -m evomaster_client.cli run -p 'evomaster_client.proxy' -m 'evomaster_client.proxy.em_app'
 ```
 
 Running tests:
