### 2020-10-13 3ac4ec9
SUT Handler for Flask apps (start/stop server using threads)
