### 2020-11-03 #2153 @mado89
Just wanted to ask if there is still work left from my side. I've ticked of all my boxes some time ago. While I think there is room for improvement concerning usability I think this tool can be accepted for publication in JOSS.
