### 2020-12-01 #2153 @arcuri82
Hi,
* I made a new release [1.1.0](https://github.com/EMResearch/EvoMaster/releases/tag/v1.1.0)
* I created a new entry on Zenodo, [4300745](https://zenodo.org/record/4300745). However, that is the first time I use Zenodo, so not sure if I did it correctly
