### 2021-01-03 #2153 @kyleniemeyer
Hi @arcuri82, I'm just doing some final checks on your submission before publishing.

I noticed some issues with in-text citations in your article (not using a single command when citing multiple things), and made a small PR to fix this: https://github.com/EMResearch/EvoMaster/pull/232

Can you merge this? I will then publish the article. There is no need to create a new Zenodo archive. Thanks!
